% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_incorrect_type}
\alias{abort_incorrect_type}
\title{abort_incorrect_type}
\usage{
abort_incorrect_type(type_expected, obj)
}
\arguments{
\item{type_expected}{String, containing name of the type expected}

\item{obj}{R object of wrong type}
}
\description{
Report incorrect type provided
}
