% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_length_mismatch}
\alias{abort_length_mismatch}
\title{abort_length_mismatch}
\usage{
abort_length_mismatch(name1, name2, length1, length2)
}
\arguments{
\item{name1}{String, containing name of first variable}

\item{name2}{String, containing name of second variable}

\item{length1}{Numeric, containing length of first variable}

\item{length2}{Numeric, containing length of second variable}
}
\description{
Report length mismatch and abort
}
