% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_missing_column}
\alias{abort_missing_column}
\title{abort_missing_column}
\usage{
abort_missing_column(parameter)
}
\arguments{
\item{parameter}{name of missing required column}
}
\description{
Data.frame is missing necessary columns
}
