% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_name_missing}
\alias{abort_name_missing}
\title{abort_name_missing}
\usage{
abort_name_missing(name)
}
\arguments{
\item{name}{String, containing name of missing sequence}
}
\description{
Report name missing and abort
}
