% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_nonexistant_file}
\alias{abort_nonexistant_file}
\title{abort_nonexistant_file}
\usage{
abort_nonexistant_file(filename)
}
\arguments{
\item{filename}{String, containing name of the missing file or directory}
}
\description{
Report directory / file does not exist
}
