% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_provide_at_least_one}
\alias{abort_provide_at_least_one}
\title{abort_provide_at_least_one}
\usage{
abort_provide_at_least_one(parameters)
}
\arguments{
\item{parameters}{a list of parameter where one must be selected}
}
\description{
Report missing parameter
}
