% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Get the abundance data for sequences, bins, samples, and treatments in
a \href{https://mothur.org/strollur/reference/strollur.html}{strollur}
object}
\usage{
abundance(data, type = "sequence", bin_type = "otu", by_sample = FALSE)
}
\arguments{
\item{data, }{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{type, }{string containing the type of data you want the number of.
Options include: "sequence", "bin", "sample" and "treatment".
Default = "sequence".}

\item{bin_type, }{string containing the bin type you would like the abundance
data for. Default = "otu".}

\item{by_sample, }{Boolean. When by_sample is TRUE, the abundance data will
be parsed by sample. Default = FALSE.}
}
\value{
data.frame
}
\description{
Get the abundance data for sequences, bins, samples, and
treatments in a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

miseq <- miseq_sop_example()

# To the total abundance for each sequence
abundance(data = miseq, type = "sequence")

# To the total abundance for each sequence parsed by sample
abundance(data = miseq, type = "sequence", by_sample = TRUE)

# To the total abundance for each "otu" bin
abundance(data = miseq, type = "bin", bin_type = "otu")

# To the total abundance for each "otu" bin parsed by sample
abundance(data = miseq, type = "bin", bin_type = "otu", by_sample = TRUE)

# To the total abundance for each "asv" bin
abundance(data = miseq, type = "bin", bin_type = "asv")

# To the total abundance for each "asv" bin parsed by sample
abundance(data = miseq, type = "bin", bin_type = "asv", by_sample = TRUE)

# To the total abundance for each sample
abundance(data = miseq, type = "sample")

# To the total abundance for each treatment
abundance(data = miseq, type = "treatment")

}
