% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_metadata}
\alias{add_metadata}
\title{add_metadata}
\usage{
add_metadata(data, metadata)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{metadata, }{a data.frame containing your metadata}
}
\description{
Add metadata to a \link{dataset} object
}
\examples{

data <- new_dataset("just for fun", 2)
metadata <- readr::read_tsv(rdataset_example("mouse.dpw.metadata"),
                            col_names = TRUE, show_col_types = FALSE)
add_metadata(data, metadata)

}
