% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_report}
\alias{add_report}
\title{add_report}
\usage{
add_report(data, table, type, sequence_name = "sequence_names")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing your report.}

\item{type, }{a string containing the type of report. For example: "align".}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}
}
\description{
Add a report to a \link{dataset} object
}
\examples{

data <- new_dataset("just for fun", 2)
align_report <- readr::read_tsv(rdataset_example("alignment_data.tsv"),
   col_names = TRUE, show_col_types = FALSE)
add_report(data, align_report, "align", "QueryName")

}
