% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_sequences}
\alias{add_sequences}
\title{add_sequences}
\usage{
add_sequences(data, sequence_names, sequences, comments)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{sequences}{a vector of strings containing sequence nucleotide strings}

\item{comments}{a vector of strings containing sequence comments}
}
\value{
double containing the number of sequences added
}
\description{
Add sequence data to an instance of the 'Dataset' class
}
\examples{

 dataset <- new_dataset("miseq_sop", 4)
 sequences <- read_fasta(rdataset_example("final.fasta"))
 add_sequences(dataset, sequences$sequence_names, sequences$sequences, "")

}
\seealso{
[sequence_data$add_sequences()]
}
