% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{added_message}
\alias{added_message}
\title{added_message}
\usage{
added_message(num = NULL, tag = "sequences")
}
\arguments{
\item{num}{integer containing the number of items added.}

\item{tag}{string containing item added. Default = 'sequences'.}
}
\description{
Report dataset additions
}
