% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{alert_eliminated_name}
\alias{alert_eliminated_name}
\title{alert_eliminated_name}
\usage{
alert_eliminated_name(name, return_value)
}
\arguments{
\item{name}{String, containing name of eliminated sequence}

\item{return_value}{String, containing return_value}
}
\description{
Report eliminated name
}
