% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{alert_missing_name}
\alias{alert_missing_name}
\title{alert_missing_name}
\usage{
alert_missing_name(name, return_value)
}
\arguments{
\item{name}{String, containing name of missing sequence}

\item{return_value}{String, containing return_value}
}
\description{
Report missing name
}
