% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bin_representative_sequences}
\alias{assign_bin_representative_sequences}
\title{assign_bin_representative_sequences}
\usage{
assign_bin_representative_sequences(
  data,
  bin_names,
  sequence_names,
  type = "otu"
)
}
\arguments{
\item{data}{a data.frame containing bin_names and representative
sequences for your data.}

\item{bin_names}{a vector strings containing bin names or if using the
'data' parameter a string containing the name of the column in 'data'
that contains the bin names. Default column name is 'bin_names'.}

\item{sequence_names}{a vector of strings containing bin representative
sequences or if using the 'data' parameter a string containing the name
of the column in 'data' that contains the bin names. Default column name
is 'sequence_names'.}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
double containing the number of representative sequences assigned
}
\description{
Assign representative sequences to bins.
}
\examples{

  miseq <- miseq_sop_example()

  # For examples sake, select first 531 sequences to be the representatives
  num_bins <- miseq$get_num_bins("otu")
  rep_names <- miseq$get_sequence_names()[1:num_bins]
  bin_names <- miseq$get_bin_names()

  assign_bin_representative_sequences(miseq$data,
                                      bin_names, rep_names, "otu")

}
\seealso{
[dataset$assign_bin_representative_sequences()]
}
