% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bin_taxonomy}
\alias{assign_bin_taxonomy}
\title{assign_bin_taxonomy}
\usage{
assign_bin_taxonomy(
  data,
  table,
  type = "otu",
  reference = NULL,
  bin_name = "bin_names",
  taxonomy = "taxonomies"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing bin taxonomy assignments}

\item{type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{bin_name, }{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'bin_names'.}

\item{taxonomy, }{a string containing the name of the column in 'table' that
contains the bin taxonomies. Default column name is 'taxonomies'.}
}
\value{
double containing the number of bins assigned
}
\description{
Assign bin classifications to a \link{dataset} object

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the concensus taxonomy for each bin for you.
}
\examples{

otu_data <- read_mothur_cons_taxonomy(rdataset_example(
                        "final.cons.taxonomy"))

data <- new_dataset("my_dataset", 2)
assign_bins(data, otu_data)
assign_bin_taxonomy(data, otu_data)

}
