% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bins}
\alias{assign_bins}
\title{assign_bins}
\usage{
assign_bins(data, bin_names, abundances, samples, sequence_names, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector strings containing of bin labels}

\item{abundances}{a vector of integers containing abundances. Note: You must
provide either abundances or seq_ids.}

\item{samples}{a vector of strings containing sample assignments}

\item{sequence_names}{a vector of strings containing sequence names.
Note: You must provide either abundances or seq_ids.}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\description{
Add bin assignments to an instance of the 'Dataset' class.
}
\examples{

  # To assign sequences to bins:

  dataset <- new_dataset("my_dataset", 4)
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  assign_bins(dataset, bin_ids, 0, "", seq_ids, "otu")

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5

  # To add abundance only bin assignments:

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  assign_bins(dataset, bin_ids, abundances, "", "", "otu")

  # bins would look like:
  #            bin1             bin2        bin3
  # (rabund)   110              525         80

  # To add abundance bin assignments parsed by sample:

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  assign_bins(dataset, bin_ids, sample_abundances, samples, "", "otu")

  # (shared) bins would look like:
  # sample   bin1   bin2   bin3
  # sample1  10     500    80
  # sample2  100    0      0
  # sample3  0      25     0
  # sample5  1      0      0

  # To assign sequences to bins with their abundances parsed by sample:

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin1", "bin1", "bin1", "bin1", "bin1",
               "bin2", "bin2", "bin2",
               "bin3", "bin3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  assign_bins(dataset, bin_ids, abundances, samples, seq_ids, "otu")

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5
  # (rabund)   716              85          65

}
\seealso{
[sequence_data$assign_bins()]
}
