% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_sequence_abundance}
\alias{assign_sequence_abundance}
\title{assign_sequence_abundance}
\usage{
assign_sequence_abundance(
  data,
  table,
  sequence_name = "sequence_names",
  abundance = "abundances",
  sample = "samples",
  treatment = "treatments"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing sequence abundance assignments}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}

\item{abundance, }{a string containing the name of the column in 'table'
that contains the sequence abundances. Default column name is 'abundances'.}

\item{sample, }{a string containing the name of the column in 'table'
that contains the sequence samples. Default column name is 'samples'.
(Optional)}

\item{treatment, }{a string containing the name of the column in 'table'
that contains the sequence treatments. Default column name is 'treatments'.}
}
\value{
double containing the number of sequences assigned
}
\description{
Set sequence abundance and optionally assign sample and treatment data to a
 \link{dataset} object
}
\examples{

data <- new_dataset("my_dataset", 2)
sequence_abundance <- readr::read_tsv(rdataset_example(
                                      "mothur2_count_table.tsv"),
                                      show_col_types = FALSE)

assign_sequence_abundance(data, sequence_abundance, "names")

}
