% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_treatments}
\alias{assign_treatments}
\title{assign_treatments}
\usage{
assign_treatments(data, table, sample = "samples", treatment = "treatments")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing sample treatment assignments}

\item{sample, }{a string containing the name of the column in 'table'
that contains the samples. Default column name is 'samples'.}

\item{treatment, }{a string containing the name of the column in 'table'
that contains the treatments. Default column name is 'treatments'.}
}
\value{
double containing the number of samples assigned to treatments
}
\description{
Assign samples to treatments in a \link{dataset} object
}
\examples{

data <- new_dataset("my_dataset", 2)
sequence_abundance <- readr::read_tsv(rdataset_example(
                                      "mothur2_count_table.tsv"),
                                      show_col_types = FALSE)

assign_sequence_abundance(data, sequence_abundance, "names")

sample_assignments <- readr::read_table(file = rdataset_example("mouse.time.design"),
                             col_names = TRUE, show_col_types = FALSE)

assign_treatments(data, sample_assignments)

}
