% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{assigned_message}
\alias{assigned_message}
\title{assigned_message}
\usage{
assigned_message(num, tag = "sequences")
}
\arguments{
\item{num}{integer containing the number of items assigned}

\item{tag}{string containing item assigned Default = 'sequences'.}
}
\value{
No return value, called for side effects.
}
\description{
Report dataset assignments
}
\examples{
assigned_message(10)
assigned_message(10, "sequence abundances")
}
