% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear}
\alias{clear}
\title{Clear data from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
clear(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Clear data from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- miseq_sop_example()
clear(data)

}
