% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{convert_qual}
\alias{convert_qual}
\title{convert_qual}
\usage{
convert_qual(quality_string, format = "illumina1.8+", table = NULL)
}
\arguments{
\item{quality_string, }{string containing quality data string}

\item{format}{string, containing the FASTQ format. Options include:
\verb{illumina1.8+}, \code{solexa}, \code{sanger}. Default = \verb{illumina1.8+}.}

\item{table}{quality conversion table for solexa format. Default = \code{NULL}.}
}
\value{
vector of integers
}
\description{
Convert quality string to quality scores
}
\examples{

quality_string <- "3AA?ABBDBFFBEGGEGGGGAFFGGGGGHHHCGGGGGGHFGHGGCFDEFGGGH"
convert_qual(quality_string)

}
