% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fastq.R
\name{convert_qual_string}
\alias{convert_qual_string}
\title{convert_qual_string}
\usage{
convert_qual_string(quality_score, format = "illumina1.8+", table = NULL)
}
\arguments{
\item{quality_score}{vector containing the quality scores}

\item{format}{string, containing the FASTQ format. Options include:
\verb{illumina1.8+}, \code{solexa}, \code{sanger}. Default = \verb{illumina1.8+}.}

\item{table}{quality conversion table for solexa format. Default = \code{NULL}.}
}
\value{
vector of characters
}
\description{
Convert quality scores to quality string
}
\examples{

quality_score <- c(
  2, 29, 29, 32, 32, 33, 32, 33, 33,
  37, 37, 37, 38, 38, 38
)

convert_qual_string(quality_score)

}
