% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_pointer}
\alias{copy_pointer}
\title{copy_pointer}
\usage{
copy_pointer(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
pointer to an instance of the C++ 'Dataset' class.
}
\description{
For internal use only, copy an instance of the C++ 'Dataset' class.
}
\keyword{internal}
