% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deserialize_dobject}
\alias{deserialize_dobject}
\title{deserialize_dobject}
\usage{
deserialize_dobject(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\description{
For internal use only, deserialize_dobject an instance of the C++ 'Dataset'
 class.
}
