% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{.abort_incorrect_format}
\alias{.abort_incorrect_format}
\title{.abort_incorrect_format}
\usage{
.abort_incorrect_format(type_expected, type_received)
}
\arguments{
\item{type_expected}{String, containing name of the type expected}

\item{type_received}{String, containing name of the type received}
}
\description{
Report incorrect artifact format provided
}
\keyword{internal}
