% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{.abort_not_writable}
\alias{.abort_not_writable}
\title{.abort_not_writable}
\usage{
.abort_not_writable(filename)
}
\arguments{
\item{filename}{String, containing name of the file without write
permissions}
}
\description{
Report directory / file is not writable
}
\keyword{internal}
