% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.split_white_space}
\alias{.split_white_space}
\title{.split_white_space}
\usage{
.split_white_space(line)
}
\arguments{
\item{line, }{String containing data to split}
}
\value{
A vector of Strings
}
\description{
Split string at white space
}
\keyword{internal}
