% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{export}
\usage{
export(dataset, tags = NULL)
}
\arguments{
\item{dataset}{a 'sequence_data' object}

\item{tags}{a vector of strings containing the items you wish to export.
Options are 'sequence_data', 'bin_data', 'metadata',
'references', 'sequence_tree', 'sample_tree', 'alignment_report',
'contigs_assembly_report' and 'chimera_report'. By default, everything is
 exported.}
}
\value{
A list of data.frames
}
\description{
The export function will create a list containing the data in the dataset.
}
\examples{

miseq <- miseq_sop_example()
table <- export(miseq)

}
\seealso{
[sequence_data$export()]
}
