% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_processors.R
\name{get_available_processors}
\alias{get_available_processors}
\title{Get available processors}
\usage{
get_available_processors()
}
\value{
Integer
}
\description{
Get the number of available cores
}
\examples{

get_available_processors()

}
