% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin}
\alias{get_bin}
\title{get_bin}
\usage{
get_bin(data, bin_name, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{bin_name, }{string containing the bin name}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\value{
String, containing names of the sequences in a given bin
}
\description{
Get the names of the sequences in a given bin in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_bin(data, "Otu005", "otu")

}
