% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_abundance}
\alias{get_bin_abundance}
\title{get_bin_abundance}
\usage{
get_bin_abundance(data, bin_name, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{bin_name, }{string containing the bin name}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\value{
double, containing the abundance of a given bin
}
\description{
Get the abundance of a given bin in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_bin_abundance(data, "Otu005", "otu")

}
