% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_assignments}
\alias{get_bin_assignments}
\title{get_bin_assignments}
\usage{
get_bin_assignments(data, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
data.frame
}
\description{
Get data.frame containing bin abundance data
}
\examples{

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  assign_bins(dataset, bin_ids, sample_abundances, samples, "")

  # (shared) bins would look like:
  # sample   bin1   bin2   bin3
  # sample1  10     500    80
  # sample2  100    0      0
  # sample3  0      25     0
  # sample5  1      0      0

  get_bin_assignments(dataset)

}
