% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_names}
\alias{get_bin_names}
\title{get_bin_names}
\usage{
get_bin_names(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string indicating the type of clusters. Default = "otu".}
}
\value{
vector containing the names of bins
}
\description{
Get the names of the bins in a \link{dataset} object
}
\examples{

  miseq <- miseq_sop_example()
  get_bin_names(miseq, "phylotype")

}
