% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_representative_sequences}
\alias{get_bin_representative_sequences}
\title{get_bin_representative_sequences}
\usage{
get_bin_representative_sequences(data, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{type, }{string indicating the type of clusters. Default = "otu".}
}
\value{
data.frame
}
\description{
Get the representative sequences of the bins in an instance of the
 'Dataset' class.
}
\examples{

  miseq <- miseq_sop_example()

  # For examples sake, select first 531 sequences to be the representatives
  num_bins <- miseq$get_num_bins("otu")
  rep_names <- miseq$get_sequence_names()[1:num_bins]
  bin_names <- miseq$get_bin_names()

  assign_bin_representative_sequences(miseq$data,
                                      bin_names, rep_names, "otu")



  get_bin_representative_sequences(miseq$data, "otu")

}
\seealso{
[dataset$get_bin_representative_sequences()]
}
