% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_taxonomy_report}
\alias{get_bin_taxonomy_report}
\title{get_bin_taxonomy_report}
\usage{
get_bin_taxonomy_report(data, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{type, }{string indicating the type of bin clusters. Default = "otu".}
}
\value{
data.frame
}
\description{
Get the bin classifications of an instance of the 'Dataset' class.
}
\examples{

bin_ids <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- new_dataset("my_dataset", 4)
assign_bins(dataset, bin_ids, abunds, "", "")
assign_bin_taxonomy(dataset, bin_ids, taxonomies)
get_bin_taxonomy_report(dataset)

}
