% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_custom_report_types}
\alias{get_custom_report_types}
\title{get_custom_report_types}
\usage{
get_custom_report_types(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
vector of strings
}
\description{
Get the custom report types of a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_custom_report_types(data)

}
