% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dataset_name}
\alias{get_dataset_name}
\title{get_dataset_name}
\usage{
get_dataset_name(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
String, containing the name of the dataset
}
\description{
Get the name of an instance of the 'Dataset' class.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
get_dataset_name(dataset)

}
