% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_list}
\alias{get_list}
\title{get_list}
\usage{
get_list(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
2 column data.frame sequences assigned to bins
}
\description{
Get data frame containing sequence bin assignments of a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_list(data)

}
