% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
data.frame
}
\description{
Get the metadata of a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_metadata(data)

}
