% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_bins}
\alias{get_num_bins}
\title{get_num_bins}
\usage{
get_num_bins(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\value{
Integer, the number of bins of a specific type in a \link{dataset}
 object
}
\description{
Get the number of bins of a specific type in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_num_bins(data, "otu")
get_num_bins(data, "asv")
get_num_bins(data, "phylotype")

}
