% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_processors}
\alias{get_num_processors}
\title{get_num_processors}
\usage{
get_num_processors(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
Integer, containing number of processors
}
\description{
Get the number of processors used to summarize a \link{dataset} object
}
\examples{
data <- new_dataset("my_dataset", 2)
get_num_processors(data)

}
