% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_samples}
\alias{get_num_samples}
\title{get_num_samples}
\usage{
get_num_samples(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Integer, the number of samples in an instance of the 'Dataset' class.
}
\description{
Get the number of samples in an instance of the 'Dataset' class.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
samples <- c("sample1", "sample2", "sample5", "sample1", "sample3", "sample1")
sample_abundances <- c(10, 100, 1, 500, 25, 80)
assign_bins(dataset, bin_ids, sample_abundances, samples, "")
get_num_samples(dataset)

}
