% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_sequences}
\alias{get_num_sequences}
\title{get_num_sequences}
\usage{
get_num_sequences(data, distinct = FALSE, sample = "")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{distinct}{Boolean. When distinct is TRUE the number of unique
sequence is returned. Default = FALSE.}

\item{sample, }{string containing the name of the sample you want number of
sequences for.}
}
\value{
double, the number of sequences in a \link{dataset} object
}
\description{
Get the number of sequences in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_num_sequences(data)
get_num_sequences(data, TRUE)
get_num_sequences(data, FALSE, "F3D0")
get_num_sequences(data, TRUE, "F3D0")

}
