% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_rabund}
\alias{get_rabund}
\title{get_rabund}
\usage{
get_rabund(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
a 2 column data.frame containing bin names and bin abundances
}
\description{
Get data.frame containing bin abundance data in a \link{dataset} object
}
\examples{

  data <- miseq_sop_example()
  get_rabund(data)

}
