% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_rabund_vector}
\alias{get_rabund_vector}
\title{get_rabund_vector}
\usage{
get_rabund_vector(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
vector containing each bins abundance
}
\description{
Get vector containing bin abundance data in a \link{dataset} object
}
\examples{

  data <- miseq_sop_example()
  get_rabund_vector(data)

}
