% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_references}
\alias{get_references}
\title{get_references}
\usage{
get_references(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
data.frame
}
\description{
Get a table containing resource references in a \link{dataset} object
}
