% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_reports}
\alias{get_reports}
\title{get_reports}
\usage{
get_reports(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
list
}
\description{
Get a list containing the reports in a \link{dataset} object
}
