% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sample_totals}
\alias{get_sample_totals}
\title{get_sample_totals}
\usage{
get_sample_totals(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
vector containing the number of sequences in each sample in a
\link{dataset} object
}
\description{
Get the number of sequences in each sample in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sample_totals(data)

}
