% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
vector of strings containing the names of the samples in a
\link{dataset} object
}
\description{
Get the samples in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_samples(data)

}
