% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_abundance_table}
\alias{get_sequence_abundance_table}
\title{get_sequence_abundance_table}
\usage{
get_sequence_abundance_table(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
data.frame containing sequence abundance data in a \link{dataset}
object
}
\description{
Get the abundances of each sequence in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sequence_abundance_table(data)

}
