% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_abundances_by_sample}
\alias{get_sequence_abundances_by_sample}
\title{get_sequence_abundances_by_sample}
\usage{
get_sequence_abundances_by_sample(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
2D vector of integers ([num_seqs][num_samples]) containing the
abundances of each sequence in an instance of the 'Dataset' class parsed by
sample.
}
\description{
Get the abundances of each sequence in an instance of the 'Dataset' class
parsed by sample.
}
\examples{

names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, names, abundances, samples, "")
get_sequence_abundances_by_sample(dataset)

}
