% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_names}
\alias{get_sequence_names}
\title{get_sequence_names}
\usage{
get_sequence_names(data, sample = "")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{sample}{a string containing the name of the sample you
would like sequence names for. For all samples in dataset, sample = "".}
}
\value{
vector of string containing the names of the sequences a
\link{dataset} object
}
\description{
Get the names of the sequences in a \link{dataset} object
}
\examples{


data <- miseq_sop_example()

# to get the names of all the sequences in the dataset

get_sequence_names(data)

# to get the names of the sequences in sample 'F3D0' from the dataset

get_sequence_names(data, "F3D0")

}
