% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_names_by_sample}
\alias{get_sequence_names_by_sample}
\title{get_sequence_names_by_sample}
\usage{
get_sequence_names_by_sample(data, samples = as.character(c()))
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{samples}{a vector of strings containing the names of the samples you
would like sequence names for. By default all samples are included.}
}
\value{
2D vector of strings ([num_seqs][num_samples]) containing the
names of each sequence a \link{dataset} object parsed by sample.
}
\description{
Get the names of the sequences in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()

get_sequence_names_by_sample(data)

}
