% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_report}
\alias{get_sequence_report}
\title{get_sequence_report}
\usage{
get_sequence_report(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
data.frame
}
\description{
Get sequence report data: starts, ends, lengths, ambigs, longest
homopolymers and numns.
}
\examples{

 data <- miseq_sop_example()
 get_sequence_report(data)

}
