% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_summary}
\alias{get_sequence_summary}
\title{get_sequence_summary}
\usage{
get_sequence_summary(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
list of data.frames containing the 'sequence_summary' table and
'scrap_summary' table if sequences have been removed
}
\description{
Get a summary of the sequence report data, as well as reports of containing
scrapped data.
}
\examples{

 data <- miseq_sop_example()

 # Sequence summary, nothing has been scrapped

 get_sequence_summary(data)

 # Sequence summary, after removing sample 'F3D0'

 xdev_remove_samples(data, c("F3D0"))
 get_sequence_summary(data)

}
