% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_taxonomy_report}
\alias{get_sequence_taxonomy_report}
\title{get_sequence_taxonomy_report}
\usage{
get_sequence_taxonomy_report(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
data.frame
}
\description{
Get the sequence classifications of a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sequence_taxonomy_report(data)

}
