% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequences}
\alias{get_sequences}
\title{get_sequences}
\usage{
get_sequences(data, sample = "")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sample}{a string containing the name of the sample you
would like sequence names for. For all samples in dataset, sample = "".}
}
\value{
vector of string containing nucleotide strings of the sequences in
the 'Dataset' class.
}
\description{
Get the nucleotide strings for each sequence in an instance of the 'Dataset'
 class.
}
\examples{

 dataset <- new_dataset("miseq_sop", 4)
 sequences <- read_fasta(rdataset_example("final.fasta"))
 add_sequences(dataset, sequences$names, sequences$sequences, "")
 get_sequences(dataset)

}
