% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_shared_vector}
\alias{get_shared_vector}
\title{get_shared_vector}
\usage{
get_shared_vector(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
2D vector ([num_bins][num_samples]) containing the abundances of
each bin parsed by sample.
}
\description{
Get 2D vector containing bin abundance data by sample in a \link{dataset} object
}
\examples{

 data <- miseq_sop_example()
 get_shared_vector(data)

}
