% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_sample}
\alias{has_sample}
\title{Determine if a given sample is in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
has_sample(data, sample)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}.}

\item{sample}{a string containing the name of a sample.}
}
\value{
boolean indicating whether the dataset has a given sample
}
\description{
Determine if a given sample is in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

data <- miseq_sop_example()
has_sample(data, "F3D0")
has_sample(data, "not a valid sample")

}
