% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(table)
}
\arguments{
\item{table}{a table containing the data from a 'sequence_data' object. You
can create the table using 'sequence_data$export()'.}
}
\value{
A 'sequence_data' object
}
\description{
The import function will create a 'sequence_data' object from the exported
table of a 'sequence_data' object.
}
\examples{

miseq <- miseq_sop_example()
dataset <- import(miseq$export())
dataset

}
\seealso{
[sequence_data$export()]
}
