% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_aligned}
\alias{is_aligned}
\title{Determine if a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} contains aligned sequences.}
\usage{
is_aligned(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
}
\value{
Boolean
}
\description{
Determine if a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} contains aligned sequences.
}
\examples{

dataset <- miseq_sop_example()
is_aligned(dataset)

}
