% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{load}
\usage{
load(file)
}
\arguments{
\item{file}{a string containing the .rds file name.}
}
\value{
A 'dataset' object
}
\description{
The load function will create a 'dataset' object from an RDS file.
}
\examples{

dataset <- load(rdataset_example("miseq_sop.rds"))
dataset

}
