% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_bins}
\alias{merge_bins}
\title{merge_bins}
\usage{
merge_bins(data, bin_names, reason = "merged", type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{bin_names, }{a vector of strings containing the names of the bins you
would like merge. The resulting merged bin will be stored in the first
bin_id in the vector.}

\item{reason, }{a string indicating why you are merging bins. Default =
"merged".}

\item{type, }{a string indicating the type of bin clusters. Default = "otu"}
}
\description{
Designed with package integration in mind, the merge bins function allows
you to merge bins in a \link{dataset} object
}
\examples{

 data <- miseq_sop_example()

 # to merge otu5 and otu6

 bins_to_merge <- c("Otu005", "Otu006")

 merge_bins(data, bins_to_merge)

 # If you look at the scrap report, you will see Otu006 with the trash code
 # set to "merged".

 get_scrap_report(data, "bin")

}
