% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_sequences}
\alias{merge_sequences}
\title{merge_sequences}
\usage{
merge_sequences(data, sequence_names, reason = "merged")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing the names of the sequences you
sequences you would like merge. The resulting merged sequence will be stored
in the first sequence name in the vector.}

\item{reason}{a string indicating why you are merging sequences.
Default = "merged"}
}
\description{
Merge sequences in an instance of the 'Dataset' class.
}
\examples{

sequence_names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4")
abundances <- c(250, 400, 500, 25, 40, 50, 25, 25, 4)

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, sequence_names, abundances, samples, "")

seqs_to_merge <- c("seq1", "seq4")
merge_sequences(dataset, seqs_to_merge, "identical")

}
