% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{name}
\alias{name}
\title{name}
\usage{
name(data, type = "sequences", bin_type = "otu", sample = "", distinct = FALSE)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you would like. Options
include: "dataset", "sequences", "bins", "samples", "treatments", "reports".
Default = "sequences".}

\item{bin_type, }{string containing the bin type you would like the names
for. Default = "otu".}

\item{sample, }{string. sample is only used when 'type' <- "sequences" or
'type' <- "bins" . sample should contain the name of the sample you want
names for. Default = "".}

\item{distinct, }{Boolean. distinct is only used when 'type' <- "bins" and
the sample parameter is used. The distinct parameter allows you to get the
names of the bins that are unique to a given sample. When distinct is TRUE,
the names function will return the names of the bins that ONLY contain
sequences from the given sample. When distinct is FALSE the bins return
contains sequences from a given sample, but may ALSO contain sequences from
other samples. Default = FALSE.}
}
\value{
vector of strings, containing the names requested
}
\description{
Get the names of a given type of data in a \link{dataset} object
}
\examples{

miseq <- miseq_sop_example()

# To get the name of the dataset
name(data = miseq, type = "dataset")

# To get the names of the sequences in the dataset
name(data = miseq, type = "sequences")

# To get the names of the sequences in sample 'F3D0' in the dataset
name(data = miseq, type = "sequences", sample = "F3D0")

# To get the names of the samples in the dataset
name(data = miseq, type = "samples")

# To get the names of the treatments in the dataset
name(data = miseq, type = "treatments")

# To get the names of the bins in the dataset
name(data = miseq, type = "bins")

# To get the names of the bins in the dataset that are unique to 'F3D0'
name(data = miseq, type = "bins", sample = "F3D0", distinct = TRUE)

# To get the names of the bins in the dataset that include sequences
# from 'F3D0'
name(data = miseq, type = "bins", sample = "F3D0", distinct = FALSE)

# To get the names of the reports in the dataset
name(data = miseq, type = "reports")

}
