% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_pointer}
\alias{new_pointer}
\title{new_pointer}
\usage{
new_pointer(dataset_name = "", processors = 1L)
}
\arguments{
\item{dataset_name, }{string containing dataset name}

\item{processors, }{number of processors to use}
}
\value{
pointer to an instance of the C++ 'Dataset' class.
}
\description{
For internal use only, create an instance of the C++ 'Dataset' class.
}
\keyword{internal}
