% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num}
\alias{num}
\title{num}
\usage{
num(data, type = "sequences", bin_type = "otu", distinct = FALSE, sample = "")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you want the number of.
Options include: "sequences", "samples", "treatments", "bins".
Default = "sequences".}

\item{bin_type, }{string containing the bin type you would like the number of
bins for. Default = "otu".}

\item{distinct, }{Boolean. distinct is only used when 'type' <- "sequences".
When distinct is TRUE the number of unique sequences is returned.
Default = FALSE.}

\item{sample, }{string. sample is only used when 'type' <- "sequences". sample
should contain the name of the sample you want number of sequences for.}
}
\value{
double
}
\description{
Find the number of sequences, samples, treatments or bins of a given type in
a \link{dataset} object
}
\examples{

miseq <- miseq_sop_example()

# To get number of sequences in the dataset
num(data = miseq, type = "sequences")

# To get number of unique sequences in the dataset
num(data = miseq, type = "sequences", distinct = TRUE)

# To get number of sequences in the dataset from sample 'F3D0'
num(data = miseq, type = "sequences", sample = "F3D0")

# To get number of unique sequences in the dataset from sample 'F3D0'
num(data = miseq, type = "sequences", distinct = TRUE, sample = "F3D0")

# To get the number of samples in the dataset
num(data = miseq, type = "samples")

# To get the number of treatments in the dataset
num(data = miseq, type = "treatments")

# To get the number of "otu" bins in the dataset
num(data = miseq, type = "bins", bin_type = "otu")

# To get the number of "asv" bins in the dataset
num(data = miseq, type = "bins", bin_type = "asv")

# To get the number of "phylotype" bins in the dataset
num(data = miseq, type = "bins", bin_type = "phylotype")

}
