\name{rdataset-package}
\alias{rdataset-package}
\alias{rdataset}
\docType{package}
\title{
  Store and transfer data used in the analysis of microbial DNA
}
\description{
  The 'rdataset' package stores the data associated with your microbial DNA
  analysis. This includes nucleotide sequences, abundance, sample and treatment
  assignments, taxonomic classifications, asv, otu and phylotype clusters,
  metadata, trees and various reports. It is designed to facilitate data
  analysis across multiple R packages with utility functions to import
  from mothur, qiime2, DADA2 and phyloseq.
}
\details{
  Get started by taking a look at our most frequently used functions:
  new_dataset(), add(), assign(), names(), count(), abundance(), report() and
  summary().
}
\author{
Sarah Westcott, Schloss Lab, University of Michigan
Maintainer: Sarah Westcott <swestcot@umich.edu>
}
\keyword{ package }
