% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataset_example.R
\name{rdataset_example}
\alias{rdataset_example}
\title{rdataset_example}
\usage{
rdataset_example(file = NULL)
}
\arguments{
\item{file}{Name of file.}
}
\description{
rdataset comes bundled with some example files in its `inst/extdata`
directory. This function make them easy to access them.
}
\details{
Get path to rdataset example
}
\examples{
rdataset_example()
rdataset_example("final.fasta.gz")
}
