% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biom.R
\name{read_biom}
\alias{read_biom}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
from a biom formatted file.}
\usage{
read_biom(biom)
}
\arguments{
\item{biom}{string containing the name of the biom file.}
}
\value{
A
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
The read_biom function reads a \href{https://biom-format.org}{biom formatted}
file and creates a \code{strollur::strollur} object.
}
\examples{

if (requireNamespace("h5lite", quietly = TRUE)) {
  data <- strollur::read_biom(biom = strollur_example("miseq_sop.otu.biom"))
  data
} else {
  message(paste(
    "To use this functionality you have to install the",
    "h5lite package."
  ))
}

}
