% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a \href{https://www.ncbi.nlm.nih.gov/genbank/fastaformat/}{FASTA}
formatted sequence file}
\usage{
read_fasta(fasta)
}
\arguments{
\item{fasta}{FASTA file name (required)}
}
\value{
A data.frame containing the FASTA sequence data
}
\description{
Read a \href{https://www.ncbi.nlm.nih.gov/genbank/fastaformat/}{FASTA}
formatted sequence file
}
\examples{

fasta_data <- strollur::read_fasta(strollur_example("final.fasta.gz"))

# fasta_data is a data.frame.
# To access the names of the sequences in the file, run the following:

fasta_data$sequence_name

# To access the sequences in the file, run the following:

fasta_data$sequence

}
