% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_file.R
\name{read_fasta_file}
\alias{read_fasta_file}
\title{read_fasta_file}
\usage{
read_fasta_file(filename)
}
\arguments{
\item{filename}{FASTA file name (required)}
}
\value{
A list containing the FASTA sequence data
}
\description{
Read a FASTA formatted sequence file
}
