% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{read_fastq}
\usage{
read_fastq(fastq, path = NULL, format = "illumina1.8+")
}
\arguments{
\item{fastq}{string, containing the FASTQ file name. Accepts compressed \code{.gz}
files as well.}

\item{path}{string, optional path to FASTQ file.}

\item{format}{string, containing the FASTQ format. Options include:
\verb{illumina1.8+}, \code{solexa}, \code{sanger}. Default = \verb{illumina1.8+}.}
}
\value{
A data.frame containing the FASTQ sequence data
}
\description{
Read a \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2847217/}{FASTQ}
formatted file
}
\examples{

fastq_data <- strollur::read_fastq(strollur_example("tiny.fastq.gz"))

# fastq_data is a data.frame.
# To access the names of the sequences in the file, run the following:

fastq_data$sequence_name

# To access the sequences in the file, run the following:

fastq_data$sequence

# To access the quality scores in the file, run the following:

fastq_data$quality_score

}
