% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_count_file.R
\name{read_mothur_count_file}
\alias{read_mothur_count_file}
\title{read_mothur_count_file}
\usage{
read_mothur_count_file(filename)
}
\arguments{
\item{filename}{count file name (required)}
}
\value{
data.table
}
\description{
Read a mothur formatted count file
}
\examples{

 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

 # returns
 # id   group abundance
 # <char>  <char>     <int>
 #  1:   seq1 sample2       250
 #  2:   seq1 sample3       400
 #  3:   seq1 sample4       500
 #  4:   seq2 sample2        25
 #  5:   seq2 sample3        40
 #  6:   seq2 sample4        50
 #  7:   seq3 sample2        25
 #  8:   seq3 sample3        25
 #  9:   seq4 sample4         4

# read a count file with groups
sample_table <- read_mothur_count_file(rdataset_example("test.count_table"))

# read a count file without groups

abund_table <- read_mothur_count_file(
    rdataset_example("test_nogroups.count_table"))

}
