% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_list.R
\name{read_mothur_list}
\alias{read_mothur_list}
\title{read_mothur_list}
\usage{
read_mothur_list(list)
}
\arguments{
\item{list}{file name. The
\href{https://mothur.org/wiki/list_file/}{list file} can be created using
several of mothur's commands.
\href{https://mothur.org/wiki/cluster/}{cluster},
\href{https://mothur.org/wiki/cluster.split/}{cluster.split},
\href{https://mothur.org/wiki/cluster.fit/}{cluster.fit} and
\href{https://mothur.org/wiki/phylotype/}{phylotype}.}
}
\value{
A data.frame containing the sequence otu assignments
}
\description{
Read a mothur formatted \href{https://mothur.org/wiki/list_file/}{list file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

# read mothur's list file into data.frame
otu_data <-
     strollur::read_mothur_list(strollur_example("final.opti_mcc.list.gz"))

# create a new empty `strollur::strollur` object
data <- strollur::new_dataset()

# assign sequences to 'otu' bins
strollur::assign(data = data, table = otu_data,
                 type = "bin", bin_type = "otu")

}
