% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_oligos.R
\name{read_mothur_oligos}
\alias{read_mothur_oligos}
\title{read_mothur_oligos}
\usage{
read_mothur_oligos(oligos)
}
\arguments{
\item{oligos}{file name. a mothur formatted
\href{https://mothur.org/wiki/oligos_file/}{oligos file}}
}
\value{
A data.frame containing the oligos data.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/oligos_file/}{oligos file}
}
\examples{

oligos <-
     strollur::read_mothur_oligos(strollur_example("paired_read.oligos"))

# Create a new dataset and add your oligos data

data <- strollur::new_dataset() |>
  strollur::add(
    table = oligos,
    type = "report",
    report_type = "oligos"
  )

}
