% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2}
\alias{read_qiime2}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
from a qiime2 outputs}
\usage{
read_qiime2(
  qza,
  metadata = NULL,
  dataset_name = "",
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)
}
\arguments{
\item{qza}{vector of filenames, .qza files containing your data from qiime2.}

\item{metadata}{filename, a .tsv file containing metadata}

\item{dataset_name}{A string containing a name for your dataset.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the unpacked artifacts
and temporary directories will be removed. Default = TRUE.}
}
\value{
a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
The read_qiime2 function reads various types of .qza files created by
\href{https://qiime2.org}{qiime2}, and creates a \code{strollur::strollur} object.
}
\examples{

# Using the example files from moving-pictures, we add FASTA data, assign
# taxonomy and abundance for features, and add a newick tree and
# metadata.

qza_files <- c(
  strollur_example("rep_seqs.qza"),
  strollur_example("table.qza"),
  strollur_example("taxonomy.qza"),
  strollur_example("rooted-tree.qza")
)

if (requireNamespace("h5lite", quietly = TRUE)) {
  data <- strollur::read_qiime2(
    qza = qza_files,
    metadata = strollur_example("sample_metadata.tsv"),
    dataset_name = "qiime2_moving_pictures"
  )
  data
} else {
  message(paste(
    "To use this functionality you have to install the",
    "h5lite package."
  ))
}

}
\references{
Bolyen,E., Rideout,J.R., Dillon,M.R. et al. (2019),
Reproducible, interactive, scalable and extensible microbiome data
science using QIIME 2. Bioinformatics 37:852-857.
\url{doi:10.1038/s41587-019-0209-9}
}
