% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2_taxonomy.R
\name{read_qiime2_taxonomy}
\alias{read_qiime2_taxonomy}
\title{Read a \href{https://qiime2.org}{qiime2} qza containing taxonomy data}
\usage{
read_qiime2_taxonomy(qza, dir_path = NULL, remove_unpacked_artifacts = TRUE)
}
\arguments{
\item{qza}{file name, a qiime2 .qza file containing taxonomy data.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the artifact's
temporary directories will be removed after processing. Default = TRUE.}
}
\value{
A list containing artifact
}
\description{
Read a \href{https://qiime2.org}{qiime2} qza containing taxonomy data
}
\examples{

artifact <- strollur::read_qiime2_taxonomy(strollur_example(
  "taxonomy.qza"
))

# access the taxonomy table

artifact$data

}
