% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_bins}
\alias{remove_bins}
\title{remove_bins}
\usage{
remove_bins(data, bin_names, trash_tags, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector of strings containing the names of the bins you
would like removed.}

\item{trash_tags}{a vector of strings containing the reasons you are
removing each bin}

\item{type}{a string indicating the type of clusters.}
}
\description{
Remove bins from an instance of the 'Dataset' class.
}
\examples{

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  assign_bins(dataset, bin_ids, abundances, "", "")
  bins_to_remove <- c("bin1")
  trash_tag <- c("bad_bin")
  remove_bins(dataset, bins_to_remove, trash_tag)

}
