% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_lineages}
\alias{remove_lineages}
\title{remove_lineages}
\usage{
remove_lineages(data, contaminants, trash_tag = "contaminant")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{contaminants}{vector of strings containing the taxonomies you would
like to remove}

\item{trash_tag}{a string containing reason you are removing the lineages.
Default = "contaminant".}
}
\description{
Remove contaminants from an instance of the 'Dataset' class.
}
\examples{
dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       otu_list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
 "Eukaryota")

remove_lineages(dataset$data, contaminants)

}
