% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_samples}
\alias{remove_samples}
\title{remove_samples}
\usage{
remove_samples(data, samples)
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{samples, }{vector of strings containing the names of the samples to
remove.}
}
\description{
Remove samples from a \link{dataset} object
}
\examples{

data <- miseq_sop_example()

get_num_samples(data)

# To remove samples 'F3D0' and 'F3D1'

remove_samples(data, c("F3D0", "F3D1"))

get_num_samples(data)

}
