% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_sequences}
\alias{remove_sequences}
\title{remove_sequences}
\usage{
remove_sequences(data, names, trash_tags)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{names}{vector of strings containing the names of the sequences to remove}

\item{trash_tags}{vector of strings containing the reasons for the sequences
removals}
}
\description{
Remove sequences from an instance of the 'Dataset' class.
}
\examples{

seq_names <- c("seq1", "seq2", "seq3", "seq4")

dataset <- new_dataset("my_dataset", 4)
add_sequences(dataset, seq_names, "", "")

names_to_remove <- c("seq1", "seq3")
trash_tags <- c("screen_seqs-too_short", "screen_seqs-contains_ns")

remove_sequences(dataset, names_to_remove, trash_tags)

}
