% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_namespace}
\alias{require_namespace}
\title{require_namespace}
\usage{
require_namespace(package_name)
}
\arguments{
\item{package_name}{the name of the package.}
}
\value{
A logical TRUE or FALSE boolean depending on whether the
package was added to the search path.
}
\description{
a wrapper for `requireNamespace`. Allowing us to more
easily create mock test for this functionality.
}
\keyword{internal}
