% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{save}
\usage{
save(dataset, file)
}
\arguments{
\item{dataset}{a 'sequence_data' object}

\item{file}{a string containing the file name.}
}
\value{
A file containing the 'sequence_data' object
}
\description{
The save function will save the 'sequence_data' object to file.
}
\examples{

dataset <- read_mothur(
  fasta = rdataset_example("final.fasta"),
  count = rdataset_example("final.count_table"),
  taxonomy = rdataset_example("final.taxonomy"),
  design = rdataset_example("mouse.time.design"),
  otu_list = rdataset_example("final.opti_mcc.list"),
  dataset_name = "miseq_sop"
)

save(dataset, "miseq_sop.rds")

}
