% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataset.R
\name{save_dataset}
\alias{save_dataset}
\title{Save the
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
to file.}
\usage{
save_dataset(data, file)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{file}{a string containing the file name.}
}
\value{
A file containing the \code{strollur::strollur} object
}
\description{
The save_dataset function will save the
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
to file.
}
\examples{

data <- strollur::read_mothur(
  fasta = strollur_example("final.fasta.gz"),
  count = strollur_example("final.count_table.gz"),
  taxonomy = strollur_example("final.taxonomy.gz"),
  design = strollur_example("mouse.time.design"),
  otu_list = strollur_example("final.opti_mcc.list.gz"),
  dataset_name = "miseq_sop"
)

file_name <- file.path(tempdir(), "miseq_sop.rds")
strollur::save_dataset(data, file = file_name)

}
\seealso{
\code{\link[=load_dataset]{load_dataset()}}
}
