% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_data.R
\name{sequence_data}
\alias{sequence_data}
\title{sequence_dataset}
\description{
'sequence_data' is an R6 class that represents sequence
 FASTA and abundance data.
}
\examples{

## ------------------------------------------------
## Method `sequence_data$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))


## ------------------------------------------------
## Method `sequence_data$add_seqs`
## ------------------------------------------------


  dataset <- sequence_data$new("my_dataset")
  fasta_data <- read_fasta_file(rdataset_example("test.fasta"))
  dataset$add_seqs(fasta_data$names, fasta_data$sequences)


## ------------------------------------------------
## Method `sequence_data$assign_sample_abundance`
## ------------------------------------------------


 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("mydata")
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_seqs(unique_names, sequences)
dataset$assign_sample_abundance(names, groups, abundances)

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{pointer to C++ class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_data-new}{\code{sequence_data$new()}}
\item \href{#method-sequence_data-print}{\code{sequence_data$print()}}
\item \href{#method-sequence_data-add_seqs}{\code{sequence_data$add_seqs()}}
\item \href{#method-sequence_data-assign_sample_abundance}{\code{sequence_data$assign_sample_abundance()}}
\item \href{#method-sequence_data-clear}{\code{sequence_data$clear()}}
\item \href{#method-sequence_data-export}{\code{sequence_data$export()}}
\item \href{#method-sequence_data-get_dataset_name}{\code{sequence_data$get_dataset_name()}}
\item \href{#method-sequence_data-get_groups}{\code{sequence_data$get_groups()}}
\item \href{#method-sequence_data-get_group_totals}{\code{sequence_data$get_group_totals()}}
\item \href{#method-sequence_data-get_names}{\code{sequence_data$get_names()}}
\item \href{#method-sequence_data-get_num_groups}{\code{sequence_data$get_num_groups()}}
\item \href{#method-sequence_data-get_num_seqs}{\code{sequence_data$get_num_seqs()}}
\item \href{#method-sequence_data-get_num_unique}{\code{sequence_data$get_num_unique()}}
\item \href{#method-sequence_data-get_sample_table}{\code{sequence_data$get_sample_table()}}
\item \href{#method-sequence_data-get_seqs}{\code{sequence_data$get_seqs()}}
\item \href{#method-sequence_data-has_group}{\code{sequence_data$has_group()}}
\item \href{#method-sequence_data-is_aligned}{\code{sequence_data$is_aligned()}}
\item \href{#method-sequence_data-clone}{\code{sequence_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new FASTA sequence dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$new(
  name,
  fasta = NULL,
  processors = parallelly::availableCores()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of dataset}

\item{\code{fasta, }}{String name of FASTA file}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `sequence_data` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequences data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_seqs}{}}}
\subsection{Method \code{add_seqs()}}{
Add new sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_seqs(names, sequences, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data$new("my_dataset")
  fasta_data <- read_fasta_file(rdataset_example("test.fasta"))
  dataset$add_seqs(fasta_data$names, fasta_data$sequences)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sample_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sample_abundance}{}}}
\subsection{Method \code{assign_sample_abundance()}}{
Add group assignment data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sample_abundance(
  names = NULL,
  groups = NULL,
  abundances = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{groups}}{a vector of group assignments}

\item{\code{abundances}}{a vector of sample abundances}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("mydata")
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_seqs(unique_names, sequences)
dataset$assign_sample_abundance(names, groups, abundances)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all sequences from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-export"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-export}{}}}
\subsection{Method \code{export()}}{
Get List containing dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_dataset_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_dataset_name}{}}}
\subsection{Method \code{get_dataset_name()}}{
Get dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_dataset_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
Get names of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_groups(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the sequence you want groups for, optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_group_totals"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_group_totals}{}}}
\subsection{Method \code{get_group_totals()}}{
Get the number of sequences represented in samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_group_totals(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of integers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_names}{}}}
\subsection{Method \code{get_names()}}{
Get names of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_names(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_groups}{}}}
\subsection{Method \code{get_num_groups()}}{
Get number of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_seqs}{}}}
\subsection{Method \code{get_num_seqs()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_seqs(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{The name of the group you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_unique"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_unique}{}}}
\subsection{Method \code{get_num_unique()}}{
Get number of unique sequences in dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_unique()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sample_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sample_table}{}}}
\subsection{Method \code{get_sample_table()}}{
Get sample table returns data.table containing the 3 columns: name,
group (optional) and abundance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sample_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_seqs}{}}}
\subsection{Method \code{get_seqs()}}{
Get vector containing FASTA nucleotide strings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_seqs(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-has_group"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-has_group}{}}}
\subsection{Method \code{has_group()}}{
Determine if a group is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$has_group(group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-is_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-is_aligned}{}}}
\subsection{Method \code{is_aligned()}}{
Determine if the dataset is aligned and its alignment length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$is_aligned()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
bool
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
