% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serialize}
\alias{serialize}
\title{serialize}
\usage{
serialize(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Rcpp::RawVector
}
\description{
Serialize an instance of the 'Dataset' class.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
raw <- serialize(dataset)
load_dataset(dataset, raw)

}
