% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serialize_dobject}
\alias{serialize_dobject}
\title{serialize_dobject}
\usage{
serialize_dobject(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\description{
For internal use only, serialize_dobject an instance of the C++ 'Dataset'
class.
}
