% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_abundance}
\alias{set_abundance}
\title{set_abundance}
\usage{
set_abundance(data, sequence_names, sequence_abundances, reason = "update")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{sequence_abundances}{vector of integers containing the abundances of
each sequence.}

\item{reason}{a string containing the trash tag to be applied to any sequences
set to 0 abundance. Default = "update".}
}
\description{
Set sequence abundances for an instance of the 'Dataset' class without
sample data
}
\examples{

sequence_names <- c("seq1", "seq2", "seq3",  "seq4")
abundances <- c(1250, 65, 50, 4)

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, sequence_names, abundances, "", "")
get_sequence_abundances(dataset)

seqs_to_update <- c("seq1", "seq3")
new_abunds <- c(1000, 100)

set_abundance(dataset, seqs_to_update, new_abunds)
get_sequence_abundances(dataset)

}
