% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_abundances}
\alias{set_abundances}
\title{set_abundances}
\usage{
set_abundances(data, sequence_names, abundances, reason = "update")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{abundances}{2D vector of integers ([num_seqs][num_samples]) containing
the abundances of each sequence parsed by sample.}

\item{reason}{a string containing the trash tag to be applied to any sequences
set to 0 abundance. Default = "update".}
}
\description{
Set sequence abundances for an instance of the 'Dataset' class with
sample data
}
\examples{

sequence_names <- c("seq1", "seq1", "seq1", "seq2", "seq2", "seq2", "seq3",
                    "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4", "sample2", "sample3",
             "sample4", "sample2", "sample3", "sample4")
abundances <- c(250, 400, 500, 25, 40, 50, 25, 25, 4)

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, sequence_names, abundances, samples, "")

seqs_to_update <- c("seq4")
new_abunds <- list(c(20, 10, 4))

set_abundances(dataset, seqs_to_update, new_abunds)

}
