% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_dataset_name}
\alias{set_dataset_name}
\title{set_dataset_name}
\usage{
set_dataset_name(data, dataset_name)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{dataset_name, }{a string containing the desired name}
}
\description{
Set the name of a \link{dataset} object.
}
\examples{

data <- new_dataset("my_dataset", 2)
set_dataset_name(data, "new_dataset_name")

}
