% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_num_processors}
\alias{set_num_processors}
\title{set_num_processors}
\usage{
set_num_processors(data, processors)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{processors, }{a integer containing the desired number of processors}
}
\description{
Set the number of processors used to summarize an instance of the
 'Dataset' class.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
set_num_processors(dataset, 8)

}
