% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_sequences}
\alias{set_sequences}
\title{set_sequences}
\usage{
set_sequences(data, sequence_names, sequences, comments)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{sequences}{a vector of strings containing sequence nucleotide strings}

\item{comments}{a vector of strings containing sequence comments}
}
\description{
Set sequence nucleotide strings for an instance of the 'Dataset' class
}
\examples{

sequence_names <- c("seq1", "seq2", "seq3", "seq4")
sequences <- c("ATTGC", "ACTGC", "AGTGC", "TTTGC")

dataset <- new_dataset("my_dataset", 4)

add_sequences(dataset, sequence_names, "", "")
set_sequences(dataset, sequence_names, sequences, "")

}
