% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_dataframe}
\alias{sort_dataframe}
\title{sort_dataframe}
\usage{
sort_dataframe(data, order, named_col)
}
\arguments{
\item{data}{the data.frame to be sorted}

\item{order}{vector containing the order desired}

\item{named_col}{name of column in data.frame to match order}
}
\value{
sorted data.frame
}
\description{
Sort dataframe
}
\examples{

# sort results alphabetically

miseq <- miseq_sop_example()

sequence_names <- names(miseq)

fasta <- report(miseq, type = fasta)

sorted_fasta <- sort_dataframe(fasta,
  order = sort(sequence_names),
  named_col = "sequence_name"
)

}
