% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_at_char}
\alias{split_at_char}
\title{split_at_char}
\usage{
split_at_char(line, delim = ",")
}
\arguments{
\item{line, }{String containing data to split}

\item{delim, }{Character to split data by}
}
\value{
A vector of Strings
}
\description{
Split string by deliminating character
}
\examples{
(split_at_char("This is the first piece, and this is the second", ","))

}
