% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize_reports}
\alias{summarize_reports}
\title{summarize_reports}
\usage{
summarize_reports(report, count, processors)
}
\arguments{
\item{report}{DataFrame, containing all numeric columns}

\item{count}{NumericalVector, containing abundances of sequences}

\item{processors}{Integer, number of cores to use. Default = all available}
}
\value{
DataFrame with summary values
}
\description{
Summarizes a report
}
