% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarize the sequences data, custom reports, and scrapped data in a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
summary(data, type = "sequence", report_type = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{type}{string containing the type of data you want the number of.
Options include: "sequence", "report" and "scrap". Default = "sequence".}

\item{report_type}{string containing the report type you would summarized.
For example, the miseq_sop_example includes contigs assembly data and can be
accessed with report_type = "contigs_report". Default = NULL.}

\item{verbose}{boolean indicating whether or not you want progress messages.
Default = TRUE.}
}
\value{
data.frame
}
\description{
Summarize the sequences data, custom reports, and scrapped data in a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

miseq <- strollur::miseq_sop_example()

# To get the summary of your FASTA data
strollur::summary(data = miseq, type = "sequence")

# summarize contigs_report
strollur::summary(data = miseq,
                  type = "report", report_type = "contigs_report")

# remove sample 'F3D0' to produce a scrap report
strollur::xdev_remove_samples(data = miseq, samples = c("F3D0"))

# summarize FASTA data after removal of sample F3D0
strollur::summary(data = miseq, type = "sequence")

# summarize scrapped data -
# sequences and bins scrapped by removing the sample "F3D0"
strollur::summary(data = miseq, type = "scrap")

}
