% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_table_format.R
\name{taxonomy_table_format}
\alias{taxonomy_table_format}
\title{taxonomy_table_format}
\description{
strollur accepts two formats for classification data: \code{full} or \code{tidy}.
}
\details{
The \code{full} taxonomy table is a \code{data.frame} with two columns: \code{sequence_name}
or \code{bin_name}, and \code{taxonomy}. The \code{taxonomy} column contains the full
classification string.

\strong{The \code{full} taxonomy table format:}

\if{html}{\out{<div class="sourceCode">}}\preformatted{sequence_name
M00967_43_000000000-A3JHG_1_1102_26308_14496
M00967_43_000000000-A3JHG_1_1103_22654_26891

taxonomy
Bacteria(100);"Bacteroidetes"(100);"Bacteroidia"(100);"Bacteroidales"(100);
Bacteria(100);Firmicutes(99);Clostridia(99);Clostridiales(99);
}\if{html}{\out{</div>}}

The \code{tidy} taxonomy table is a \code{data.frame} with 4 columns: \code{sequence_name}
or \code{bin_name}, \code{taxonomy}, \code{level}, and \code{confidence}. For example:

\strong{The \code{tidy} taxonomy table format:}

\if{html}{\out{<div class="sourceCode">}}\preformatted{sequence_name                                    taxonomy
M00967_43_000000000-A3JHG_1_1102_26308_14496     Bacteria
M00967_43_000000000-A3JHG_1_1102_26308_14496     "Bacteroidetes"
M00967_43_000000000-A3JHG_1_1102_26308_14496     "Bacteroidia"
M00967_43_000000000-A3JHG_1_1102_26308_14496     "Bacteroidales"
M00967_43_000000000-A3JHG_1_1103_22654_26891     Bacteria
M00967_43_000000000-A3JHG_1_1103_22654_26891     Firmicutes
M00967_43_000000000-A3JHG_1_1103_22654_26891     Clostridia
M00967_43_000000000-A3JHG_1_1103_22654_26891     Clostridiales

level    confidence
1        100
2        100
3        100
4        100
1        100
2        99
3        99
4        99
}\if{html}{\out{</div>}}
}
\keyword{internal}
