% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{totals}
\alias{totals}
\title{totals}
\usage{
totals(data, type = "samples")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you want the totals of.
Options include: "samples", "treatments".
Default = "samples".}
}
\value{
data.frame
}
\description{
Find the total number of sequences in your samples or treatments in a
\link{dataset} object
}
\examples{

miseq <- miseq_sop_example()

# To get a table with the sample totals in the dataset
totals(data = miseq, type = "samples")

# To get a table with the treatment totals in the dataset
totals(data = miseq, type = "treatments")

}
