% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_taxonomy.R
\name{write_mothur_taxonomy}
\alias{write_mothur_taxonomy}
\title{write_mothur_taxonomy}
\usage{
write_mothur_taxonomy(data, filename = NULL)
}
\arguments{
\item{data}{A 'dataset' object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.taxonomy}
}
\value{
name of taxonomy file
}
\description{
Write a mothur formatted
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}
}
\examples{

miseq <- miseq_sop_example()
write_mothur_taxonomy(miseq)

}
