% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_bin_taxonomy_tidy}
\alias{xdev_assign_bin_taxonomy_tidy}
\title{Assign bin classifications to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_assign_bin_taxonomy_tidy(
  data,
  table,
  bin_type = "otu",
  reference = NULL,
  bin_name = "bin_name",
  level = "level",
  taxonomy = "taxonomy",
  confidence = "confidence",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing bin taxonomy assignments}

\item{bin_type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{bin_name}{a string containing the name of the column in 'table'
that contains the bin names. Default column name is 'bin_name'.}

\item{level}{a string containing the name of the column in 'table' that
contains the taxonomy levels. Default column name is 'level'.}

\item{taxonomy}{a string containing the name of the column in 'table' that
contains the bin taxonomies. Default column name is 'taxonomy'.}

\item{confidence}{a string containing the name of the column in 'table'
that contains the taxonomies confidence. Default column name is 'confidence'.}

\item{verbose}{a logical whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Assign bin classifications to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

bin_classifications <- readRDS(strollur_example("miseq_tidy_bin_taxonomy.rds"))
str(bin_classifications)

data <- strollur::read_mothur(otu_shared =
                                strollur_example("final.opti_mcc.shared"))

xdev_assign_bin_taxonomy_tidy(data, bin_classifications)

# With the reference parameter you can add information about the reference
# you used to classify your bins. You can also add references using the
# 'strollur::xdev_add_references' function.

reference <- new_reference("trainset9_032012.pds.zip", "9_032012",
              "classification_otu by mothur2 v1.0.0 using default options",
               "",
"https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset9_032012.pds.zip")

xdev_assign_bin_taxonomy_tidy(data, bin_classifications,
                                 reference = reference)

}
