% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_fastq_scores}
\alias{xdev_assign_sequence_fastq_scores}
\title{xdev_assign_sequence_fastq_scores}
\usage{
xdev_assign_sequence_fastq_scores(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_name",
  sequence = "sequence",
  quality_score = "quality_score",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing FASTQ data you are trying to add}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.}

\item{sequence}{a string containing the name of the column in 'table' that
contains the sequence nucleotide strings. Default column name is
'sequence'. (Optional)}

\item{quality_score}{a string containing the name of the column in 'table'
that contains the quality scores stored as vector\if{html}{\out{<int>}}. Default column
name is 'quality_score'.}

\item{verbose}{a logical whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Add FASTQ data to a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.
Scores are stored as vector\if{html}{\out{<int>}} for each sequence.
}
\examples{

table <- strollur::read_fastq(strollur_example("tiny.fastq.gz"))
data <- strollur::new_dataset("example")
strollur::xdev_assign_sequence_fastq_scores(data, table)
data

}
