% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_taxonomy}
\alias{xdev_assign_sequence_taxonomy}
\title{xdev_assign_sequence_taxonomy}
\usage{
xdev_assign_sequence_taxonomy(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_name",
  taxonomy = "taxonomy",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing sequence taxonomy assignments}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.}

\item{taxonomy}{a string containing the name of the column in 'table' that
contains the sequence taxonomies. Default column name is 'taxonomy'.}

\item{verbose}{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Assign sequence classifications to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object

Note, if you assign sequence taxonomies and assign bins, \code{strollur} will find
the consensus taxonomy for each bin for you.
}
\examples{

taxonomy <- strollur::read_mothur_taxonomy(strollur_example(
                        "final.taxonomy.gz"))

data <- strollur::new_dataset("my_dataset")

strollur::xdev_assign_sequence_taxonomy(data,
                                          table = taxonomy)

# With the reference parameter you can add information about the reference
# you used to classify your sequences. You can also add references using the
# 'strollur::xdev_add_references' function.

reference <- strollur::new_reference("trainset9_032012.pds.zip", "9_032012",
              "classification by mothur2 v1.0 using default options", "",
"https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset9_032012.pds.zip")

strollur::xdev_assign_sequence_taxonomy(data,
                                        table = taxonomy,
                                        reference = reference)

}
