% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_abundances_by_sample}
\alias{xdev_get_abundances_by_sample}
\title{Get the sequence abundance data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} parsed by sample}
\usage{
xdev_get_abundances_by_sample(data, samples = as.character(c()))
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{samples}{a vector of strings containing the names of the samples you
would like sequence names for. By default all samples are included.}
}
\value{
2D vector of float containing data requested parsed by sample.
}
\description{
Get the sequence abundance data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} parsed by sample
}
\examples{

data <- strollur::miseq_sop_example()

# To get the sequence names parsed by sample
abunds <- strollur::xdev_get_abundances_by_sample(data)

}
