% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_list_vector}
\alias{xdev_get_list_vector}
\title{Get vector of strings containing the sequences bin data}
\usage{
xdev_get_list_vector(data, type = "otu")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
vector of strings containing the names of the sequences in each bin
separated by commas
}
\description{
Get vector of strings containing the sequences bin data
}
\examples{

data <- strollur::miseq_sop_example()
strollur::xdev_get_list_vector(data)

}
