% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_sample_distances}
\alias{xdev_get_sample_distances}
\title{Get distances between samples in your `\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_get_sample_distances(data)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
Rcpp::DataFrame with 3 columns (sample1, sample2, distance)
}
\description{
Get distances between samples in your \code{strollur::strollur} object
}
\examples{

 data <- strollur::miseq_sop_example()
 strollur::xdev_get_sample_distances(data)

}
