% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_has_sequence_taxonomy}
\alias{xdev_has_sequence_taxonomy}
\title{xdev_has_sequence_taxonomy}
\usage{
xdev_has_sequence_taxonomy(data)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
boolean
}
\description{
Determine if a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object has sequence taxonomy assignments
}
\examples{

 data <- strollur::miseq_sop_example()
 strollur::xdev_has_sequence_taxonomy(data)

}
