% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_bins}
\alias{xdev_remove_bins}
\title{Remove bins from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_remove_bins(data, bin_names, trash_tags, bin_type = "otu")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}.}

\item{bin_names}{a vector of strings containing the names of the bins you
would like removed.}

\item{trash_tags}{a vector of strings containing the reasons you are
removing each bin}

\item{bin_type}{a string indicating the type of clusters.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
Designed with package integration in mind, the remove bins function allows
you to remove bins from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

  data <- strollur::new_dataset(dataset_name = "my_dataset")

  bin_names <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)

  strollur::xdev_assign_bins(data,
                   table = data.frame(bin_name = bin_names,
                                      abundance = abundances),
                   bin_type = "otu")

  strollur::count(data,
                    type = "bin",
                    bin_type = "otu")

  bins_to_remove <- c("bin1")
  trash_tag <- c("bad_bin")

  strollur::xdev_remove_bins(data,
                   bin_names = bins_to_remove,
                   trash_tags = trash_tag)

  strollur::count(data,
                    type = "bin",
                    bin_type = "otu")

}
