% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_lineages}
\alias{xdev_remove_lineages}
\title{xdev_remove_lineages}
\usage{
xdev_remove_lineages(data, contaminants, reason = "contaminant")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.}

\item{contaminants}{vector of strings containing the taxonomies you would
like to remove}

\item{reason}{a string containing reason you are removing the lineages.
Default = "contaminant".}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Designed with package integration in mind, the remove lineages function
allows you to remove contaminents from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur}
}
\examples{
data <- strollur::read_mothur(fasta = strollur_example("final.fasta.gz"),
                       count = strollur_example("final.count_table.gz"),
                       taxonomy = strollur_example("final.taxonomy.gz"),
                       design = strollur_example("mouse.time.design"),
                       otu_list = strollur_example("final.opti_mcc.list.gz"),
                       dataset_name = "miseq_sop")

contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
 "Eukaryota")

strollur::xdev_remove_lineages(data, contaminants = contaminants)

}
