% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_samples}
\alias{xdev_remove_samples}
\title{xdev_remove_samples}
\usage{
xdev_remove_samples(data, samples, reason = "remove_samples")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.}

\item{samples}{vector of strings containing the names of the samples to
remove.}

\item{reason, }{string containing the reason for removal.
Default = "remove_samples".}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Designed with package integration in mind, the remove samples function allows
you to remove samples from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- strollur::miseq_sop_example()

strollur::count(data, type = "sample")

# To remove samples 'F3D0' and 'F3D1'

strollur::xdev_remove_samples(data,
                               samples = c("F3D0", "F3D1"))

strollur::count(data, type = "sample")

}
