% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_bin_abundances}
\alias{xdev_set_bin_abundances}
\title{xdev_set_bin_abundances}
\usage{
xdev_set_bin_abundances(
  data,
  bin_names,
  abundances,
  type = "otu",
  reason = "update"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{bin_names, }{a vector strings containing of bin names to set the
abundances for.}

\item{abundances, }{2D vector ([num_seqs][num_samples]) containing the
abundances of each bin parsed by sample.}

\item{type}{a string indicating the type of clusters. Default = "otu".}

\item{reason, }{a string containing the trash tag to be applied to any bins
set to 0 abundance. Default = "update".}
}
\description{
Designed with package integration in mind, the set bin abundances function
allows you to change the abundances of bins in a \link{dataset} object
with sample data.
}
\examples{

  # For example sake, let's create a dataset with 3 bins:

  data <- new_dataset(dataset_name = "my_dataset")

  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5", "sample1", "sample3",
               "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)

  xdev_assign_bins(data = data, table = data.frame(bin_names = bin_ids,
                                              abundances = sample_abundances,
                                              samples = samples))

  # You can see bin1's abundances parsed by sample using abundance:
  abundance(data, type = "bins", by_sample = TRUE)

  # You can change bin1's abundances as follows:

  new_bin1_abunds <- list(c(10,50,0,0))
  bins <- c("bin1")

  xdev_set_bin_abundances(data = data,
                          bin_names = bins,
                          abundances = new_bin1_abunds)

  abundance(data, type = "bins", by_sample = TRUE)

}
