% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_num_processors}
\alias{xdev_set_num_processors}
\title{xdev_set_num_processors}
\usage{
xdev_set_num_processors(data, processors)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{processors, }{a integer containing the desired number of processors}
}
\description{
Designed with package integration in mind, set the number of processors used
 to summarize a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- new_dataset(dataset_name = "my_dataset")
xdev_set_num_processors(data = data, processors = 1)

}
