% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_summarize}
\alias{xdev_summarize}
\title{xdev_summarize}
\usage{
xdev_summarize(data, type = "sequences", report_type = NULL)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{type, }{string containing the type of data you want the number of.
Options include: "sequences", "reports" and "scrap". Default = "sequences".}

\item{report_type, }{string containing the report type you would summarized.
For example, the miseq_sop_example includes contigs assembly data and can be
accessed with report_type = "contigs_report". Default = NULL.}
}
\value{
data.frame()
}
\description{
Summarize the sequences data, custom reports, and scrapped data in a
\link{strollur} object
}
\examples{

 data <- miseq_sop_example()

 # summarize FASTA data
 xdev_summarize(data = data, type = "sequences")

 # summarize contigs_report
 xdev_summarize(data = data, type = "reports",
                 report_type = "contigs_report")

 # remove sample 'F3D0'
 xdev_remove_samples(data = data, samples = c("F3D0"))

 # summarize FASTA data after removal of sample F3D0
 xdev_summarize(data = data, type = "sequences")

 # summarize scrapped data
 xdev_summarize(data = data, type = "scrap")

}
