% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xint_deserialize_dobject}
\alias{xint_deserialize_dobject}
\title{xint_deserialize_dobject}
\usage{
xint_deserialize_dobject(data)
}
\arguments{
\item{data, }{a \link{strollur::strollur} object}
}
\description{
For internal use only, deserialize_dobject an instance of the C++ 'Dataset'
 class.
}
