% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xint_serialize_dobject}
\alias{xint_serialize_dobject}
\title{xint_serialize_dobject}
\usage{
xint_serialize_dobject(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\description{
For internal use only, xint_serialize_dobject an instance of the C++ 'Dataset'
class.
}
