% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abundance}
\alias{abundance}
\title{abundance}
\usage{
abundance(data, type = "sequence", by_sample = FALSE)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you want the number of.
Options include: "sequences", "bins".
Default = "sequences".}

\item{by_sample, }{Boolean. When by_sample is TRUE, the abundance data will
be parsed by sample. Default = FALSE.}
}
\value{
data.frame
}
\description{
Get a table containing the requested abundance data in a \link{dataset}
object
}
\examples{

miseq <- miseq_sop_example()

# To the total abundance for each sequence
abundance(data = miseq, type = "sequence")

# To the total abundance for each sequence parsed by sample
abundance(data = miseq, type = "sequence", by_sample = TRUE)

}
