% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\title{assign}
\usage{
assign(
  data,
  table,
  type = "bins",
  bin_type = "otu",
  table_names = list(sequence_name = "sequence_names", abundance = "abundances", sample =
    "samples", treatment = "treatments", taxonomy = "taxonomies", bin_name = "bin_names"),
  reference = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing the data you wish to assign}

\item{type, }{a string containing the type of data. Options include:
'sequence_abundance', 'sequence_taxonomy', 'bins',
 'bin_representatives', 'bin_taxonomy' and 'treatments'.
 Default = "bins".}

\item{bin_type, }{string containing the bin type you would like the number of
bins for. Default = "otu".}

\item{table_names, }{named list used to indicate the names of the columns in
the table. By default:

table_names <- list(sequence_name = "sequence_names",
                    abundance = "abundances",
                    sample = "samples",
                    treatment = "treatments",
                    taxonomy = "taxonomies",
                    bin_name = "bin_names")

In table_names, 'sequence_name' is a string containing the name of the column
in 'table' that contains the sequence names. Default column name is
'sequence_names'.

In table_names, 'abundance' is a string containing the name of the column in
'table' that contains the abundances. Default column name is 'abundances'.

In table_names, 'sample' is a string containing the name of the column in
'table' that contains the samples. Default column name is 'samples'.

In table_names, 'treatment' is a string containing the name of the
column in 'table' that contains the treatment names. Default column name is
 'treatments'.

In table_names, 'taxonomy' is a string containing the name of the
column in 'table' that contains the classifications. Default column name
is 'taxonomies'.

In table_names, 'bin_name' is a string containing the name of the
column in 'table' that contains the bin names. Default column name is
'bin_names'.}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{verbose, }{boolean indicating whether or not you want progress messages.
Default = TRUE.}
}
\value{
double - The number of items assigned
}
\description{
Assign sequence abundances, sequence classifications, bins, bin
representative sequences, bin classifications or treatments to a
\link{dataset} object
}
\examples{

# Assign sequence classifications

# create a new empty dataset named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

sequence_classifications <- read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

assign(
  data = data,
  table = sequence_classifications, type = "sequence_taxonomy"
)

# Assigning bins

# read mothur's otu list file into data.frame
otu_data <- read_mothur_list(list = strollur_example(
  "final.opti_mcc.list.gz"
))

# read mothur's asv list file into data.frame
asv_data <- read_mothur_list(list = strollur_example(
  "final.asv.list.gz"
))

# read mothur's phylotype list file into data.frame
phylo_data <- read_mothur_list(list = strollur_example(
  "final.tx.list.gz"
))

# read otu bin representative sequences into a data.frame
bin_reps <- readr::read_tsv(
  strollur_example(
    "otu_representative_sequences.tsv"
  ),
  show_col_types = FALSE
)

# assign 'otu' bins using sequence names
assign(data = data, table = otu_data, bin_type = "otu")

# assign 'asv' bins using sequence names
assign(data = data, table = asv_data, bin_type = "asv")

# assign 'phylotype' bins using sequence names
assign(data = data, table = phylo_data, bin_type = "phylotype")

# assign 'otu' bin representative sequences
assign(data = data, table = bin_reps, type = "bin_representatives")

# To assign abundance only bins

# create a new empty dataset named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# read mothur's shared file
otu_data <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# assign abundance only otus parsed by sample
assign(data = data, table = otu_data, bin_type = "otu")

# Assigning bin classifications

# read bin taxonomies
otu_data <- read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

# assign otu consensus taxonomies
assign(
  data = data, table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

# Assign treatments

sample_assignments <- readr::read_table(
  strollur_example("mouse.time.design"),
  col_names = TRUE, show_col_types = FALSE
)

assign(data = data, table = sample_assignments, type = "treatments")

}
