% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bin_representative_sequences}
\alias{assign_bin_representative_sequences}
\title{assign_bin_representative_sequences}
\usage{
assign_bin_representative_sequences(
  data,
  table,
  bin_type = "otu",
  reference = NULL,
  bin_name = "bin_names",
  sequence_name = "sequence_names"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing bin representative assignments}

\item{bin_type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{bin_name, }{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'bin_names'.}

\item{sequence_name}{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'sequence_names'.}
}
\value{
double containing the number of representative sequences assigned
}
\description{
Assign representative sequences to bins.
}
\examples{

  miseq <- miseq_sop_example()

  num_bins <- num(data = miseq, type = "bins", bin_type = "otu")

  # For examples sake, select first 531 sequences to be the representatives
  table <- data.frame(bin_names = get_bin_names(miseq, "otu"),
                      sequence_names = get_sequence_names(miseq)[1:num_bins]
                      )

  assign_bin_representative_sequences(data = miseq,
                                      table = table,
                                      bin_type = "otu")

}
