% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bin_taxonomy}
\alias{assign_bin_taxonomy}
\title{assign_bin_taxonomy}
\usage{
assign_bin_taxonomy(data, bin_names, taxonomies, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector of strings containing bin names}

\item{taxonomies}{a vector of strings containing bin classifications}

\item{type}{a string indicating the type of clusters. Default = "otu"}
}
\value{
double containing the number of bins assigned
}
\description{
Assign bin classifications to an instance of the 'Dataset' class.

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the concensus taxonomy for each bin for you.
}
\examples{

otu_data <- read_mothur_cons_taxonomy(rdataset_example(
                        "final.cons.taxonomy"))

dataset <- new_dataset("my_dataset", 4)
assign_bins(dataset, otu_data$bin_names, otu_data$abundances, "", "")
assign_bin_taxonomy(dataset, otu_data$bin_names, otu_data$taxonomies)

}
\seealso{
[sequence_data$assign_bin_taxonomy()]
}
