% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bin_taxonomy}
\alias{assign_bin_taxonomy}
\title{assign_bin_taxonomy}
\usage{
assign_bin_taxonomy(data, bin_names, taxonomies, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector of strings containing bin names}

\item{taxonomies}{a vector of strings containing bin classifications}

\item{type}{a string indicating the type of clusters. Default = "otu"}
}
\description{
Assign bin classifications to an instance of the 'Dataset' class.

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the concensus taxonomy for each bin for you.
}
\examples{

bin_names <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- new_dataset("my_dataset", 4)
assign_bins(dataset, bin_names, abunds, "", "")
assign_bin_taxonomy(dataset, bin_names, taxonomies)

}
\seealso{
[sequence_data$assign_bin_taxonomy()]
}
