% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bins}
\alias{assign_bins}
\title{assign_bins}
\usage{
assign_bins(data, bin_names, abundances, samples, sequence_names, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector strings containing of bin labels}

\item{abundances}{a vector containing bin abundances. Note: You must
provide either abundances or seq_ids.}

\item{samples}{a vector of strings containing sample assignments}

\item{sequence_names}{a vector of strings containing sequence names.
Note: You must provide either abundances or seq_ids.}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
double containing the number of bins assigned
}
\description{
Add bin assignments to an instance of the 'Dataset' class.
}
\examples{

  # To assign sequences to bins:

  dataset <- new_dataset("miseq_sop", 4)

  otu_data <- read_mothur_list(rdataset_example(
                            "final.opti_mcc.list"))
  assign_bins(dataset, otu_data$bin_names, 0, "", otu_data$sequence_names)

  # To add abundance only bin assignments:

  dataset <- new_dataset("miseq_sop", 4)

  otu_data <- read_mothur_rabund(rdataset_example(
                            "final.opti_mcc.rabund"))
  assign_bins(dataset, otu_data$bin_names, otu_data$abundances, "", "")

  # To add abundance bin assignments parsed by sample:

  dataset <- new_dataset("miseq_sop", 4)
  bin_table <- readr::read_tsv(rdataset_example(
                               "mothur2_bin_assignments_shared.tsv"))

  assign_bins(dataset, bin_table$bin_names, bin_table$abundances,
                 bin_table$samples, "")

  # To assign sequences to bins with their abundances parsed by sample:

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin1", "bin1", "bin1", "bin1", "bin1",
               "bin2", "bin2", "bin2",
               "bin3", "bin3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  assign_bins(dataset, bin_ids, abundances, samples, seq_ids, "otu")

}
\seealso{
[dataset$assign_bins()]
}
