% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_bins}
\alias{assign_bins}
\title{assign_bins}
\usage{
assign_bins(
  data,
  table,
  bin_type = "otu",
  reference = NULL,
  bin_name = "bin_names",
  abundance = "abundances",
  sample = "samples",
  sequence_name = "sequence_names"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing bin_data assignments}

\item{bin_type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{bin_name, }{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'bin_names'.}

\item{abundance, }{a string containing the name of the column in 'table'
that contains the bin abundances. Default column name is 'abundances'. Note:
 You must provide either abundances or sequence_names in the table.}

\item{sample, }{a string containing the name of the column in 'table' that
contains the sample names for datasets where the abundances are broken down
by sample. Default column name is 'samples'.}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.
Note: You must provide either abundances or sequence_names in the table.}
}
\value{
double containing the number of bins assigned
}
\description{
Add bin assignments to a \link{dataset} object
}
\examples{

  # To assign sequences to bins:

  data <- new_dataset("miseq_sop", 2)
  otu_data <- read_mothur_list(rdataset_example("final.opti_mcc.list"))

  assign_bins(data = data, table = otu_data, bin_type = "otu")

  # To add abundance only bin assignments:

  data <- new_dataset("miseq_sop", 2)
  otu_data <- read_mothur_rabund(rdataset_example("final.opti_mcc.rabund"))

  assign_bins(data = data, table = otu_data, bin_type = "otu")

  # To add abundance bin assignments parsed by sample:

  data <- new_dataset("miseq_sop", 2)
  otu_data <- readr::read_tsv(rdataset_example(
                                "mothur2_bin_assignments_shared.tsv"))

  assign_bins(data = data, table = otu_data, bin_type = "otu")

}
