% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_sequence_abundance}
\alias{assign_sequence_abundance}
\title{assign_sequence_abundance}
\usage{
assign_sequence_abundance(
  data,
  sequence_names,
  abundances,
  samples,
  treatments
)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{abundances}{a vector of integers containing sequence abundances}

\item{samples}{a vector of strings containing sample assignments}

\item{treatments}{a vector of strings containing treatment assignments}
}
\description{
Set sequence abundance and optionally assign sample and treatment data to
 an instance of the 'Dataset' class.
}
\examples{

sequence_names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500, 25, 40, 50, 25, 25, 4)

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, sequence_names, abundances, samples, "")

}
\seealso{
[sequence_data$assign_sequence_abundance()]
}
