% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_sequence_taxonomy}
\alias{assign_sequence_taxonomy}
\title{assign_sequence_taxonomy}
\usage{
assign_sequence_taxonomy(data, sequence_names, taxonomies)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{sequence_names}{a vector of strings containing sequence names}

\item{taxonomies}{a vector of strings containing sequence classifications}
}
\value{
double containing the number of sequences assigned
}
\description{
Assign sequence classifications to an instance of the 'Dataset' class.

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the concensus taxonomy for each bin for you.
}
\examples{

sequence_names <- c("seq1", "seq2", "seq3", "seq4")
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- new_dataset("my_dataset", 4)
assign_sequence_taxonomy(dataset, sequence_names, taxonomies)

}
\seealso{
[dataset$assign_sequence_taxonomy()]
}
