% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_treatments}
\alias{assign_treatments}
\title{assign_treatments}
\usage{
assign_treatments(data, samples, treatments)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{samples}{a vector of strings containing sample names}

\item{treatments}{a vector of strings containing treatment names}
}
\description{
Assign samples to treatments in an instance of the 'Dataset' class.
}
\examples{

names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)
treatments <- c("early", "early", "late")

dataset <- new_dataset("my_dataset", 4)
assign_sequence_abundance(dataset, names, abundances, samples, "")
assign_treatments(dataset, unique(samples), treatments)

}
\seealso{
[sequence_data$assign_treatments()]
}
