% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear}
\alias{clear}
\title{clear}
\usage{
clear(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\description{
Clear all data from an instance of the 'Dataset' class.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
clear(dataset)

}
