% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear}
\alias{clear}
\title{clear}
\usage{
clear(data, tags)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{tags}{a vector of strings containing the items you wish to clear.
Options are 'sequence_data' and 'bin_data'. By default, everything is
cleared.}
}
\description{
Clear all data from an instance of the 'Dataset' class.
}
\examples{
dataset <- miseq_sop_example()
clear(dataset$data, "")

}
