% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear}
\alias{clear}
\title{clear}
\usage{
clear(data)
}
\arguments{
\item{data, }{a \link{strollur::strollur} object}
}
\description{
Clear data from a \link{strollur::strollur} object
}
\examples{

data <- miseq_sop_example()
clear(data)

}
