% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{clone}
\usage{
clone(dataset)
}
\arguments{
\item{dataset}{a 'sequence_data' object}
}
\value{
A 'sequence_data' object
}
\description{
The clone function will create a deep copy of 'sequence_data' object. Note:
'sequence_data' is an R6 class with reference semantics and requires a deep
copy.
}
\examples{
# For dataset's including sequence data:

dataset <- read_mothur(
  fasta = rdataset_example("final.fasta"),
  count = rdataset_example("final.count_table"),
  taxonomy = rdataset_example("final.taxonomy"),
  design = rdataset_example("mouse.time.design"),
  list = rdataset_example("final.opti_mcc.list"),
  dataset_name = "miseq_sop"
)

copy_dataset <- clone(dataset)
copy_dataset

}
