% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_dataset}
\alias{copy_dataset}
\title{copy_dataset}
\usage{
copy_dataset(dataset)
}
\arguments{
\item{dataset}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Rcpp::XPtr<Dataset> pointer to an instance of the 'Dataset' c++
 class.
}
\description{
Create a pointer to an instance of the 'Dataset' c++ class.

The 'Dataset' class is the c++ implementation of the R6 'dataset'
object. This class allows package developers access to additional
functionality. 'Dataset' stores nucleotide sequences, abundance, sample and
treatment assignments, taxonomic classifications, asv / otu clusters. It
creates various reports and summaries. It is designed to facilitate data
transfer and access across multiple R packages.
}
\examples{

# to create a new dataset and allow for 4 processors during summary,
# run the following:

dataset1 <- new_dataset("soil", 4)

# to create a new dataset that is a copy of dataset1

dataset2 <- copy_dataset(dataset1)

}
\seealso{
[dataset$new()]
}
