% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copy_dataset}
\alias{copy_dataset}
\title{copy_dataset}
\usage{
copy_dataset(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Create a new \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object from an existing dataset.
}
\examples{

miseq <- miseq_sop_example()

# to create a new dataset that is a copy of miseq

data <- copy_dataset(miseq)

}
\seealso{
The 'new' method in the \href{https://mothur.org/strollur/reference/strollur.html}{strollur} class
}
