% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_dataset.R
\name{dataset}
\alias{dataset}
\title{dataset}
\description{
'dataset' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `dataset$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

data <- new_dataset("soil")


## ------------------------------------------------
## Method `dataset$add_sample_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))
 assign_bins(data, df)

 tree <- ape::read.tree(rdataset_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)


## ------------------------------------------------
## Method `dataset$add_sequence_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)


## ------------------------------------------------
## Method `dataset$get_bin_assignments`
## ------------------------------------------------

  data <- dataset$new("my_dataset")

  bin_table <- readr::read_tsv(rdataset_example(
                               "mothur2_bin_assignments_shared.tsv"),
                               show_col_types = FALSE)

  assign_bins(data, bin_table)

  shared <- data$get_bin_assignments()


## ------------------------------------------------
## Method `dataset$get_bin_names`
## ------------------------------------------------

  data <- miseq_sop_example()
  data$get_bin_names("otu")


## ------------------------------------------------
## Method `dataset$get_bin_types`
## ------------------------------------------------


data <- miseq_sop_example()
data$get_bin_types()


## ------------------------------------------------
## Method `dataset$get_metadata`
## ------------------------------------------------

  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add_report(data, metadata, "metadata")

  data$get_metadata()


## ------------------------------------------------
## Method `dataset$get_num_bins`
## ------------------------------------------------


  otu_data <- read_mothur_list(rdataset_example(
                            "final.opti_mcc.list"))

  data <- dataset$new("my_dataset")
  assign_bins(data, otu_data)
  data$get_num_bins()


## ------------------------------------------------
## Method `dataset$get_sample_tree`
## ------------------------------------------------


 tree <- ape::read.tree(rdataset_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")
 assign_bins(data, df)
 data$add_sample_tree(tree)
 data$get_sample_tree()


## ------------------------------------------------
## Method `dataset$get_sequence_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Rcpp::XPtr<Dataset> pointer to 'Dataset' c++ class. This
allows package developers an easy access point to the underlying C++ code
with additional functionality.}

\item{\code{raw}}{Rcpp::RawVector containing the serialized data of the
'Dataset' c++ class. This allows the load and save functions to work with
the class.}

\item{\code{sequence_tree}}{a tree that relates sequences to eachother}

\item{\code{sample_tree}}{a tree that relates samples to eachother}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataset-new}{\code{dataset$new()}}
\item \href{#method-dataset-print}{\code{dataset$print()}}
\item \href{#method-dataset-add_sample_tree}{\code{dataset$add_sample_tree()}}
\item \href{#method-dataset-add_sequence_tree}{\code{dataset$add_sequence_tree()}}
\item \href{#method-dataset-clear}{\code{dataset$clear()}}
\item \href{#method-dataset-get_bin_assignments}{\code{dataset$get_bin_assignments()}}
\item \href{#method-dataset-get_bin_names}{\code{dataset$get_bin_names()}}
\item \href{#method-dataset-get_bin_types}{\code{dataset$get_bin_types()}}
\item \href{#method-dataset-get_dataset_name}{\code{dataset$get_dataset_name()}}
\item \href{#method-dataset-get_metadata}{\code{dataset$get_metadata()}}
\item \href{#method-dataset-get_num_bins}{\code{dataset$get_num_bins()}}
\item \href{#method-dataset-get_num_sequences}{\code{dataset$get_num_sequences()}}
\item \href{#method-dataset-get_num_treatments}{\code{dataset$get_num_treatments()}}
\item \href{#method-dataset-get_num_samples}{\code{dataset$get_num_samples()}}
\item \href{#method-dataset-get_samples}{\code{dataset$get_samples()}}
\item \href{#method-dataset-get_sample_tree}{\code{dataset$get_sample_tree()}}
\item \href{#method-dataset-get_sequence_report}{\code{dataset$get_sequence_report()}}
\item \href{#method-dataset-get_summary}{\code{dataset$get_summary()}}
\item \href{#method-dataset-get_sequence_tree}{\code{dataset$get_sequence_tree()}}
\item \href{#method-dataset-get_treatments}{\code{dataset$get_treatments()}}
\item \href{#method-dataset-has_sample}{\code{dataset$has_sample()}}
\item \href{#method-dataset-clone}{\code{dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of dataset (optional)}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}

\item{\code{dataset}}{a `dataset` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `dataset` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

data <- new_dataset("soil")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-add_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-add_sample_tree}{}}}
\subsection{Method \code{add_sample_tree()}}{
Add phylo tree relating the samples in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$add_sample_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))
 assign_bins(data, df)

 tree <- ape::read.tree(rdataset_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-add_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-add_sequence_tree}{}}}
\subsection{Method \code{add_sequence_tree()}}{
Add phylo tree relating the sequences in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$add_sequence_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-clear"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-clear}{}}}
\subsection{Method \code{clear()}}{
Remove 'sample_tree', or 'sequence_tree' data from your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$clear(tags = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tags}}{a vector of strings containing the items you wish to clear.
Options are 'metadata', 'references', 'sequence_tree', 'sample_tree',
'alignment_report', 'contigs_assembly_report' and '"'chimera_report'.
By default, everything is cleared.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_bin_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_bin_assignments}{}}}
\subsection{Method \code{get_bin_assignments()}}{
Get data frame containing sequence bin assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_bin_assignments(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  data <- dataset$new("my_dataset")

  bin_table <- readr::read_tsv(rdataset_example(
                               "mothur2_bin_assignments_shared.tsv"),
                               show_col_types = FALSE)

  assign_bins(data, bin_table)

  shared <- data$get_bin_assignments()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_bin_names"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_bin_names}{}}}
\subsection{Method \code{get_bin_names()}}{
Get the names of the bins in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_bin_names(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of strings containing the bin names
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  data <- miseq_sop_example()
  data$get_bin_names("otu")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_bin_types"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_bin_types}{}}}
\subsection{Method \code{get_bin_types()}}{
Get bin table types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_bin_types()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of strings
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
data <- miseq_sop_example()
data$get_bin_types()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_dataset_name"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_dataset_name}{}}}
\subsection{Method \code{get_dataset_name()}}{
Get dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_dataset_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get data.frame containing metadata for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame()
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add_report(data, metadata, "metadata")

  data$get_metadata()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_num_bins"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_num_bins}{}}}
\subsection{Method \code{get_num_bins()}}{
Get the number of bins in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_num_bins(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  otu_data <- read_mothur_list(rdataset_example(
                            "final.opti_mcc.list"))

  data <- dataset$new("my_dataset")
  assign_bins(data, otu_data)
  data$get_num_bins()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_num_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_num_sequences}{}}}
\subsection{Method \code{get_num_sequences()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_num_sequences(distinct = FALSE, sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distinct}}{Boolean. When distinct is TRUE the number of unique
sequence is returned.}

\item{\code{sample}}{The name of the sample you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_num_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_num_treatments}{}}}
\subsection{Method \code{get_num_treatments()}}{
Get the number of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_num_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_num_samples}{}}}
\subsection{Method \code{get_num_samples()}}{
Get the number of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_samples"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_samples}{}}}
\subsection{Method \code{get_samples()}}{
Get names of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sample_tree}{}}}
\subsection{Method \code{get_sample_tree()}}{
Get phylo tree relating the samples in your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sample_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 tree <- ape::read.tree(rdataset_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")
 assign_bins(data, df)
 data$add_sample_tree(tree)
 data$get_sample_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get data.frame sequence report data. Sequence report data includes: start
positions, end positions, number of bases, number of ambiguous bases,
length of longest homopolymer, and the number of N's.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print summaries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sequence_tree}{}}}
\subsection{Method \code{get_sequence_tree()}}{
Get phylo tree relating the sequences in your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sequence_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_treatments}{}}}
\subsection{Method \code{get_treatments()}}{
Get names of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-has_sample"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-has_sample}{}}}
\subsection{Method \code{has_sample()}}{
Determine if a sample is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$has_sample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
