% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_dataset.R
\name{dataset}
\alias{dataset}
\title{dataset}
\description{
'dataset' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `dataset$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

data <- new_dataset("soil")


## ------------------------------------------------
## Method `dataset$add_sample_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 tree <- ape::read.tree(strollur_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)


## ------------------------------------------------
## Method `dataset$add_sequence_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)


## ------------------------------------------------
## Method `dataset$get_bin_types`
## ------------------------------------------------


data <- miseq_sop_example()
data$get_bin_types()


## ------------------------------------------------
## Method `dataset$get_metadata`
## ------------------------------------------------

  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(strollur_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add(data = data, table = metadata, type = "metadata")

  data$get_metadata()


## ------------------------------------------------
## Method `dataset$get_sample_tree`
## ------------------------------------------------


 tree <- ape::read.tree(strollur_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")

 # assign abundance 'otu' bins
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 data$add_sample_tree(tree)
 data$get_sample_tree()


## ------------------------------------------------
## Method `dataset$get_sequence_tree`
## ------------------------------------------------


 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Rcpp::XPtr<Dataset> pointer to 'Dataset' c++ class. This
allows package developers an easy access point to the underlying C++ code
with additional functionality.}

\item{\code{raw}}{Rcpp::RawVector containing the serialized data of the
'Dataset' c++ class. This allows the load and save functions to work with
the class.}

\item{\code{sequence_tree}}{a tree that relates sequences to eachother}

\item{\code{sample_tree}}{a tree that relates samples to eachother}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataset-new}{\code{dataset$new()}}
\item \href{#method-dataset-print}{\code{dataset$print()}}
\item \href{#method-dataset-add_sample_tree}{\code{dataset$add_sample_tree()}}
\item \href{#method-dataset-add_sequence_tree}{\code{dataset$add_sequence_tree()}}
\item \href{#method-dataset-clear}{\code{dataset$clear()}}
\item \href{#method-dataset-get_bin_types}{\code{dataset$get_bin_types()}}
\item \href{#method-dataset-get_metadata}{\code{dataset$get_metadata()}}
\item \href{#method-dataset-get_sample_tree}{\code{dataset$get_sample_tree()}}
\item \href{#method-dataset-get_sequence_report}{\code{dataset$get_sequence_report()}}
\item \href{#method-dataset-get_summary}{\code{dataset$get_summary()}}
\item \href{#method-dataset-get_sequence_tree}{\code{dataset$get_sequence_tree()}}
\item \href{#method-dataset-clone}{\code{dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of dataset (optional)}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}

\item{\code{dataset}}{a `dataset` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `dataset` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

data <- new_dataset("soil")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-add_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-add_sample_tree}{}}}
\subsection{Method \code{add_sample_tree()}}{
Add phylo tree relating the samples in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$add_sample_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 tree <- ape::read.tree(strollur_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-add_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-add_sequence_tree}{}}}
\subsection{Method \code{add_sequence_tree()}}{
Add phylo tree relating the sequences in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$add_sequence_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-clear"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-clear}{}}}
\subsection{Method \code{clear()}}{
Clear data from datasest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_bin_types"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_bin_types}{}}}
\subsection{Method \code{get_bin_types()}}{
Get bin table types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_bin_types()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of strings
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
data <- miseq_sop_example()
data$get_bin_types()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get data.frame containing metadata for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame()
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(strollur_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add(data = data, table = metadata, type = "metadata")

  data$get_metadata()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sample_tree}{}}}
\subsection{Method \code{get_sample_tree()}}{
Get phylo tree relating the samples in your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sample_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 tree <- ape::read.tree(strollur_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")

 # assign abundance 'otu' bins
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 data$add_sample_tree(tree)
 data$get_sample_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get data.frame sequence report data. Sequence report data includes: start
positions, end positions, number of bases, number of ambiguous bases,
length of longest homopolymer, and the number of N's.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print summaries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-get_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-get_sequence_tree}{}}}
\subsection{Method \code{get_sequence_tree()}}{
Get phylo tree relating the sequences in your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$get_sequence_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
