% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_exports.R
\name{Dataset}
\alias{Dataset}
\title{Dataset}
\description{
The 'Dataset' class is the backend C++ implementation of the R6
'sequence_data' object. Along with the rcpp_module this class allows
package developers access to additional functionality. 'Dataset' stores
nucleotide sequences, abundance, sample and treatment assignments,
taxonomic classifications, asv / otu clusters. It creates various reports
and summaries. It is designed to facilitate data transfer and access across
multiple R packages.
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructs a new Dataset object
\itemize{
\item \emph{name}, a string containing the dataset name
\item \emph{processors}, an integer indicating the number of processors
\item Returns: 'Dataset' object
\item \code{data <- new(Dataset, "my_dataset", 4)}
\item @seealso [sequence_data$new()]
}}

\item{\code{add_align_report}}{Add alignment data
\itemize{
\item \emph{names}, a vector of strings containing the sequence names
\item \emph{search_scores}, a vector of doubles containing kmer search score
used to find the closest reference sequence match
\item \emph{sim_scores}, a vector of doubles containing the similarity score
between the reference sequence and the query sequence
\item \emph{longest_inserts}, a vector of integers containing the longest
insert added to the query sequence
\item \code{data$add_align_report(names, search_scores, sim_scores,
 longest_inserts)}
}}

\item{\code{add_contigs_report}}{Add paired read assembly data
\itemize{
\item \emph{names}, a vector of strings containing the sequence names
\item \emph{overlap_lengths}, a vector of integers containing length of the
overlap between the paired reads
\item \emph{overlap_starts}, a vector of integers containing the start
position of the paired reads overlap
\item \emph{overlap_ends}, a vector of integers containing the end
position of the paired reads overlap
\item \emph{mismatches}, a vector of integers containing the number of
mismatches between the paired reads
\item \emph{expected_errors}, a vector of doubles containing the number of
expected errors based on the quality scores of the paired reads
\item \code{data$add_contigs_report(names, overlap_lengths, overlap_starts,
 overlap_ends, mismatches, expected_errors)}
}}

\item{\code{add_sequences}}{Add sequences to dataset

\itemize{
\item \emph{names}, a vector of strings containing the sequence names
\item \emph{sequences}, a vector of strings containing the sequence
nucleotides
\item \emph{comments}, a vector of strings containing the sequence comments
\item \code{data$add_sequences(names, sequences, comments)}
\item @seealso [sequence_data$add_sequences()]
}}

\item{\code{assign_otus}}{Add otu assignments to dataset
\itemize{
\item \emph{otu_ids}, a vector of strings containing the otu names
\item \emph{abundances}, a vector of integers containing the otus abundances
\item \emph{samples}, a vector of strings containing samples
\item \emph{seq_ids}, a vector of strings containing the sequence names
\item \code{data$assign_otus(otu_ids, abundances, samples, seq_ids)}
\item @seealso [sequence_data$assign_otus()]
}}

\item{\code{assign_otu_taxonomy}}{Assign otu classification
\itemize{
\item \emph{otu_ids}, a vector of strings containing the otu names
\item \emph{taxonomies}, a vector of strings containing the otu
classifications
\item \code{data$assign_otu_taxonomy(otu_ids, taxonomies)}
\item @seealso [sequence_data$assign_sequence_taxonomy()]
}}

\item{\code{assign_sequence_abundance}}{Add sequence abundance data with optional
 sample / treatment assignments
\itemize{
\item \emph{names}, a vector of strings containing the sequence names
\item \emph{abundances}, a vector of integers containing the sequence
abundances
\item \emph{samples}, a vector of strings containing samples
\item \emph{treatments}, a vector of strings containing treatments
\item \code{data$assign_sequence_abundance(names, abundances,
 samples, treatments)}
\item @seealso [sequence_data$assign_sequence_abundance()]
}}

\item{\code{assign_sequence_taxonomy}}{Assign sequence classification
\itemize{
\item \emph{names}, a vector of strings containing the sequence names
\item \emph{taxonomies}, a vector of strings containing the sequence
classifications
\item \code{data$assign_sequence_taxonomy(names, taxonomies)}
\item @seealso [sequence_data$assign_sequence_taxonomy()]
}}

\item{\code{assign_treatments}}{Assign samples to treatments
\itemize{
\item \emph{samples}, a vector of strings containing sample names
\item \emph{treatments}, a vector of strings containing treatment names
\item \code{data$assign_treatments(samples, treatments)}
\item @seealso [sequence_data$assign_treatments()]
}}

\item{\code{clear}}{Remove all data from dataset
\itemize{
\item \code{data$clear()}
\item @seealso [sequence_data$clear()]
}}

\item{\code{get_abundance}}{Get total abundance for each sequence
\itemize{
\item Returns a vector of integers contain the total abundance for each
sequence.
\item \code{data$get_abundance()}
}}

\item{\code{get_abundances}}{Get abundances for sequences parsed by sample
\itemize{
\item Returns a 2D vector of integers containing each sequence's
abundance data parsed by sample. abundances[num_names][num_samples]
\item \code{data$get_abundances()}
}}

\item{\code{get_align_report}}{Get data.frame containing the sequence alignment data
\itemize{
\item Returns a data.frame containing 4 columns: id, search_score(kmer search
 score used to find the closest reference sequence match), sim_score(the
 similarity score between the reference sequence and the query sequence) and
 longest_insert(the longest insert added to the query sequence).
\item \code{data$get_align_report()}
\item @seealso [sequence_data$get_align_report()]
}}

\item{\code{get_contigs_report}}{Get data.frame containing paired read assembly data
\itemize{
\item Returns a data.frame containing 8 columns: id, overlap_length(length
of the overlap between the paired reads), overlap_start(the start position
of the paired reads overlap), overlap_end(the end position of the paired
reads overlap), mismatches(the number of mismatches between the paired reads)
, num_n(the number of n's in the paired read) and ee(the number of expected
 errors based on the quality scores of paired reads).
\item \code{data$get_contigs_report()}
\item @seealso [sequence_data$get_contigs_report()]
}}

\item{\code{get_list}}{Get data.frame containing sequence otu assignments
\itemize{
\item Returns a data.frame containing 2 columns: otu_id and seq_id
\item \code{data$get_list()}
\item @seealso [sequence_data$get_list()]
}}

\item{\code{get_names}}{Get names of sequences in the dataset. You can provide a
 sample name to get the names of sequence in a given sample.
\itemize{
\item \emph{sample}, a string containing a sample name or ""
\item Returns a vector of strings containing names of sequences
\item \code{data$get_names("")}
\item @seealso [sequence_data$get_names()]
}}

\item{\code{get_names_by_sample}}{Get the names of the sequences parsed by sample
\itemize{
\item Returns a 2D vector of strings containing names of the sequences parsed
by sample. names[num_samples][num_seqs_in_sample]
\item \code{data$get_names_by_sample()}
}}

\item{\code{get_otu}}{Get names of sequences in a given otu separated by commas.
\itemize{
\item \emph{otu_id}, a string containing an otu name
\item Returns a string containing the names of sequences in a given otu
separated by commas. For example "seq1,seq4,seq10"
\item \code{data$get_otu(otu_id)}
}}

\item{\code{get_otu_abundance}}{Get total abundance of a given otu
\itemize{
\item \emph{otu_id}, a string containing an otu name
\item Returns an integer containing the otu's total abundance
\item \code{data$get_otu_abundance(otu_id)}
}}

\item{\code{get_otu_abundances}}{Get abundance of otu parsed by sample
\itemize{
\item \emph{otu_id}, a string containing an otu name
 \item Returns a vector of integers containing the otu's abundance data
  parsed by sample.
\item \code{data$get_otu_abundances(otu_id)}
}}

\item{\code{get_otu_taxonomy_report}}{Get data.frame containing the otu
classifications.
\itemize{
\item Returns a data.frame containing 3 or 4 columns: id, level, taxon and
 confidence(optional)
\item \code{data$get_otu_taxonomy_report()}
\item @seealso [sequence_data$get_otu_taxonomy_report()]
}}

\item{\code{get_rabund}}{Get data.frame containing otu abundance data
\itemize{
\item Returns a data.frame containing 2 columns: otu_id and abundance
\item \code{data$get_rabund()}
\item @seealso [sequence_data$get_rabund()]
}}

\item{\code{get_samples}}{Get names of samples in the dataset
\itemize{
\item Returns a vector of strings containing the sample names
\item \code{data$get_samples()}
\item @seealso [sequence_data$get_samples()]
}}

\item{\code{get_sample_totals}}{Get the abundances of the samples
\itemize{
 \item Returns a vector of integers containing abundance of each sample
\item \code{data$get_sample_totals()}
}}

\item{\code{get_scrap_report}}{Get list of data.frames containing sequences and otus
removed from the dataset
\itemize{
\item Returns a list of data.frames: 'sequence_scrap_report' (if sequence
data is available) and 'otu_scrap_report' (if otu data is available). Both
data frames include 2 columns: id and trash_code.
\item \code{data$get_scrap_report()}
\item @seealso [sequence_data$get_scrap_report()]
}}

\item{\code{get_sequences}}{Get nucleotide sequence data. You can provide a
 sample name to get nucleotide sequence data in a given sample.
\itemize{
\item \emph{sample}, a string containing a sample name or ""
\item Returns a vector of strings containing nucleotide sequence data
\item \code{data$get_sequences("")}
\item @seealso [sequence_data$get_sequences()]
}}

\item{\code{get_sequences_by_sample}}{Get the nucleotide sequence data parsed by
sample.
\itemize{
\item Returns a 2D vector of strings containing the nucleotide sequence data
 parsed by sample. sequences[num_samples][num_seqs_in_sample]
\item \code{data$get_sequences_by_sample()}
}}

\item{\code{get_sequence_abundance_table}}{Get data.frame containing sequence
abundance data
\itemize{
\item Returns a data.frame containing 2, 3 or 4 columns: id, abundance,
sample(if the dataset includes sample data), treatment(if the dataset
includes treatment data)
\item \code{data$get_sequence_abundance_table()}
\item @seealso [sequence_data$get_count_table()]
}}

\item{\code{get_sequence_report}}{Get data.frame sequence report data. Sequence
report data includes: start positions, end positions, number of bases,
number of ambiguous bases, length of longest homopolymer, and the number of
 N's.
\itemize{
\item Returns a data.frame containing 7 columns: id, start, end, length,
ambig, longest_homopolymer, and num_n.
\item \code{data$get_sequence_report()}
\item @seealso [sequence_data$get_sequence_report()]
}}

\item{\code{get_sequence_summary}}{Get list of data.frames containing summaries of
the sequence various sequence reports.
\itemize{
\item Returns a list of data.frames: 'sequence_summary', 'contigs_summary' (
if paired read assembly data is available), 'align_summary' (
if alignment data is available). The summaries are data.frames with 8 rows
and variable columns. The rows represent different summary cutoffs: minimum,
 2.5%-tile, 25%-tile, median, 75%-tile, 97.5%-tile, maximum, and mean.
\item \code{data$get_sequence_summary()}
\item @seealso [sequence_data$get_sequence_summary()]
}}

\item{\code{get_sequence_taxonomy_report}}{Get data.frame containing the sequence
classifications.
\itemize{
\item Returns a data.frame containing 3 or 4 columns: id, level, taxon and
 confidence(optional)
\item \code{data$get_sequence_taxonomy_report()}
\item @seealso [sequence_data$get_sequence_taxonomy_report()]
}}

\item{\code{get_shared}}{Get data.frame containing otu abundance data by sample
\itemize{
\item Returns a data.frame containing 3 columns: otu_id, sample and abundance
\item \code{data$get_shared()}
\item @seealso [sequence_data$get_shared()]
}}

\item{\code{get_treatments}}{Get names of treatments in the dataset
\itemize{
\item Returns a vector of strings containing the treatment names
\item \code{data$get_treatments()}
\item @seealso [sequence_data$get_treatments()]
}}

\item{\code{get_total}}{Get the total number of sequences in the dataset. You can
 provide a sample name to get the total number of sequences in a given
 sample.
\itemize{
\item \emph{sample}, a string containing a sample name or ""
\item Returns an integer containing the total abundance of the dataset or
sample
\item \code{data$get_total("")}
\item @seealso [sequence_data$get_num_sequences()]
}}

\item{\code{get_treatment_totals}}{Get the abundances of the treatments
\itemize{
 \item Returns a vector of integers containing abundance of each treatment
\item \code{data$get_treatment_totals()}
}}

\item{\code{get_unique_total}}{Get the number of unique sequences in the dataset.
 You can provide a sample name to get the number of unique sequences in a
 given sample.
\itemize{
\item \emph{sample}, a string containing a sample name or ""
\item Returns an integer containing the number of unique sequences or sample
\item \code{data$get_unique_total("")}
\item @seealso [sequence_data$get_num_sequences()]
}}

\item{\code{has_sample}}{Determine if dataset contains sample
\itemize{
\item \emph{sample}, a string containing a sample name
\item Returns a boolean
\item \code{data$has_sample(sample)}
\item @seealso [sequence_data$has_sample()]
}}

\item{\code{merge_otus}}{Merge otus combining abundance data and storing
it in the first otu listed.
\itemize{
\item \emph{otu_ids}, a vector of strings containing otu names
\item \emph{reason}, a string containing the reason for merging
\item \code{data$merge_otus(otu_ids, reason)}
}}

\item{\code{merge_sequences}}{Merge sequences combining abundance data and storing
it in the first sequence listed.
\itemize{
\item \emph{names}, a vector of strings containing sequence names
\item \emph{reason}, a string containing the reason for merging
\item \code{data$merge_sequences(names, reason)}
}}

\item{\code{remove_lineages}}{Remove sequences / otus classified to given
 taxonomies. You can also use confidence scores to remove sequences failing
 below a confidence threshold. For example "Bacteria;Firmicutes;" will remove
 all sequences classified to "Bacteria;Firmicutes;", but
 "Bacteria(100);Firmicutes(85);" will only remove sequences classified to
 "Bacteria;Firmicutes;" with confidence scores below 100 and 85 respectively.
\itemize{
\item \emph{taxonomies}, a vector of strings containing taxonomies
\item \emph{reason}, a string containing the reason for removing the
 contaminant.
\item \code{data$remove_lineages(taxonomies, reason)}
}}

\item{\code{remove_otus}}{Remove otus from dataset. Any sequences assigned to the
 otus being removed will also be removed.
\itemize{
\item \emph{otu_ids}, a vector of strings containing otu names
\item \emph{trash_codes}, a vector of strings containing the reasons for
 removing the otus.
\item \code{data$remove_otus(otu_ids, trash_codes)}
}}

\item{\code{remove_sequences}}{Remove sequences from dataset. If all sequences in an
otu are removed then the otu will also be removed.
\itemize{
\item \emph{names}, a vector of strings containing sequence names
\item \emph{trash_codes}, a vector of strings containing the reasons for
 removing the sequences
\item \code{data$remove_sequences(names, trash_codes)}
}}

\item{\code{set_abundance}}{Set sequence abundance data for datasets without sample
 data
\itemize{
\item \emph{names}, a vector of strings containing sequence names
\item \emph{abundances}, a vector of integers containing each sequence's
total abundance
\item \emph{reason}, a string containing the reason you are setting the
abundances. This will be used as a trash code if you set the abundance of a
sequence to 0.
\item \code{data$set_abundance(names, abundances, reason)}
}}

\item{\code{set_abundances}}{Set sequence abundances for datasets with multiple
samples
\itemize{
\item \emph{names}, a vector of strings containing sequence names
\item \emph{abundances}, a 2D vector of integers containing each sequence's
abundance data parsed by sample. abundances[num_names][num_samples]
\item \emph{reason}, a string containing the reason you are setting the
abundances. This will be used as a trash code if the sum of the abundances
for a sequence is 0.
\item \code{data$set_abundances(names, abundances, reason)}
}}

\item{\code{set_otu_abundance}}{Set otu abundance data for datasets without sample
 data
\itemize{
\item \emph{otu_ids}, a vector of strings containing otu names
\item \emph{abundances}, a vector of integers containing each otu's
total abundance
\item \emph{reason}, a string containing the reason you are setting the
abundances. This will be used as a trash code if you set the abundance of an
otu to 0.
\item \code{data$set_otu_abundance(otu_ids, abundances, reason)}
}}

\item{\code{set_otu_abundances}}{Set otu abundances for datasets with multiple
samples
\itemize{
\item \emph{otu_ids}, a vector of strings containing otu names
\item \emph{abundances}, a 2D vector of integers containing each otu's
abundance data parsed by sample. abundances[num_otu_ids][num_samples]
\item \emph{reason}, a string containing the reason you are setting the
abundances. This will be used as a trash code if the sum of the abundances
for an otu is 0.
\item \code{data$set_otu_abundances(otu_ids, abundances, reason)}
}}

\item{\code{set_sequences}}{Set nucleotide sequence data for sequences.
\itemize{
\item \emph{names}, a vector of strings containing sequence names
\item \emph{sequences}, a vector of strings containing the nucleotide
 sequence data
\item \emph{comments}, a vector of strings containing comments about the
 sequences
\item \code{data$set_sequences(names, sequences, comments)}
}}

\item{\code{dataset_name}}{a string containing the dataset name}

\item{\code{is_aligned}}{boolean indicating the dataset's alignment status}

\item{\code{num_unique}}{integer containing the number of distinct sequences in the
dataset}

\item{\code{has_align_data}}{boolean indicating the dataset includes an alignment
report}

\item{\code{has_contigs_data}}{boolean indicating the dataset includes a contigs
assembly report}
}}

\seealso{
[sequence_data]
}
