% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dataset.R
\name{export_dataset}
\alias{export_dataset}
\title{Create a human readable table containing all data from a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.}
\usage{
export_dataset(data)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
list, containing the data in the strollur::strollur object
}
\description{
Export all data from a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.
}
\examples{

dataset <- strollur::new_dataset("my_dataset")
strollur::export_dataset(dataset)

}
