% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{export_dataset}
\alias{export_dataset}
\title{export_dataset}
\usage{
export_dataset(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Rcpp::List, containing the data in the 'Dataset
}
\description{
Export all data from an instance of the 'Dataset' class.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
export_dataset(dataset)

}
