% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{export_dataset}
\alias{export_dataset}
\title{export_dataset}
\usage{
export_dataset(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
Rcpp::List, containing the data in the 'Dataset
}
\description{
Export all data from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.
}
\examples{

dataset <- new_dataset("my_dataset", 2)
export_dataset(dataset)

}
