% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{export_dataset}
\alias{export_dataset}
\title{export_dataset}
\usage{
export_dataset(data, tags = as.character(c()))
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{tags}{a vector of strings containing the items you wish to export.
Options are 'sequence_data' and 'bin_data'. By default, everything is
 exported.}
}
\value{
Rcpp::List, containing the data in the 'Dataset
}
\description{
Export all data from an instance of the 'Dataset' class.
}
\examples{

dataset <- new_dataset("my_dataset", 2)
export_dataset(dataset, c(""))

}
