% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin}
\alias{get_bin}
\title{get_bin}
\usage{
get_bin(data, bin_name, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_name, }{string containing the bin name}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\value{
String, containing names of the sequences in a given bin
}
\description{
Get the names of the sequences in a given bin in an instance of the
'Dataset' class.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
dataset <- new_dataset("my_dataset", 4)
assign_bins(dataset, bin_ids, 0, "", seq_ids)
get_bin(dataset, "bin1")

}
