% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_abundances}
\alias{get_bin_abundances}
\title{get_bin_abundances}
\usage{
get_bin_abundances(data, bin_name, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{bin_name, }{string containing the bin name}

\item{type, }{string indicating the type of clusters. Default = "otu".}
}
\value{
vector containing the abundance of a given bin parsed by sample
}
\description{
Get the abundance of a given bin parsed by sample in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_bin_abundances(data, "Otu005", "otu")

}
