% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_assignments}
\alias{get_bin_assignments}
\title{get_bin_assignments}
\usage{
get_bin_assignments(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
data.frame containing 2, 3 or 4 columns: bin_names, abundances,
samples (if assigned), and treatments (if assigned)
}
\description{
Get data.frame containing bin abundance data of a \link{dataset} object
}
\examples{

  data <- miseq_sop_example()
  get_bin_assignments(data, "otu")

}
