% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_names}
\alias{get_bin_names}
\title{get_bin_abundances}
\usage{
get_bin_names(data, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{type, }{string indicating the type of clusters. Default = "otu".}
}
\value{
vector containing the names of bins
}
\description{
Get the names of the bins in an instance of the  'Dataset' class.
}
\examples{

  miseq <- miseq_sop_example()
  get_bin_names(miseq$data, "phylotype")

}
