% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_representative_sequences}
\alias{get_bin_representative_sequences}
\title{get_bin_representative_sequences}
\usage{
get_bin_representative_sequences(data, bin_type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{bin_type, }{string indicating the type of clusters. Default = "otu".}
}
\value{
data.frame containing 2 columns representative_names and
 representative_sequences
}
\description{
Get the representative sequences of the bins in a \link{dataset} object
}
\examples{

  miseq <- miseq_sop_example()

  num_bins <- count(data = miseq, type = "bins", bin_type = "otu")

  # For examples sake, select first 531 sequences to be the representatives
  table <- data.frame(bin_names = names(data = miseq,
                                       type = "bins",
                                       bin_type = "otu"),
                      sequence_names = names(data = miseq,
                                            type = "sequences")[1:num_bins]
                      )

  assign(data = miseq, table = table,
         type = "bin_representatives", bin_type = "otu")


  get_bin_representative_sequences(data = miseq, bin_type= "otu")

}
