% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_taxonomy_report}
\alias{get_bin_taxonomy_report}
\title{get_bin_taxonomy_report}
\usage{
get_bin_taxonomy_report(data, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string indicating the type of bin clusters. Default = "otu".}
}
\value{
data.frame
}
\description{
Get the bin classifications of a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_bin_taxonomy_report(data, "otu")

}
