% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_types}
\alias{get_bin_types}
\title{Get bin table types of a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
get_bin_types(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
}
\value{
vector of strings
}
\description{
Get bin table types of a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

data <- miseq_sop_example()
get_bin_types(data)

}
