% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_bin_types}
\alias{get_bin_types}
\title{get_bin_types}
\usage{
get_bin_types(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
vector of strings
}
\description{
Get bin table types of an instance of the 'Dataset' class.
}
\examples{

data <- miseq_sop_example()
get_bin_types(data$data)

}
