% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dataset_name}
\alias{get_dataset_name}
\title{get_dataset_name}
\usage{
get_dataset_name(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
String, containing the name of the dataset
}
\description{
Get the name of a \link{dataset} object
}
\examples{
data <- new_dataset("my_dataset", 2)
get_dataset_name(data)

}
