% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_bins}
\alias{get_num_bins}
\title{get_num_bins}
\usage{
get_num_bins(data, type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\value{
Integer, the number of bins of a specific type in an instance of
the 'Dataset' class.
}
\description{
Get the number of bins of a specific type in an instance of the 'Dataset'
 class.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
bin_ids <- c("bin1", "bin2", "bin3")
abundances <- c(110, 525, 80)
assign_bins(dataset, bin_ids, abundances, "", "")
get_num_bins(dataset)

}
