% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_processors}
\alias{get_num_processors}
\title{get_num_processors}
\usage{
get_num_processors(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Integer, containing number of processors
}
\description{
Get the number of processors used to summarize an instance of the
 'Dataset' class.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
get_num_processors(dataset)

}
