% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_samples}
\alias{get_num_samples}
\title{get_num_samples}
\usage{
get_num_samples(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
Integer, the number of samples in a \link{dataset} object
}
\description{
Get the number of samples in a \link{dataset} object
}
\examples{

data <- new_dataset()
bin_table <- readr::read_tsv(rdataset_example(
                              "mothur2_bin_assignments_shared.tsv"),
                              show_col_types = FALSE)
assign_bins(data, bin_table)
get_num_samples(data)

}
