% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_sequences}
\alias{get_num_sequences}
\title{get_num_sequences}
\usage{
get_num_sequences(data, distinct = FALSE, sample = "")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{distinct}{Boolean. When distinct is TRUE the number of unique
sequence is returned.}

\item{sample, }{string containing the name of the sample you want number of
sequences for.}
}
\value{
An integer
}
\description{
Get the number of sequences in an instance of the 'Dataset' class.
}
