% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_treatments}
\alias{get_num_treatments}
\title{get_num_treatments}
\usage{
get_num_treatments(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
Integer, the number of treatments in a \link{dataset} object
}
\description{
Get the number of treatments in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_num_treatments(data)

}
