% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sample_treatment_assignments}
\alias{get_sample_treatment_assignments}
\title{get_sample_treatment_assignments}
\usage{
get_sample_treatment_assignments(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
2 column data.frame containing the sample treatment assignments in a
 \link{dataset} object
}
\description{
Get treatment assignments for samples in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sample_treatment_assignments(data)

}
