% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_scrap_report}
\alias{get_scrap_report}
\title{get_scrap_report}
\usage{
get_scrap_report(data, type = "sequence")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type}{a string indicating the type of scrap report you would like.
Default = 'sequence'.}
}
\value{
data.frame containing sequences or bins removed from the
\link{dataset} object during analysis
}
\description{
Get a scrap report containing sequences and bins eliminated from a
\link{dataset} object
}
\examples{

  data <- miseq_sop_example()
  remove_bins(data, c("Otu005"), c("bad_bin"))

  sequence_scrap_report <- get_scrap_report(data, "sequence")
  otu_scrap_report <- get_scrap_report(data, "bin")

}
