% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_abundance_table}
\alias{get_sequence_abundance_table}
\title{get_sequence_abundance_table}
\usage{
get_sequence_abundance_table(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
data.frame
}
\description{
Get the abundances of each sequence in an instance of the 'Dataset' class
parsed by sample.
}
\examples{

dataset <- new_dataset("my_dataset", 4)
bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
samples <- c("sample1", "sample2", "sample5", "sample1", "sample3", "sample1")
treatments <- c("early", "early", "late", "early", "late", "early")
sample_abundances <- c(10, 100, 1, 500, 25, 80)
assign_bins(dataset, bin_ids, sample_abundances, samples, treatments)

get_sequence_abundance_table(dataset)

}
