% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_abundances}
\alias{get_sequence_abundances}
\title{get_sequence_abundances}
\usage{
get_sequence_abundances(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
vector containing the total abundance for each sequence a
\link{dataset} object
}
\description{
Get the total abundance for each sequence in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sequence_abundances(data)

}
