% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_abundances_by_sample}
\alias{get_sequence_abundances_by_sample}
\title{get_sequence_abundances_by_sample}
\usage{
get_sequence_abundances_by_sample(data)
}
\arguments{
\item{data, }{a \link{dataset} object}
}
\value{
2D vector ([num_seqs][num_samples]) containing the
abundances of each sequence in an instance of the 'Dataset' class parsed by
sample.
}
\description{
Get the abundances of each sequence in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
get_sequence_abundances_by_sample(data)

}
