% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_report}
\alias{get_sequence_report}
\title{get_sequence_report}
\usage{
get_sequence_report(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
data.frame
}
\description{
Get sequence report data: starts, ends, lengths, ambigs, longest
homopolymers and numns.
}
\examples{

 dataset <- new_dataset("miseq_sop", 4)
 sequences <- read_fasta(rdataset_example("final.fasta"))
 add_sequences(dataset, sequences$names, sequences$sequences, "")
 get_sequence_report(dataset)

}
