% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequence_summary}
\alias{get_sequence_summary}
\title{get_sequence_summary}
\usage{
get_sequence_summary(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
data.frame
}
\description{
Get a summary of the sequence report data, as well as reports of containing
scrapped data.
}
\examples{

 dataset <- new_dataset("miseq_sop", 4)
 sequences <- read_fasta(rdataset_example("final.fasta"))
 add_sequences(dataset, sequences$sequence_names, sequences$sequences, "")
 get_sequence_summary(dataset)

}
