% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sequences}
\alias{get_sequences}
\title{get_sequences}
\usage{
get_sequences(data, sample = "")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{sample, }{a string containing the name of the sample you
would like sequence names for. For all samples in dataset, sample = "".}
}
\value{
vector of string containing nucleotide strings of the sequences in
a \link{dataset} object
}
\description{
Get the nucleotide strings for each sequence in a \link{dataset} object
}
\examples{

 data <- miseq_sop_example()
 get_sequences(data)

}
