% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_treatments}
\alias{get_treatments}
\title{get_treatments}
\usage{
get_treatments(data)
}
\arguments{
\item{data, }{a \link{dataset} object.}
}
\value{
vector of strings containing the names of the treatments in a
\link{dataset} object
}
\description{
Get the treatments in a \link{dataset} object
}
\examples{

 data <- miseq_sop_example()
 get_treatments(data)

}
