% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_sample}
\alias{has_sample}
\title{has_sample}
\usage{
has_sample(data, sample)
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{sample}{a string containing the name of a sample.}
}
\value{
boolean indicating whether the dataset has a given sample
}
\description{
Determine if a given sample is in a \link{dataset} object
}
\examples{

data <- miseq_sop_example()
has_sample(data, "F3D0")
has_sample(data, "not a valid sample")

}
