% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_sequence_strings}
\alias{has_sequence_strings}
\title{has_sequence_strings}
\usage{
has_sequence_strings(data)
}
\arguments{
\item{data, }{a \link{dataset} object.}
}
\value{
boolean indicating whether the dataset has sequence nucleotide
strings.
}
\description{
Determine if a \link{dataset} object contains sequence nucleotide strings.
}
\examples{

data <- miseq_sop_example()
has_sequence_strings(data)

}
