% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_sequence_strings}
\alias{has_sequence_strings}
\title{has_sequence_strings}
\usage{
has_sequence_strings(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
boolean indicating whether the dataset has sequence nucleotide
strings.
}
\description{
Determine if an instance of the 'Dataset' class contains sequence
nucleotide strings.
}
\examples{

seq_names <- c("seq1", "seq2", "seq3", "seq4")
sequences <- c("ATTGC", "ACTGC", "AGTGC", "TTTGC")

dataset <- new_dataset("my_dataset", 4)
add_sequences(dataset, seq_names, sequences, "")
has_sequence_strings(dataset)

}
