% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(table, tags = NULL)
}
\arguments{
\item{table}{a table containing the data from a 'sequence_data' object. You
can create the table using 'export(dataset)'.}

\item{tags}{a vector of strings containing the items you wish to export.
Options are 'sequence_data', 'bin_data', 'metadata',
'references', 'sequence_tree', 'sample_tree', 'alignment_report',
'contigs_assembly_report' and 'chimera_report'. By default, everything is
 imported.}
}
\value{
A 'sequence_data' object
}
\description{
The import function will create a 'sequence_data' object from the exported
table of a 'sequence_data' object.
}
\examples{

miseq <- miseq_sop_example()
dataset <- import(export(miseq))
dataset

}
\seealso{
[sequence_data$export()]
}
