% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{import_dataset}
\usage{
import_dataset(table)
}
\arguments{
\item{table}{a table containing the data from a \link{strollur::strollur}
object. You can create the table using 'export(data)'.}
}
\value{
a \link{strollur::strollur} object
}
\description{
The import_dataset function will create a
\link{strollur::strollur} object from the exported table of a
\link{strollur::strollur} object.
}
\examples{

miseq <- miseq_sop_example()
data <- import_dataset(export_dataset(miseq))
data

}
\seealso{
[dataset$export()]
}
