% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{Import strollur object from exported data.frame.}
\usage{
import_dataset(table)
}
\arguments{
\item{table}{a table containing the data from a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur}
object. You can create the table using 'export(data)'.}
}
\value{
a
  \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
The import_dataset function will create a
  \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
  from the exported table of a
  \href{https://mothur.org/strollur/reference/strollur.html}{strollur}
  object.
}
\examples{

miseq <- miseq_sop_example()
data <- import_dataset(export_dataset(miseq))
data

}
\seealso{
[export_dataset()]
}
