% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{import_dataset}
\usage{
import_dataset(table, tags = NULL)
}
\arguments{
\item{table}{a table containing the data from a \link{dataset} object. You
can create the table using 'export(dataset)'.}

\item{tags}{a vector of strings containing the items you wish to export.
Options are 'sequence_data', 'bin_data', 'metadata',
'references', 'sequence_tree', 'sample_tree', and 'reports'.
By default, everything is imported.}
}
\value{
a \link{dataset} object
}
\description{
The import_dataset function will create a \link{dataset} object from the
exported table of a \link{dataset} object.
}
\examples{

miseq <- miseq_sop_example()
data <- import_dataset(export_dataset(miseq))
data

}
\seealso{
[dataset$export()]
}
