% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_aligned}
\alias{is_aligned}
\title{is_aligned}
\usage{
is_aligned(data)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}
}
\value{
Boolean
}
\description{
Determine if the instance of the 'Dataset' class contains aligned sequences.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
is_aligned(dataset)

}
