% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_aligned}
\alias{is_aligned}
\title{is_aligned}
\usage{
is_aligned(data)
}
\arguments{
\item{data, }{a \link{strollur} object}
}
\value{
Boolean
}
\description{
Determine if a \link{strollur} object contains aligned sequences.
}
\examples{

dataset <- miseq_sop_example()
is_aligned(dataset)

}
