% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_equal.R
\name{is_equal}
\alias{is_equal}
\title{is_equal}
\usage{
is_equal(data, data2)
}
\arguments{
\item{data, }{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{data2, }{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
}
\value{
a logical
}
\description{
Determine if two
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}s
are equal.
}
\examples{

miseq <- strollur::miseq_sop_example()

data <- strollur::copy_dataset(miseq)

strollur::is_equal(miseq, data)

}
