% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{load_dataset}
\usage{
load_dataset(file)
}
\arguments{
\item{file}{a string containing the .rds file name.}
}
\value{
a \link{dataset} object
}
\description{
The load_dataset function will create a \link{dataset} object from an RDS
file.
}
\examples{

dataset <- load_dataset(rdataset_example("miseq_sop.rds"))
dataset

}
