% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_dataset}
\alias{load_dataset}
\title{load_dataset}
\usage{
load_dataset(data, raw)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{raw}{a Rcpp::RawVector containing the serialize 'Dataset' class data.}
}
\description{
Load an instance of the 'Dataset' class from serialized raw data.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
raw <- serialize(dataset)
load_dataset(dataset, raw)

}
