% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{load_dataset}
\usage{
load_dataset(file)
}
\arguments{
\item{file}{a string containing the .rds file name.}
}
\value{
a
  \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
The load_dataset function will create a
  \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
  from an RDS file.
}
\examples{

data <- load_dataset(strollur_example("miseq_sop.rds"))
data

}
