% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_bins}
\alias{merge_bins}
\title{merge_bins}
\usage{
merge_bins(data, bin_names, reason = "merged", type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector of strings containing the names of the bins you
would like merge. The resulting merged bin will be stored in the first
bin_id in the vector.}

\item{reason}{a string indicating why you are merging bins}

\item{type}{a string indicating the type of bin clusters. Default = "otu"}
}
\description{
Merge bins in an instance of the 'Dataset' class.
}
\examples{

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  assign_bins(dataset, bin_ids, abundances, "", "", "otu")
  bins_to_merge <- c("bin1", "bin3")
  merge_bins(dataset, bins_to_merge)

}
