% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_sequences}
\alias{merge_sequences}
\title{merge_sequences}
\usage{
merge_sequences(data, sequence_names, reason = "merged")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{sequence_names, }{a vector of strings containing the names of the
sequences you would like merge. The resulting merged sequence will be stored
in the first sequence name in the vector.}

\item{reason}{a string indicating why you are merging sequences.
Default = "merged"}
}
\description{
Designed with package integration in mind, the merge sequences function
allows you to merge sequences in a \link{dataset} object.
}
\examples{

data <- miseq_sop_example()

get_num_sequences(data)

# For the sake of example let's merge the first 3 sequences from
# miseq_sop_example:

seqs_to_merge <- c("M00967_43_000000000-A3JHG_1_2101_16474_12783",
                   "M00967_43_000000000-A3JHG_1_1113_12711_3318",
                   "M00967_43_000000000-A3JHG_1_2108_14707_9807")

merge_sequences(data, seqs_to_merge)

# If you look at the scrap report, you will see the second two sequence
# names, listed with the trash code set to "merged".

get_scrap_report(data)

# You can see from the get_num_sequences function that the merged sequence's
# abundances are added to the first sequence.

get_num_sequences(data)

}
