% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miseq_sop_example.R
\name{miseq_sop_example}
\alias{miseq_sop_example}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
using the analysis files from the
\href{https://mothur.org/wiki/miseq_sop/}{MiSeq_SOP} example.}
\usage{
miseq_sop_example()
}
\value{
A \code{strollur::strollur} object
}
\description{
The miseq_sop_example function will create
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
using the analysis files from the
\href{https://mothur.org/wiki/miseq_sop/}{MiSeq_SOP} example.
}
\examples{

miseq <- strollur::miseq_sop_example()

}
