% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miseq_sop_example.R
\name{miseq_sop_example}
\alias{miseq_sop_example}
\title{miseq_sop_example}
\usage{
miseq_sop_example()
}
\value{
A 'sequence_data' object
}
\description{
The miseq_sop_example function will create 'sequence_data' object using the
analysis files from the \href{https://mothur.org/wiki/miseq_sop/}{MiSeq_SOP}
example.
}
\examples{

miseq_dataset <- miseq_sop_example()

}
