% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dataset.R
\name{new_dataset}
\alias{new_dataset}
\title{new_dataset}
\usage{
new_dataset(dataset_name = "")
}
\arguments{
\item{dataset_name}{string, a string containing the dataset name.
Default = ""}
}
\value{
a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Create a new
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- strollur::new_dataset()

# to create a new dataset named "soil", run the following:

data <- strollur::new_dataset(dataset_name = "soil")

}
\seealso{
The 'new' method in the
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} class
}
