% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_dataset}
\alias{new_dataset}
\title{new_dataset}
\usage{
new_dataset(dataset_name, processors)
}
\arguments{
\item{dataset_name}{string, a string containing the dataset name}

\item{processors}{integer, number of cores to use during summary functions.}
}
\value{
Rcpp::XPtr<Dataset> pointer to an instance of the 'Dataset' c++
 class.
}
\description{
Create a pointer to an instance of the 'Dataset' c++ class.

The 'Dataset' class is the c++ implementation of the R6 'sequence_data'
object. This class allows package developers access to additional
functionality. 'Dataset' stores nucleotide sequences, abundance, sample and
treatment assignments, taxonomic classifications, asv / otu clusters. It
creates various reports and summaries. It is designed to facilitate data
transfer and access across multiple R packages.
}
\examples{

# to create a new dataset and allow for 4 processors during summary,
# run the following:

dataset <- new_dataset("soil", 4)

}
\seealso{
[sequence_data$new()]
}
