% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_dataset}
\alias{new_dataset}
\title{new_dataset}
\usage{
new_dataset(dataset_name = "", processors = 2L)
}
\arguments{
\item{dataset_name}{string, a string containing the dataset name.
Default = ""}

\item{processors}{integer, number of cores to use during summary functions.
Default = 2}
}
\value{
a \link{dataset} object
}
\description{
Create a new \link{dataset} object
}
\examples{

data <- new_dataset()

# to create a new dataset named "soil" and allow for 2 processors during
# summary functions, run the following:

data <- new_dataset("soil", 2)

# to create a new dataset named "soil" and allow for all available
# processors during summary functions, run the following:

data <- new_dataset("soil", get_available_processors())

}
\seealso{
The 'new' method in the \link{dataset} class
}
