% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_dataset}
\alias{new_dataset}
\title{new_dataset}
\usage{
new_dataset(dataset_name = "", processors = NULL)
}
\arguments{
\item{dataset_name}{string, a string containing the dataset name.
Default = ""}

\item{processors}{integer, number of cores to use during summary functions.
Default = all available}
}
\value{
a \link{strollur::strollur} object
}
\description{
Create a new \link{strollur::strollur} object
}
\examples{

data <- new_dataset()

# to create a new dataset named "soil" and allow for all available
# processors during summary functions, run the following:

data <- new_dataset(dataset_name = "soil")

# to create a new dataset named "soil" and allow for 2
# processors during summary functions, run the following:

data <- new_dataset(dataset_name = "soil", processors = 2)

}
\seealso{
The 'new' method in the \link{strollur::strollur} class
}
