% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dada2.R
\name{read_dada2}
\alias{read_dada2}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
from dada2 outputs}
\usage{
read_dada2(sequence_table, dataset_name = "")
}
\arguments{
\item{sequence_table}{A dada2 sequence table}

\item{dataset_name}{A string containing a name for your dataset.}
}
\value{
A \code{strollur::strollur} object
}
\description{
This function reads a dada2 sequence table and creates a \code{strollur::strollur}
object. The dada2 sequence table is a 2D matrix containing the abundance
counts by sample for each ASV. The sample names are stored as row names and
the sequence nucleotide strings are stored as column names.

To generate the dada2 sequence table from your own files you can follow
\href{https://benjjneb.github.io/dada2/tutorial.html}{this dada2 tutorial}.
}
\examples{

seqtab <- readRDS(strollur_example("dada2.rds"))
dim(seqtab)

data <- strollur::read_dada2(sequence_table = seqtab,
                               dataset_name = "dada2 example")

}
\references{
Callahan,B.J., McMurdie,P.J., Rosen,M.J., Han,A.W., Johnson,A.J.A.
and Holmes,S.P. (2016), DADA2: High-resolution sample inference from
Illumina amplicon data. Nature Methods 13:581-583.
\url{doi:10.1038/nmeth.3869}
}
