% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dada2.R
\name{read_dada2}
\alias{read_dada2}
\title{read_dada2}
\usage{
read_dada2(sequence_table, dataset_name = "")
}
\arguments{
\item{sequence_table}{A dada2 sequence table}

\item{dataset_name}{A string containing a name for your dataset.}
}
\value{
A 'dataset' object
}
\description{
This function will read a dada2 sequence table and create a 'dataset'
object. The dada2 sequence table is a 2D matrix containing the abundance
counts by sample for each ASV. The sample names are stored as row names and
the sequence nucleotide strings are stored as column names.

To generate the dada2 sequence table you can follow
\href{https://benjjneb.github.io/dada2/tutorial.html}{this dada2 tutorial}.
}
\examples{
\dontrun{
# Follow the dada2 tutorial above to generate seqtab from your fastq files.

# dim(seqtab)
# [1]  20 293

# data <- read_dada2(seqtab, "dada2")
}
}
