% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{read_fasta}
\usage{
read_fasta(fasta)
}
\arguments{
\item{fasta}{FASTA file name (required)}
}
\value{
A data.frame containing the FASTA sequence data
}
\description{
Read a FASTA formatted sequence file
}
\examples{

fasta_data <- read_fasta(rdataset_example("final.fasta"))

# fasta_data is a data.frame.
# To access the names of the sequences in the file, run the following:

fasta_data$names

# To access the sequences in the file, run the following:

fasta_data$sequences

}
