% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_file.R
\name{read_fasta_file}
\alias{read_fasta_file}
\title{read_fasta_file}
\usage{
read_fasta_file(fasta)
}
\arguments{
\item{fasta}{FASTA file name (required)}
}
\value{
A list containing the FASTA sequence data
}
\description{
Read a FASTA formatted sequence file
}
\examples{

fasta_data <- read_fasta_file(rdataset_example("test.fasta"))

# fasta_data is a named list.
# To access the names of the sequences in the file, run the following:

fasta_data$names

# To access the sequences in the file, run the following:

fasta_data$sequences

}
