% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur.R
\name{read_mothur}
\alias{read_mothur}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
from mothur outputs}
\usage{
read_mothur(
  fasta = NULL,
  count = NULL,
  taxonomy = NULL,
  otu_list = NULL,
  asv_list = NULL,
  phylo_list = NULL,
  design = NULL,
  cons_taxonomy = NULL,
  otu_shared = NULL,
  asv_shared = NULL,
  phylo_shared = NULL,
  sample_tree = NULL,
  sequence_tree = NULL,
  dataset_name = ""
)
}
\arguments{
\item{fasta}{filename, a FASTA formatted file containing sequence strings.
\href{https://mothur.org/wiki/fasta_file/}{fasta file}}

\item{count}{filename, a mothur
\href{https://mothur.org/wiki/count_file/}{count file}}

\item{taxonomy}{filename, a mothur
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}, created by
\href{https://mothur.org/wiki/classify.seqs/}{classify.seqs}}

\item{otu_list}{filename, a mothur
\href{https://mothur.org/wiki/list_file/}{list file} containing otu bin
assignments. The otu_list file is created by
\href{https://mothur.org/wiki/cluster/}{cluster},
\href{https://mothur.org/wiki/cluster.split/}{cluster.split}, and
\href{https://mothur.org/wiki/cluster.fit/}{cluster.fit}}

\item{asv_list}{filename, a mothur
\href{https://mothur.org/wiki/list_file/}{list file} containing asv bin
assignments. The asv_list file is created by
\href{https://mothur.org/wiki/cluster/}{cluster} using the 'unique' method.}

\item{phylo_list}{filename, a mothur
\href{https://mothur.org/wiki/list_file/}{list file} containing phylotype bin
assignments. The phylo_list file is created by
\href{https://mothur.org/wiki/phylotype/}{phylotype}.}

\item{design}{filename, a mothur
\href{https://mothur.org/wiki/design_file/}{design file}}

\item{cons_taxonomy}{filename, a mothur consensus taxonomy file
\href{https://mothur.org/wiki/constaxonomy_file/}{constaxonomy file}. The
cons_taxonomy file is created by
\href{https://mothur.org/wiki/classify.otu/}{classify.otu}.}

\item{otu_shared}{filename, a mothur
\href{https://mothur.org/wiki/shared_file/}{shared file} containing otu bin
sample abundance assignments.}

\item{asv_shared}{filename, a mothur
\href{https://mothur.org/wiki/shared_file/}{shared file} containing asv bin
sample abundance assignments.}

\item{phylo_shared}{filename, a mothur
\href{https://mothur.org/wiki/shared_file/}{shared file} containing phylotype
bin sample abundance assignments.}

\item{sample_tree}{filename, a tree that relates samples.
The sample tree is created by
\href{https://mothur.org/wiki/tree.shared/}{tree.shared}. We recommend
running tree.shared with subsample = true, and using the 'ave.tre' output
for best results.}

\item{sequence_tree}{filename, a tree that relates sequences.
The sequence tree is created by
\href{https://mothur.org/wiki/clearcut/}{clearcut}. We DO NOT recommend
using sequence trees. With the ever growing size of modern datasets,
sequence tree can be difficult / impossible to build without hitting a
memory limitation.}

\item{dataset_name}{A string containing a name for your dataset.}
}
\value{
A
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
The read_mothur function reads various
\href{https://mothur.org/wiki/tags/#file_types}{file types} created by
mothur, and creates a \code{strollur::strollur} object.

To generate the various input files you can follow Pat's
\href{https://mothur.org/wiki/miseq_sop/}{Miseq example analysis}.
}
\note{
\itemize{
\item \emph{consensus taxonomy}, The \code{strollur::strollur} object will
generate consensus taxonomies for you based on the sequence taxonomy
assignment. You only need to provide the ".cons.taxonomy" file if you are not
providing sequence taxonomy assignments.
\item \emph{shared / rabund file}, The \code{strollur::strollur} object will
generate shared and rabund data for you based on the otu assignment in the
list file and the count data. You only need to provide the ".shared" file if
you are not providing the list and count files.
}
}
\examples{
# For dataset's including sequence data:

data <- strollur::read_mothur(
  fasta = strollur_example("final.fasta.gz"),
  count = strollur_example("final.count_table.gz"),
  taxonomy = strollur_example("final.taxonomy.gz"),
  design = strollur_example("mouse.time.design"),
  otu_list = strollur_example("final.opti_mcc.list.gz"),
  asv_list = strollur_example("final.asv.list.gz"),
  phylo_list = strollur_example("final.tx.list.gz"),
  sample_tree = strollur_example("final.opti_mcc.jclass.ave.tre"),
  dataset_name = "miseq_sop"
)

# For dataset's with only otu data:

data <- strollur::read_mothur(
  otu_shared = strollur_example("final.opti_mcc.shared"),
  cons_taxonomy = strollur_example(
    "final.cons.taxonomy"
  ),
  design = strollur_example("mouse.time.design"),
  sample_tree = strollur_example("final.opti_mcc.jclass.ave.tre"),
  dataset_name = "miseq_sop"
)

}
\references{
Schloss,P.D., Westcott,S.L., Ryabin,T., Hall,J.R., Hartmann,M.,
Hollister,E.B., Lesniewski,R.A., Oakley,B.B., Parks,D.H., Robinson,C.J.,
Sahl,J.W., Stres,B., Thallinger,G.G., Van Horn,D.J. and Weber,C.F. (2009),
Introducing mothur: Open-source, platform-independent, community-supported
software for describing and comparing microbial communities. Applied and
Environmental Microbiology 75:7537-7541. \url{doi:10.1128/AEM.01541-09}
}
