% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur.R
\name{read_mothur}
\alias{read_mothur}
\title{read_mothur}
\usage{
read_mothur(
  fasta = NULL,
  count = NULL,
  taxonomy = NULL,
  list = NULL,
  design = NULL,
  shared = NULL,
  rabund = NULL,
  cons_taxonomy = NULL,
  dataset_name = "",
  list_type = "otu",
  shared_type = "otu",
  rabund_type = "otu"
)
}
\arguments{
\item{fasta}{filename, a FASTA formatted file containing sequence strings.
\href{https://mothur.org/wiki/fasta_file/}{fasta file}}

\item{count}{filename, a mothur
\href{https://mothur.org/wiki/count_file/}{count file}}

\item{taxonomy}{filename, a mothur
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}}

\item{list}{filename, a mothur
\href{https://mothur.org/wiki/list_file/}{list file}}

\item{design}{filename, a mothur
\href{https://mothur.org/wiki/design_file/}{design file}}

\item{shared}{filename, a mothur
\href{https://mothur.org/wiki/shared_file/}{shared file}}

\item{rabund}{filename, a mothur
\href{https://mothur.org/wiki/rabund_file/}{rabund file}}

\item{cons_taxonomy}{filename, a mothur consensus taxonomy file
\href{https://mothur.org/wiki/constaxonomy_file/}{constaxonomy file}}

\item{dataset_name}{A string containing a name for your dataset.}

\item{list_type}{string containing the type of list file. Options include:
'otu', 'asv' or 'phylotype'. Default = 'otu'.}

\item{shared_type}{string containing the type of shared file. Options
include: 'otu', 'asv' or 'phylotype'. Default = 'otu'.}

\item{rabund_type}{string containing the type of rabund file. Options
include: 'otu', 'asv' or 'phylotype'. Default = 'otu'.}
}
\value{
A 'sequence_data' object
}
\description{
The read_mothur function will read various
\href{https://mothur.org/wiki/tags/#file_types}{file types} created by
mothur, and create a "sequence_data" object.

To generate the various output files you can follow Pat's
\href{https://mothur.org/wiki/miseq_sop/}{Miseq example analysis}.
}
\note{
\itemize{
\item \emph{consensus taxonomy}, The 'sequence_data' object will generate
consensus taxonomies for you based on the sequence taxonomy assignment. You
only need to provide the ".cons.taxonomy" file if you are not providing
sequence taxonomy assignments.
\item \emph{shared / rabund file}, The 'sequence_data' object will generate
shared and rabund data for you based on the otu assignment in the list file
and the count data. You only need to provide the ".shared" file if you are
not providing the list and count files.
}
}
\examples{
# For dataset's including sequence data:

dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

# For dataset's with only otu data:

dataset <- read_mothur(shared = rdataset_example("final.opti_mcc.shared"),
                       cons_taxonomy = rdataset_example(
                       "final.cons.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       dataset_name = "miseq_sop")

}
