% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_cons_taxonomy.R
\name{read_mothur_cons_taxonomy}
\alias{read_mothur_cons_taxonomy}
\title{read_mothur_cons_taxonomy}
\usage{
read_mothur_cons_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{file name, a mothur
\href{https://mothur.org/wiki/constaxonomy_file/}{consensus taxonomy file}.
The cons_taxonomy file is created by
\href{https://mothur.org/wiki/classify.otu/}{classify.otu}.}
}
\value{
A data.frame containing the bin names, bin abundances and bin
taxonomies.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/constaxonomy_file/}{cons_taxonomy file}
}
\examples{

# You can add the otu assignments and bin taxonomies to the your data set
# using the following:

# read mothur's consensus taxonomy file into a data.frame
otu_data <- strollur::read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

data <- strollur::new_dataset()

# assign abundance only 'otu' bins
strollur::assign(data = data, table = otu_data,
                 type = "bin", bin_type = "otu")

# assign consensus taxonomies to 'otu' bins
strollur::assign(
  data = data, table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

}
