% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_cons_taxonomy.R
\name{read_mothur_cons_taxonomy}
\alias{read_mothur_cons_taxonomy}
\title{read_mothur_cons_taxonomy}
\usage{
read_mothur_cons_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{file name, a mothur
\href{https://mothur.org/wiki/constaxonomy_file/}{consensus taxonomy file}.
 The cons_taxonomy file is created by
\href{https://mothur.org/wiki/classify.otu/}{classify.otu}.}
}
\value{
A data.frame containing the bin names, bin abundances and bin
taxonomies.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/constaxonomy_file/}{cons_taxonomy file}
}
\examples{

# You can add the otu assignments and bin taxonomies to the your data set
# using the following:

otu_data <- read_mothur_cons_taxonomy(rdataset_example(
  "final.cons.taxonomy"
))

data <- dataset$new()
data$assign_bins(otu_data)
data$assign_bin_taxonomy(otu_data)

}
