% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_cons_taxonomy.R
\name{read_mothur_cons_taxonomy}
\alias{read_mothur_cons_taxonomy}
\title{read_mothur_cons_taxonomy}
\usage{
read_mothur_cons_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{file name (required)}
}
\value{
A data.frame containing the bin names, bin abundances and bin
taxonomies.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/constaxonomy_file/}{constaxonomy file}
}
\examples{

# You can add the otu assignments and bin taxonomies to the your data set
# using the following:

otu_data <- read_mothur_cons_taxonomy(rdataset_example(
                        "final.cons.taxonomy"))

dataset <- sequence_data$new()
dataset$assign_bins(otu_data)
dataset$assign_bin_taxonomy(otu_data)

}
