% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_list.R
\name{read_mothur_list}
\alias{read_mothur_list}
\title{read_mothur_list}
\usage{
read_mothur_list(list)
}
\arguments{
\item{list}{file name (required)}
}
\value{
A data.table containing the sequence otu assignments
}
\description{
Read a mothur formatted \href{https://mothur.org/wiki/list_file/}{list file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

otu_data <- read_mothur_list(rdataset_example("final.opti_mcc.list"))

dataset <- sequence_data$new()
dataset$assign_otus(otu_data$otu_id, seq_ids = otu_data$seq_id)

}
