% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_rabund.R
\name{read_mothur_rabund}
\alias{read_mothur_rabund}
\title{Read a mothur formatted
\href{https://mothur.org/wiki/rabund_file/}{rabund file}}
\usage{
read_mothur_rabund(rabund)
}
\arguments{
\item{rabund}{file name (required)}
}
\value{
A data.frame containing the sequence otu assignments
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/rabund_file/}{rabund file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

# read rabund file into data.frame
otu_data <- strollur::read_mothur_rabund(
  rabund =
    strollur_example("final.opti_mcc.rabund")
)

data <- strollur::new_dataset()

# assign abundance only 'otu' bins
strollur::assign(data = data, table = otu_data,
                 type = "bin", bin_type = "otu")

}
