% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_rabund.R
\name{read_mothur_rabund}
\alias{read_mothur_rabund}
\title{read_mothur_rabund}
\usage{
read_mothur_rabund(rabund)
}
\arguments{
\item{rabund}{file name (required)}
}
\value{
A data.frame containing the sequence otu assignments
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/rabund_file/}{rabund file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

otu_data <- read_mothur_rabund(rdataset_example("final.opti_mcc.rabund"))

data <- dataset$new()
data$assign_bins(otu_data)

}
