% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_shared.R
\name{read_mothur_shared}
\alias{read_mothur_shared}
\title{Read a mothur formatted
\href{https://mothur.org/wiki/shared_file/}{shared file}}
\usage{
read_mothur_shared(shared)
}
\arguments{
\item{shared}{file name (required)}
}
\value{
A data.frame containing the sequence otu assignments
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/shared_file/}{shared file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

# read mothur shared file into data.frame
otu_data <-
     strollur::read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# create a new empty `strollur::strollur` object
data <- strollur::new_dataset()

# assign abundance only 'otu' bins parsed by sample
strollur::assign(data = data, table = otu_data,
                  type = "bin", bin_type = "otu")

}
