% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_shared.R
\name{read_mothur_shared}
\alias{read_mothur_shared}
\title{read_mothur_shared}
\usage{
read_mothur_shared(shared)
}
\arguments{
\item{shared}{file name (required)}
}
\value{
A data.frame containing the sequence otu assignments
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/shared_file/}{shared file}
}
\examples{

# You can add your otu assignments to the your data set using the following:

otu_data <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))

dataset <- sequence_data$new()
dataset$assign_bins(otu_data$bin_id, otu_data$abundance, otu_data$sample)

}
