% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_taxonomy.R
\name{read_mothur_taxonomy}
\alias{read_mothur_taxonomy}
\title{read_mothur_taxonomy}
\usage{
read_mothur_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{file name. a mothur
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}, created by
\href{https://mothur.org/wiki/classify.seqs/}{classify.seqs}}
}
\value{
A data.frame containing the sequences names and sequences taxonomies.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}
}
\examples{

# You can add the sequences and their taxonomies to the your data set
# using the following:

# read mothur's taxonomy file into a data.frame
classification_data <- strollur::read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

# create a new empty `strollur::strollur` object
data <- strollur::new_dataset()

# assign sequence classifications
strollur::assign(data = data, table = classification_data,
                 type = "sequence_taxonomy")

}
