% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur_taxonomy.R
\name{read_mothur_taxonomy}
\alias{read_mothur_taxonomy}
\title{read_mothur_taxonomy}
\usage{
read_mothur_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{file name (required)}
}
\value{
A data.frame containing the sequences names and sequences taxonomies.
}
\description{
Read a mothur formatted
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}
}
\examples{

# You can add the sequences and their taxonomies to the your data set
# using the following:

classification_data <- read_mothur_taxonomy(rdataset_example(
                        "final.taxonomy"))

dataset <- sequence_data$new()
dataset$assign_sequence_taxonomy(classification_data)

}
