% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phyloseq.R
\name{read_phyloseq}
\alias{read_phyloseq}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
from a phyloseq object}
\usage{
read_phyloseq(phyloseq_object, treatment_column_name = NULL, dataset_name = "")
}
\arguments{
\item{phyloseq_object}{the phyloseq object that is returned when using
any read function in the phyloseq package. It has to be of type "phyloseq"}

\item{treatment_column_name}{the column name inside your phyloseq object
within your sample data that is used to describe treatments. It must
be a character. Defaults to NULL.}

\item{dataset_name}{A string containing a name for your dataset.}
}
\value{
a strollur object.
}
\description{
The \code{read_phyloseq()} function reads phyloseq objects created from
the phyloseq package
(https://www.bioconductor.org/packages/release/bioc/html/phyloseq.html)
and converts it into a strollur object.
}
\examples{
miseq <- miseq_sop_example()

if (requireNamespace("phyloseq", quietly = TRUE)) {
  phylo_obj <- write_phyloseq(miseq)
  miseq_re_read <- read_phyloseq(phylo_obj)
} else {
  message(paste(
    "To use this functionality you have to install the",
    "phyloseq package."
  ))
}
}
\references{
McMurdie,P.J. and Holmes,S. (2013), phyloseq: An R Package for
Reproducible Interactive Analysis and Graphics of Microbiome Census Data.
PLoS ONE 8:e61217. \url{doi:10.1371/journal.pone.0061217}
}
