% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phyloseq.R
\name{read_phyloseq}
\alias{read_phyloseq}
\title{read_phyloseq}
\usage{
read_phyloseq(phyloseq_object, treatment_column_name = NULL, dataset_name = "")
}
\arguments{
\item{phyloseq_object}{the phyloseq object that is returned when using
any read function in the phyloseq package. It has to be of type "phyloseq"}

\item{treatment_column_name}{the column name inside your phyloseq object
within your sample data that is used to descrbe treatments. It must
be a character. Defaults to NULL.}

\item{dataset_name}{A string containing a name for your dataset.}
}
\value{
a strollur object.
}
\description{
The `read_phyloseq()` function reads phyloseq objects created from
the phyloseq package
(https://www.bioconductor.org/packages/release/bioc/html/phyloseq.html)
and converts it into a strollur object.
}
\examples{
miseq <- miseq_sop_example()
phylo_obj <- write_phyloseq(miseq)
miseq_re_read <- read_phyloseq(phylo_obj)
}
