% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2}
\alias{read_qiime2}
\title{read_qiime2}
\usage{
read_qiime2(
  qza,
  metadata = NULL,
  dataset_name = "",
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)
}
\arguments{
\item{qza}{vector of filenames, .qza files containing your data from qiime2.}

\item{metadata}{filename, a .tsv file containing metadata}

\item{dataset_name}{A string containing a name for your dataset.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the unpacked artifacts
and temporary directories will be removed. Default = TRUE.}
}
\value{
A 'dataset' object
}
\description{
The read_qiime function reads various types of .qza files created by
\href{https://qiime2.org}{qiime2}, and creates a 'dataset' object.

To generate the various input files you can follow \href{https://amplicon-docs.qiime2.org/en/latest/tutorials/moving-pictures.html}{qiime moving-pictures}.
}
\examples{

# Using the example files from moving-pictures, we add FASTA data, assign
# taxonomy and abundance for features, and add a newick tree and
# metadata.

qza_files <- c(
  rdataset_example("rep_seqs.qza"),
  rdataset_example("table.qza"),
  rdataset_example("taxonomy.qza"),
  rdataset_example("rooted-tree.qza")
)

data <- read_qiime2(
  qza = qza_files,
  metadata = rdataset_example("sample_metadata.tsv"),
  dataset_name = "qiime_moving_pictures"
)
data

}
