% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2}
\alias{read_qiime2}
\title{Create a \link{strollur::strollur} object from a qiime2 outputs}
\usage{
read_qiime2(
  qza,
  metadata = NULL,
  dataset_name = "",
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)
}
\arguments{
\item{qza}{vector of filenames, .qza files containing your data from qiime2.}

\item{metadata}{filename, a .tsv file containing metadata}

\item{dataset_name}{A string containing a name for your dataset.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the unpacked artifacts
and temporary directories will be removed. Default = TRUE.}
}
\value{
A `strollur` object
}
\description{
The read_qiime2 function reads various types of .qza files created by
\href{https://qiime2.org}{qiime2}, and creates a `strollur` object.

To generate the various input files you can follow \href{https://amplicon-docs.qiime2.org/en/latest/tutorials/moving-pictures.html}{qiime2 moving-pictures}.
}
\examples{

# Using the example files from moving-pictures, we add FASTA data, assign
# taxonomy and abundance for features, and add a newick tree and
# metadata.

qza_files <- c(
  strollur_example("rep_seqs.qza"),
  strollur_example("table.qza"),
  strollur_example("taxonomy.qza"),
  strollur_example("rooted-tree.qza")
)

data <- read_qiime2(
  qza = qza_files,
  metadata = strollur_example("sample_metadata.tsv"),
  dataset_name = "qiime2_moving_pictures"
)
data

}
