% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2}
\alias{read_qiime2}
\title{read_qiime2}
\usage{
read_qiime2(
  fasta = NULL,
  taxonomy = NULL,
  biom = NULL,
  cons_fasta = NULL,
  cons_taxonomy = NULL,
  tree = NULL,
  metadata = NULL,
  dataset_name = "",
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)
}
\arguments{
\item{fasta}{filename, a .qza file containing DNA sequence strings.}

\item{taxonomy}{filename, a .qza file containing classifications.}

\item{biom}{filename, a .qza file containing a feature table in biom
format. This will be used to assign bin abundance and sample data.}

\item{cons_fasta}{filename, a .qza file containing DNA sequence strings of
representative sequences for each bin.}

\item{cons_taxonomy}{filename, a .qza file containing conscensus
classifications for each bin.}

\item{tree}{filename, a .qza file containing a tree.}

\item{metadata}{filename, a .tsv file containing metadata}

\item{dataset_name}{A string containing a name for your dataset.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the unpacked artifacts
and temporary directories will be removed. Default = TRUE.}
}
\value{
A 'dataset' object
}
\description{
The read_qiime function reads various types of .qza files created by
\href{https://qiime2.org}{qiime2}, and creates a 'dataset' object.

To generate the various input files you can follow \href{https://amplicon-docs.qiime2.org/en/latest/tutorials/moving-pictures.html}{qiime moving-pictures}.
}
\examples{

# Using the example files from moving-pictures

# dataset <- read_qiime2(
#   cons_fasta = rdataset_example("rep_seqs.qza"),
#   biom = rdataset_example("table.qza"),
#   cons_taxonomy = rdataset_example("taxonomy.qza"),
#   tree = rdataset_example("rooted-tree.qza"),
#   metadata = rdataset_example("sample_metadata.tsv"),
#   dataset_name = "qiime_moving_pictures"
# )

}
