% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2_feature_table.R
\name{read_qiime2_feature_table}
\alias{read_qiime2_feature_table}
\title{read_qiime2_feature_table}
\usage{
read_qiime2_feature_table(
  qza,
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)
}
\arguments{
\item{qza}{file name, a qiime2 .qza file containing bin data.}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be unpacked. Default = current working directory.}

\item{remove_unpacked_artifacts}{boolean, When TRUE, the artifact's temporary
directories will be removed after processing. Default = TRUE.}
}
\value{
A list containing artifact
}
\description{
Read a \href{https://qiime2.org}{qiime2} qza containing bin data
}
\examples{

artifact <- read_qiime2_feature_table(strollur_example(
  "table.qza"
))

# access the bin assignment table

artifact$data

# to create a `strollur` object with your data

data <- new_dataset("my_data")

assign(data = data, table = artifact$data, type = "bins")

data

}
