% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2_metadata.R
\name{read_qiime2_metadata}
\alias{read_qiime2_metadata}
\title{read_qiime2_metadata}
\usage{
read_qiime2_metadata(metadata)
}
\arguments{
\item{metadata}{file name, a qiime2 .tsv file containing metadata about your
analysis.}
}
\value{
A data.frame containing metadata
}
\description{
Read a \href{https://qiime2.org}{qiime2} .tsv table containing metadata.
}
\examples{

metadata <- read_qiime2_metadata(rdataset_example(
  "sample_metadata.tsv"
))

}
