% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quality.R
\name{read_quality}
\alias{read_quality}
\title{read_quality}
\usage{
read_quality(quality, path = NULL)
}
\arguments{
\item{quality}{string, containing the quality score file name.}

\item{path}{string, optional path to quality score file.}
}
\value{
A data.frame containing the quality score sequence data
}
\description{
Read a file containing sequence quality scores
}
\examples{

quality_data <- strollur::read_quality(strollur_example("tiny.qual"))

# quality_data is a data.frame.
# To access the names of the sequences in the file, run the following:

quality_data$sequence_name

# To access the quality scores in the file, run the following:

quality_data$quality_score

}
