% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_samples}
\alias{remove_samples}
\title{remove_samples}
\usage{
remove_samples(data, samples)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{samples}{vector of strings containing the names of the samples to}
}
\description{
Remove samples from an instance of the 'Dataset' class.
}
\examples{
dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       otu_list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

get_num_samples(dataset$data)

# To remove samples 'F3D0' and 'F3D1'

remove_samples(dataset$data, c("F3D0", "F3D1"))

get_num_samples(dataset$data)

}
