% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{report}
\alias{report}
\title{report}
\usage{
report(data, type = "sequence_data", bin_type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of report you would like. Options
include: "sequence_data", "sequence_taxonomy", "bin_taxonomy",
"sequence_scrap", "bin_scrap", "metadata", "references". If you have added
custom reports for alignment, contigs_assembly or chimeras, you can get those
as well. Default = "sequence_data".}

\item{bin_type, }{string containing the bin type you would like a bin_taxonomy
report for. Default = "otu".}
}
\value{
data.frame
}
\description{
Get a data.frame containing the given report in a \link{dataset} object
}
\examples{

# First let's create a dataset from the \href{https://mothur.org/wiki/miseq_sop/}{MiSeq_SOP}

miseq <- miseq_sop_example()

# To get a report about the FASTA data

sequence_report <- report(data = miseq, type = "sequence_data")
head(sequence_report, n = 10)

# To get a report about sequence classifications

sequence_taxonomy_report <- report(data = miseq,
                                       type = "sequence_taxonomy")
head(sequence_taxonomy_report, n = 10)

# To get a report about bin classifications for 'otu' data

otu_taxonomy_report <- report(data = miseq,
                                       type = "bin_taxonomy",
                                       bin_type = "otu")
head(otu_taxonomy_report, n = 10)

# To get a report about bin classifications for 'asv' data

asv_taxonomy_report <- report(data = miseq, type = "bin_taxonomy",
                              bin_type = "asv")
head(asv_taxonomy_report, n = 10)

# To get a report about bin classifications for 'phylotype' data

phylotype_taxonomy_report <- report(data = miseq, type = "bin_taxonomy",
                                    bin_type = "phylotype")
head(phylotype_taxonomy_report, n = 10)

# To get a report about the sequences removed during your analysis:

scrapped_sequence_report <- report(data = miseq, type = "sequence_scrap")

# To get a report about the "otu" bins removed during your analysis:

scrapped_otu_report <- report(data = miseq, type = "bin_scrap",
                              bin_type = "otu")

# To get a report about the "phylotype" bins removed during your analysis:

scrapped_phylotype_report <- report(data = miseq, type = "bin_scrap",
                                    bin_type = "phylotype")

# To get the metadata associated with your data:

metadata <- report(data = miseq, type = "metadata")

# To get the resource references associated with your data:

references <- report(data = miseq, type = "references")

# To get our custom report containing the contigs assembly data:

contigs_report <- report(data = miseq, type = "contigs_report")
head(contigs_report, n = 10)

}
