% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataset.R
\name{save_dataset}
\alias{save_dataset}
\title{save_dataset}
\usage{
save_dataset(data, file)
}
\arguments{
\item{data}{a \link{strollur::strollur} object}

\item{file}{a string containing the file name.}
}
\value{
A file containing the `strollur` object
}
\description{
The save_dataset function will save the \link{strollur::strollur} object to
file.
}
\examples{

data <- read_mothur(
  fasta = strollur_example("final.fasta.gz"),
  count = strollur_example("final.count_table.gz"),
  taxonomy = strollur_example("final.taxonomy.gz"),
  design = strollur_example("mouse.time.design"),
  otu_list = strollur_example("final.opti_mcc.list.gz"),
  dataset_name = "miseq_sop"
)

save_dataset(data, "miseq_sop.rds")

}
