% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataset.R
\name{save_dataset}
\alias{save_dataset}
\title{save_dataset}
\usage{
save_dataset(dataset, file)
}
\arguments{
\item{dataset}{a \link{dataset} object}

\item{file}{a string containing the file name.}
}
\value{
A file containing the 'dataset' object
}
\description{
The save_dataset function will save the \link{dataset} object to file.
}
\examples{

dataset <- read_mothur(
  fasta = rdataset_example("final.fasta.gz"),
  count = rdataset_example("final.count_table.gz"),
  taxonomy = rdataset_example("final.taxonomy.gz"),
  design = rdataset_example("mouse.time.design"),
  otu_list = rdataset_example("final.opti_mcc.list.gz"),
  dataset_name = "miseq_sop"
)

save_dataset(dataset, "miseq_sop.rds")

}
