% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_data.R
\name{sequence_data}
\alias{sequence_data}
\title{sequence_dataset}
\description{
'sequence_data' is an R6 class that represents sequence
 FASTA and abundance data.
}
\examples{

## ------------------------------------------------
## Method `sequence_data$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))


## ------------------------------------------------
## Method `sequence_data$assign_otu_abundance`
## ------------------------------------------------

  # To assign sequences to otus:

  # otu_ids  seq_ids  abundances
  #  otu1     seq1        10
  #  otu1     seq2        100
  #  otu1     seq4        1
  #  otu2     seq3        500
  #  otu2     seq6        25
  #  otu3     seq5        80
  # ...

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otu_abundance("0.03", otu_ids,
              sequence_abundances, seq_ids = seq_ids)

  # otus would look like:
  # label  otu1             otu2        otu3 ...
  # 0.03   seq1,seq2,seq4   seq3,seq6   seq5 ...
  # 0.03   110              525         80 ...

  # To add abundance only otu assignments:

  # otu_ids  samples  abundances
  #  otu1     sample1        10
  #  otu1     sample2        100
  #  otu1     sample5        1
  #  otu2     sample1        500
  #  otu2     sample3        25
  #  otu3     sample1        80
  # ...

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otu_abundance("0.03", otu_ids,
                                     sequence_abundances, samples)

  # otus would look like:
  # label  sample   otu1   otu2   otu3 ...
  # 0.03   sample1  10     500    80 ...
  # 0.03   sample2  100    0      0 ...
  # 0.03   sample3  0      25     0 ...
  # 0.03   sample4  0      0      0 ...
  # 0.03   sample5  1      0      0 ...

## ------------------------------------------------
## Method `sequence_data$add_sequences`
## ------------------------------------------------


  dataset <- sequence_data$new("my_dataset")
  sequence_data <- read_fasta(rdataset_example("test.fasta"))
  dataset$add_sequences(sequence_data$names, sequence_data$sequences)


## ------------------------------------------------
## Method `sequence_data$assign_sequence_abundance`
## ------------------------------------------------


 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
samples <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_sequences(unique_names, sequences)
dataset$assign_sequence_abundance(names, abundances, samples)

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{pointer to C++ class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_data-new}{\code{sequence_data$new()}}
\item \href{#method-sequence_data-print}{\code{sequence_data$print()}}
\item \href{#method-sequence_data-assign_otu_abundance}{\code{sequence_data$assign_otu_abundance()}}
\item \href{#method-sequence_data-add_sequences}{\code{sequence_data$add_sequences()}}
\item \href{#method-sequence_data-assign_sequence_abundance}{\code{sequence_data$assign_sequence_abundance()}}
\item \href{#method-sequence_data-clear}{\code{sequence_data$clear()}}
\item \href{#method-sequence_data-export}{\code{sequence_data$export()}}
\item \href{#method-sequence_data-get_align_report}{\code{sequence_data$get_align_report()}}
\item \href{#method-sequence_data-get_contigs_report}{\code{sequence_data$get_contigs_report()}}
\item \href{#method-sequence_data-get_count_table}{\code{sequence_data$get_count_table()}}
\item \href{#method-sequence_data-get_dataset_name}{\code{sequence_data$get_dataset_name()}}
\item \href{#method-sequence_data-get_ids}{\code{sequence_data$get_ids()}}
\item \href{#method-sequence_data-get_num_otus}{\code{sequence_data$get_num_otus()}}
\item \href{#method-sequence_data-get_num_samples}{\code{sequence_data$get_num_samples()}}
\item \href{#method-sequence_data-get_num_sequences}{\code{sequence_data$get_num_sequences()}}
\item \href{#method-sequence_data-get_num_treatments}{\code{sequence_data$get_num_treatments()}}
\item \href{#method-sequence_data-get_oligos}{\code{sequence_data$get_oligos()}}
\item \href{#method-sequence_data-get_otu_tables}{\code{sequence_data$get_otu_tables()}}
\item \href{#method-sequence_data-get_samples}{\code{sequence_data$get_samples()}}
\item \href{#method-sequence_data-get_sample_summary}{\code{sequence_data$get_sample_summary()}}
\item \href{#method-sequence_data-get_scrap_report}{\code{sequence_data$get_scrap_report()}}
\item \href{#method-sequence_data-get_sequence_report}{\code{sequence_data$get_sequence_report()}}
\item \href{#method-sequence_data-get_sequence_summary}{\code{sequence_data$get_sequence_summary()}}
\item \href{#method-sequence_data-get_sequences}{\code{sequence_data$get_sequences()}}
\item \href{#method-sequence_data-get_taxonomy_tables}{\code{sequence_data$get_taxonomy_tables()}}
\item \href{#method-sequence_data-get_treatments}{\code{sequence_data$get_treatments()}}
\item \href{#method-sequence_data-has_sample}{\code{sequence_data$has_sample()}}
\item \href{#method-sequence_data-is_aligned}{\code{sequence_data$is_aligned()}}
\item \href{#method-sequence_data-clone}{\code{sequence_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new FASTA sequence dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$new(
  name,
  fasta = NULL,
  processors = parallelly::availableCores()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of dataset}

\item{\code{fasta, }}{String name of FASTA file}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `sequence_data` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_otu_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_otu_abundance}{}}}
\subsection{Method \code{assign_otu_abundance()}}{
Add otu data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_otu_abundance(
  label,
  otu_ids,
  abundances,
  samples = NULL,
  seq_ids = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{a String containing otu label}

\item{\code{otu_ids}}{a vector of otu labels}

\item{\code{abundances}}{a vector of abundances}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{seq_ids}}{a vector of sequence names (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  # To assign sequences to otus:

  # otu_ids  seq_ids  abundances
  #  otu1     seq1        10
  #  otu1     seq2        100
  #  otu1     seq4        1
  #  otu2     seq3        500
  #  otu2     seq6        25
  #  otu3     seq5        80
  # ...

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otu_abundance("0.03", otu_ids,
              sequence_abundances, seq_ids = seq_ids)

  # otus would look like:
  # label  otu1             otu2        otu3 ...
  # 0.03   seq1,seq2,seq4   seq3,seq6   seq5 ...
  # 0.03   110              525         80 ...

  # To add abundance only otu assignments:

  # otu_ids  samples  abundances
  #  otu1     sample1        10
  #  otu1     sample2        100
  #  otu1     sample5        1
  #  otu2     sample1        500
  #  otu2     sample3        25
  #  otu3     sample1        80
  # ...

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otu_abundance("0.03", otu_ids,
                                     sequence_abundances, samples)

  # otus would look like:
  # label  sample   otu1   otu2   otu3 ...
  # 0.03   sample1  10     500    80 ...
  # 0.03   sample2  100    0      0 ...
  # 0.03   sample3  0      25     0 ...
  # 0.03   sample4  0      0      0 ...
  # 0.03   sample5  1      0      0 ...
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_sequences}{}}}
\subsection{Method \code{add_sequences()}}{
Add new sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_sequences(names, sequences, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data$new("my_dataset")
  sequence_data <- read_fasta(rdataset_example("test.fasta"))
  dataset$add_sequences(sequence_data$names, sequence_data$sequences)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sequence_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sequence_abundance}{}}}
\subsection{Method \code{assign_sequence_abundance()}}{
Add sequence abundance data with optional sample / treatment assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sequence_abundance(
  names,
  abundances,
  samples = NULL,
  treatments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{abundances}}{a vector of sequence abundances}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{treatments}}{a vector of treatment assignments (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
samples <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_sequences(unique_names, sequences)
dataset$assign_sequence_abundance(names, abundances, samples)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all sequences from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-export"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-export}{}}}
\subsection{Method \code{export()}}{
Get List containing dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_align_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_align_report}{}}}
\subsection{Method \code{get_align_report()}}{
Get report containing align report table -
ids, search_scores, sim_scores, longest_insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_align_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_contigs_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_contigs_report}{}}}
\subsection{Method \code{get_contigs_report()}}{
Get report containing contigs report table -
ids, lengths, ostarts, oends, olengths, mismatches, numns, ee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_contigs_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_count_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_count_table}{}}}
\subsection{Method \code{get_count_table()}}{
Get count table returns data.table containing:
ids, abundances, sample(optional), treatment(optional)
This table represents mothur's count and design files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_count_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_dataset_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_dataset_name}{}}}
\subsection{Method \code{get_dataset_name()}}{
Get dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_dataset_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_ids"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_ids}{}}}
\subsection{Method \code{get_ids()}}{
Get names of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_ids(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_otus"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_otus}{}}}
\subsection{Method \code{get_num_otus()}}{
Get the number of otus in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_otus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_samples}{}}}
\subsection{Method \code{get_num_samples()}}{
Get number of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_sequences}{}}}
\subsection{Method \code{get_num_sequences()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_sequences(distinct = FALSE, sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distinct}}{Boolean. When distinct is TRUE the number of unique
sequence is returned.}

\item{\code{sample}}{The name of the sample you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_treatments}{}}}
\subsection{Method \code{get_num_treatments()}}{
Get the number of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_oligos"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_oligos}{}}}
\subsection{Method \code{get_oligos()}}{
Get data.table containing the oligo data
tag, oligo, diffs, oligo(optional), diffs(optional), sample(optional)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_oligos()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_otu_tables"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_otu_tables}{}}}
\subsection{Method \code{get_otu_tables()}}{
Get data.tables representing your OTU / ASV clusters.
id, abundance, sample(optional)
These tables represent mothur's list, rabund, shared, relabund files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_otu_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of data.tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_samples}{}}}
\subsection{Method \code{get_samples()}}{
Get names of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sample_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sample_summary}{}}}
\subsection{Method \code{get_sample_summary()}}{
Get summary of count table. sample, sample_totals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sample_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print sample summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_scrap_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_scrap_report}{}}}
\subsection{Method \code{get_scrap_report()}}{
Get report containing the scrapped sequences -
ids, trash_codes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_scrap_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get report containing the sequence report table -
ids, starts, ends, lengths, ambigs, homopolymers, numns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_summary}{}}}
\subsection{Method \code{get_sequence_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print summaries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data.tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequences}{}}}
\subsection{Method \code{get_sequences()}}{
Get vector containing sequence nucleotide strings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequences(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_taxonomy_tables"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_taxonomy_tables}{}}}
\subsection{Method \code{get_taxonomy_tables()}}{
Get data.tables containing classifications for sequences and OTUs.
id, taxonomy, abundance(optional)
These tables represent mothur's taxonomy and cons.taxonomy files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_taxonomy_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of data.tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_treatments}{}}}
\subsection{Method \code{get_treatments()}}{
Get names of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-has_sample"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-has_sample}{}}}
\subsection{Method \code{has_sample()}}{
Determine if a sample is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$has_sample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-is_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-is_aligned}{}}}
\subsection{Method \code{is_aligned()}}{
Determine if the dataset is aligned
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$is_aligned()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
bool
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
