% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_data.R
\name{sequence_data}
\alias{sequence_data}
\title{sequence_data}
\description{
'sequence_data' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `sequence_data$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")


## ------------------------------------------------
## Method `sequence_data$add_metadata`
## ------------------------------------------------


  dataset <- sequence_data$new("my_dataset")
  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)
  dataset$add_metadata(metadata)


## ------------------------------------------------
## Method `sequence_data$add_references`
## ------------------------------------------------


# To add a single reference:
dataset <- sequence_data$new("my_dataset")
dataset$add_references(reference_name = "silva.v4.fasta",
version = "1.38.1", usage = "alignment", note = "custom reference
 created by trimming silva.bacteria.fasta to the V4 region",
 url = "https://mothur.org/wiki/silva_reference_files/")

# To add multiple references:
reference <- readr::read_csv(rdataset_example("references.csv"))
dataset$add_references(reference = reference)


## ------------------------------------------------
## Method `sequence_data$add_sequences`
## ------------------------------------------------


  dataset <- sequence_data$new("my_dataset")
  sequences <- read_fasta(rdataset_example("final.fasta"))
  dataset$add_sequences(sequences$names, sequences$sequences)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- "https://mothur.org/wiki/silva_reference_files/"

dataset <- sequence_data$new("my_dataset")
dataset$add_sequences(sequences$names, sequences$sequences,
reference_name = "silva.bacteria.fasta",
reference_note = "alignment by mothur2 v1.0 using default options",
reference_version = "1.38.1", reference_url = url)


## ------------------------------------------------
## Method `sequence_data$assign_bins`
## ------------------------------------------------


  # bin_ids  seq_ids
  #  bin1     seq1
  #  bin1     seq2
  #  bin1     seq4
  #  bin2     seq3
  #  bin2     seq6
  #  bin3     seq5

  # To assign sequences to bins:

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  dataset$assign_bins(bin_ids, seq_ids = seq_ids)

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5

  # To add abundance only bin assignments:

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  dataset$assign_bins(bin_ids, abundances)

  # bins would look like:
  #            bin1             bin2        bin3
  # (rabund)   110              525         80

  # To add abundance bin assignments parsed by sample:

  # bin_ids  samples  abundances
  #  bin1     sample1        10
  #  bin1     sample2        100
  #  bin1     sample5        1
  #  bin2     sample1        500
  #  bin2     sample3        25
  #  bin3     sample1        80

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_bins(bin_ids, sample_abundances, samples)

  # (shared) bins would look like:
  # label  sample   bin1   bin2   bin3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  # To assign sequences to bins with their abundances parsed by sample:

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin1", "bin1", "bin1",
               "bin2", "bin2", "bin2",
               "bin3", "bin3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  dataset$assign_bins(bin_ids, abundances, samples, seq_ids)

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5
  # (rabund)   716              85          65

  dataset$get_shared()


## ------------------------------------------------
## Method `sequence_data$assign_bin_taxonomy`
## ------------------------------------------------


bin_ids <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_bins(bin_ids, abunds)
dataset$assign_bin_taxonomy(bin_ids, taxonomies)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- paste("https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset",
         "9_032012.pds.zip", collapse = "")

dataset$assign_bin_taxonomy(bin_ids, taxonomies,
reference_name = "trainset9_032012.pds.zip",
reference_note = "classification by mothur2 v1.0 using default options",
reference_version = "9_032012", reference_url = url)


## ------------------------------------------------
## Method `sequence_data$assign_sequence_abundance`
## ------------------------------------------------


 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)


## ------------------------------------------------
## Method `sequence_data$assign_sequence_taxonomy`
## ------------------------------------------------


names <- c("seq1", "seq2", "seq3", "seq4")
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- paste("https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset",
         "9_032012.pds.zip", collapse = "")

dataset$assign_sequence_taxonomy(names = names, taxonomies = taxonomies,
reference_name = "trainset9_032012.pds.zip",
reference_note = "classification by mothur2 v1.0 using default options",
reference_version = "9_032012", reference_url = url)


## ------------------------------------------------
## Method `sequence_data$assign_treatments`
## ------------------------------------------------


names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

treatments <- c("early", "early", "late")
dataset$assign_treatments(unique(samples), treatments)


## ------------------------------------------------
## Method `sequence_data$get_list`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  dataset$assign_bins(bin_ids, sequence_abundances, seq_ids = seq_ids)

  # (list) bins would look like:
  # bin1             bin2        bin3
  # seq1,seq2,seq4   seq3,seq6   seq5

  list <- dataset$get_list()

  #  list$bin_id  list$seq_id
  #  bin1          seq1
  #  bin1          seq2
  #  bin1          seq4
  #  bin2          seq3
  #  bin2          seq6
  #  bin3          seq5


## ------------------------------------------------
## Method `sequence_data$get_metadata`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)
  dataset$add_metadata(metadata)
  dataset$get_metadata()


## ------------------------------------------------
## Method `sequence_data$get_bin_taxonomy_report`
## ------------------------------------------------


bin_ids <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_bins(bin_ids, abunds)
dataset$assign_bin_taxonomy(bin_ids, taxonomies)
dataset$get_bin_taxonomy_report()


## ------------------------------------------------
## Method `sequence_data$get_rabund`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin2", "bin3")
  bin_abundances <- c(111, 525, 80)
  dataset$assign_bins(bin_ids, bin_abundances)

  # (rabund) bins would look like:
  # bin1  bin2  bin3
  # 110   525   80

  rabund <- dataset$get_rabund()

  #  rabund$bin_id  rabund$abundance
  #  bin1           111
  #  bin2           525
  #  bin3           80


## ------------------------------------------------
## Method `sequence_data$get_references`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  reference <- readr::read_csv(rdataset_example("references.csv"))
  dataset$add_references(reference = reference)
  dataset$get_references()


## ------------------------------------------------
## Method `sequence_data$get_sequence_taxonomy_report`
## ------------------------------------------------

names <- c("seq1", "seq2", "seq3", "seq4")
tax1 <- "Bacteria(100);Bacteroidetes(95);Bacteroidia(90);"
tax2 <- "Bacteria(100);Proteobacteria(89);Betaproteobacteria(85);"
tax3 <- "Bacteria(100);Firmicutes(99);Bacilli(90);"
tax4 <- "Bacteria(100);Proteobacteria(87);Gammaproteobacteria(82);"
taxonomies <- c(tax1, tax2, tax3, tax4)

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)
dataset$get_sequence_taxonomy_report()


## ------------------------------------------------
## Method `sequence_data$get_shared`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_bins(bin_ids, sample_abundances, samples)

  # (shared) bins would look like:
  # label  sample   bin1   bin2   bin3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  shared <- dataset$get_shared()

  # shared$bin_id  shared$sample  shared$abundance
  # bin1           sample1        10
  # bin1           sample2        100
  # bin1           sample5        1
  # bin2           sample1        500
  # bin2           sample3        25
  # bin3           sample1        80


## ------------------------------------------------
## Method `sequence_data$remove_lineages`
## ------------------------------------------------

dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
 "Eukaryota")

dataset$remove_lineages(contaminants)


## ------------------------------------------------
## Method `sequence_data$remove_samples`
## ------------------------------------------------

dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

dataset$get_num_samples()

# To remove samples 'F3D0' and 'F3D1'

dataset$remove_samples(c("F3D0", "F3D1"))

dataset$get_num_samples()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{pointer to 'Dataset' (Rcpp Module). This allows package
developers an easy access point to the underlying C++ code with
additional functionality.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_data-new}{\code{sequence_data$new()}}
\item \href{#method-sequence_data-print}{\code{sequence_data$print()}}
\item \href{#method-sequence_data-add_metadata}{\code{sequence_data$add_metadata()}}
\item \href{#method-sequence_data-add_references}{\code{sequence_data$add_references()}}
\item \href{#method-sequence_data-add_sequences}{\code{sequence_data$add_sequences()}}
\item \href{#method-sequence_data-assign_bins}{\code{sequence_data$assign_bins()}}
\item \href{#method-sequence_data-assign_bin_taxonomy}{\code{sequence_data$assign_bin_taxonomy()}}
\item \href{#method-sequence_data-assign_sequence_abundance}{\code{sequence_data$assign_sequence_abundance()}}
\item \href{#method-sequence_data-assign_sequence_taxonomy}{\code{sequence_data$assign_sequence_taxonomy()}}
\item \href{#method-sequence_data-assign_treatments}{\code{sequence_data$assign_treatments()}}
\item \href{#method-sequence_data-clear}{\code{sequence_data$clear()}}
\item \href{#method-sequence_data-export}{\code{sequence_data$export()}}
\item \href{#method-sequence_data-get_align_report}{\code{sequence_data$get_align_report()}}
\item \href{#method-sequence_data-get_contigs_report}{\code{sequence_data$get_contigs_report()}}
\item \href{#method-sequence_data-get_count_table}{\code{sequence_data$get_count_table()}}
\item \href{#method-sequence_data-get_dataset_name}{\code{sequence_data$get_dataset_name()}}
\item \href{#method-sequence_data-get_ids}{\code{sequence_data$get_ids()}}
\item \href{#method-sequence_data-get_list}{\code{sequence_data$get_list()}}
\item \href{#method-sequence_data-get_metadata}{\code{sequence_data$get_metadata()}}
\item \href{#method-sequence_data-get_num_bins}{\code{sequence_data$get_num_bins()}}
\item \href{#method-sequence_data-get_num_samples}{\code{sequence_data$get_num_samples()}}
\item \href{#method-sequence_data-get_num_sequences}{\code{sequence_data$get_num_sequences()}}
\item \href{#method-sequence_data-get_num_treatments}{\code{sequence_data$get_num_treatments()}}
\item \href{#method-sequence_data-get_oligos}{\code{sequence_data$get_oligos()}}
\item \href{#method-sequence_data-get_bin_taxonomy_report}{\code{sequence_data$get_bin_taxonomy_report()}}
\item \href{#method-sequence_data-get_rabund}{\code{sequence_data$get_rabund()}}
\item \href{#method-sequence_data-get_references}{\code{sequence_data$get_references()}}
\item \href{#method-sequence_data-get_samples}{\code{sequence_data$get_samples()}}
\item \href{#method-sequence_data-get_sample_summary}{\code{sequence_data$get_sample_summary()}}
\item \href{#method-sequence_data-get_scrap_report}{\code{sequence_data$get_scrap_report()}}
\item \href{#method-sequence_data-get_sequence_report}{\code{sequence_data$get_sequence_report()}}
\item \href{#method-sequence_data-get_sequence_taxonomy_report}{\code{sequence_data$get_sequence_taxonomy_report()}}
\item \href{#method-sequence_data-get_sequence_summary}{\code{sequence_data$get_sequence_summary()}}
\item \href{#method-sequence_data-get_sequences}{\code{sequence_data$get_sequences()}}
\item \href{#method-sequence_data-get_shared}{\code{sequence_data$get_shared()}}
\item \href{#method-sequence_data-get_treatments}{\code{sequence_data$get_treatments()}}
\item \href{#method-sequence_data-has_sample}{\code{sequence_data$has_sample()}}
\item \href{#method-sequence_data-is_aligned}{\code{sequence_data$is_aligned()}}
\item \href{#method-sequence_data-remove_lineages}{\code{sequence_data$remove_lineages()}}
\item \href{#method-sequence_data-remove_samples}{\code{sequence_data$remove_samples()}}
\item \href{#method-sequence_data-clone}{\code{sequence_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new sequence dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of dataset (optional)}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}

\item{\code{dataset}}{a `sequence_data` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `sequence_data` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add metadata about your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{a data.frame containing metadata about your dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data$new("my_dataset")
  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)
  dataset$add_metadata(metadata)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_references"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_references}{}}}
\subsection{Method \code{add_references()}}{
Add information about the references used to analyze your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_references(
  reference = NULL,
  reference_name = NULL,
  version = NULL,
  usage = NULL,
  note = NULL,
  url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference}}{a data.frame containing multiple references. The only
required column is reference_name. Optional columns can include: version,
usage, note and url. Default = NULL. Either the reference or
reference_name parameter is required.}

\item{\code{reference_name}}{a string containing the name of the reference. For
example: 'silva.bacteria.fasta' Default = NULL. Either the reference
or reference_name parameter is required.}

\item{\code{version}}{a string containing the version of the reference. For
eaxmple: '138.2' Default = NULL.}

\item{\code{usage}}{a string containing the usage of the reference in
your analysis. For example: 'alignment using mothur2' Default = NULL.}

\item{\code{note}}{a string containing the any additional notes about the
reference. For example: 'custom reference using silva version 1.38.2 with
additional sequences tailored for this analysis'. Default = NULL.}

\item{\code{url}}{a string containing a web address where the reference may be
downloaded. For example: 'https://mothur.org/wiki/silva_reference_files/'
. Default = NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# To add a single reference:
dataset <- sequence_data$new("my_dataset")
dataset$add_references(reference_name = "silva.v4.fasta",
version = "1.38.1", usage = "alignment", note = "custom reference
 created by trimming silva.bacteria.fasta to the V4 region",
 url = "https://mothur.org/wiki/silva_reference_files/")

# To add multiple references:
reference <- readr::read_csv(rdataset_example("references.csv"))
dataset$add_references(reference = reference)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_sequences}{}}}
\subsection{Method \code{add_sequences()}}{
Add new sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_sequences(
  names,
  sequences = NULL,
  comments = NULL,
  reference_name = NULL,
  reference_version = NULL,
  reference_note = NULL,
  reference_url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}

\item{\code{reference_name}}{a string containing the name of the reference used
in the classification of the sequences. For example:
'silva.bacteria.fasta' Default = NULL. (optional)}

\item{\code{reference_version}}{a string containing the version of the reference
used in the preparing of the sequences. For example: '1.38.1'.
Default = NULL. (optional)}

\item{\code{reference_note}}{a string containing the any additional notes about
the reference. For example: 'used for alignment using mothur2 v1.0'.
Default = NULL. (optional)}

\item{\code{reference_url}}{a string containing a web address where the
reference may be downloaded. Default = NULL. (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data$new("my_dataset")
  sequences <- read_fasta(rdataset_example("final.fasta"))
  dataset$add_sequences(sequences$names, sequences$sequences)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- "https://mothur.org/wiki/silva_reference_files/"

dataset <- sequence_data$new("my_dataset")
dataset$add_sequences(sequences$names, sequences$sequences,
reference_name = "silva.bacteria.fasta",
reference_note = "alignment by mothur2 v1.0 using default options",
reference_version = "1.38.1", reference_url = url)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_bins"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_bins}{}}}
\subsection{Method \code{assign_bins()}}{
Add bin data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_bins(
  bin_ids,
  abundances = NULL,
  samples = NULL,
  seq_ids = NULL,
  type = "otu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bin_ids}}{a vector of bin labels}

\item{\code{abundances}}{a vector of abundances (optional). You must provide
either abundances or seq_ids.}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{seq_ids}}{a vector of sequence names (optional) You must provide
either abundances or seq_ids.}

\item{\code{type}}{a string indicating the type of bin assignments. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  # bin_ids  seq_ids
  #  bin1     seq1
  #  bin1     seq2
  #  bin1     seq4
  #  bin2     seq3
  #  bin2     seq6
  #  bin3     seq5

  # To assign sequences to bins:

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  dataset$assign_bins(bin_ids, seq_ids = seq_ids)

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5

  # To add abundance only bin assignments:

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  dataset$assign_bins(bin_ids, abundances)

  # bins would look like:
  #            bin1             bin2        bin3
  # (rabund)   110              525         80

  # To add abundance bin assignments parsed by sample:

  # bin_ids  samples  abundances
  #  bin1     sample1        10
  #  bin1     sample2        100
  #  bin1     sample5        1
  #  bin2     sample1        500
  #  bin2     sample3        25
  #  bin3     sample1        80

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_bins(bin_ids, sample_abundances, samples)

  # (shared) bins would look like:
  # label  sample   bin1   bin2   bin3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  # To assign sequences to bins with their abundances parsed by sample:

  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin1", "bin1", "bin1",
               "bin2", "bin2", "bin2",
               "bin3", "bin3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  dataset$assign_bins(bin_ids, abundances, samples, seq_ids)

  # bins would look like:
  #            bin1             bin2        bin3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5
  # (rabund)   716              85          65

  dataset$get_shared()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_bin_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_bin_taxonomy}{}}}
\subsection{Method \code{assign_bin_taxonomy()}}{
Assign bin classification.

Note, if you assign sequence taxonomies and assign bins, 'sequence_data'
will find the concensus taxonomy for each bin for you.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_bin_taxonomy(
  bin_ids,
  taxonomies,
  type = "otu",
  reference_name = NULL,
  reference_version = NULL,
  reference_note = NULL,
  reference_url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bin_ids}}{a vector of bin names}

\item{\code{taxonomies}}{a vector of bin classifications}

\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}

\item{\code{reference_name}}{a string containing the name of the reference used
in the classification of the bins. For example:
'trainset9_032012.pds.zip' Default = NULL.}

\item{\code{reference_version}}{a string containing the version of the reference
used in the classification of the bins For example: '9_032012'.
Default = NULL.}

\item{\code{reference_note}}{a string containing the any additional notes about
the reference. For example: 'custom reference based on RDP'.
Default = NULL.}

\item{\code{reference_url}}{a string containing a web address where the
reference may be downloaded. Default = NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
bin_ids <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_bins(bin_ids, abunds)
dataset$assign_bin_taxonomy(bin_ids, taxonomies)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- paste("https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset",
         "9_032012.pds.zip", collapse = "")

dataset$assign_bin_taxonomy(bin_ids, taxonomies,
reference_name = "trainset9_032012.pds.zip",
reference_note = "classification by mothur2 v1.0 using default options",
reference_version = "9_032012", reference_url = url)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sequence_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sequence_abundance}{}}}
\subsection{Method \code{assign_sequence_abundance()}}{
Add sequence abundance data with optional sample / treatment assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sequence_abundance(
  names,
  abundances,
  samples = NULL,
  treatments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{abundances}}{a vector of sequence abundances}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{treatments}}{a vector of treatment assignments (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sequence_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sequence_taxonomy}{}}}
\subsection{Method \code{assign_sequence_taxonomy()}}{
Assign sequence classification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sequence_taxonomy(
  names,
  taxonomies,
  reference_name = NULL,
  reference_version = NULL,
  reference_note = NULL,
  reference_url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{taxonomies}}{a vector of sequence classifications}

\item{\code{reference_name}}{a string containing the name of the reference used
in the classification of the sequences. For example:
'trainset9_032012.pds.zip' Default = NULL.}

\item{\code{reference_version}}{a string containing the version of the reference
used in the classification of the sequences. For example: '9_032012'.
Default = NULL.}

\item{\code{reference_note}}{a string containing the any additional notes about
the reference. For example: 'custom reference based on RDP'.
Default = NULL.}

\item{\code{reference_url}}{a string containing a web address where the
reference may be downloaded. Default = NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
names <- c("seq1", "seq2", "seq3", "seq4")
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)

# With the additional parameters to add information about the reference
# You can also add references using the 'add_references' function.

url <- paste("https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset",
         "9_032012.pds.zip", collapse = "")

dataset$assign_sequence_taxonomy(names = names, taxonomies = taxonomies,
reference_name = "trainset9_032012.pds.zip",
reference_note = "classification by mothur2 v1.0 using default options",
reference_version = "9_032012", reference_url = url)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_treatments}{}}}
\subsection{Method \code{assign_treatments()}}{
Assign samples to treatments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_treatments(samples, treatments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{a vector of sample names}

\item{\code{treatments}}{a vector of treatment names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

treatments <- c("early", "early", "late")
dataset$assign_treatments(unique(samples), treatments)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all data from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-export"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-export}{}}}
\subsection{Method \code{export()}}{
Get List containing dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_align_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_align_report}{}}}
\subsection{Method \code{get_align_report()}}{
Get report containing align report table -
ids, search_scores, sim_scores, longest_insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_align_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_contigs_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_contigs_report}{}}}
\subsection{Method \code{get_contigs_report()}}{
Get report containing contigs report table -
ids, lengths, ostarts, oends, olengths, mismatches, numns, ee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_contigs_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_count_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_count_table}{}}}
\subsection{Method \code{get_count_table()}}{
Get count table returns data.frame containing:
ids, abundances, sample(optional), treatment(optional)
This table represents mothur's count and design files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_count_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_dataset_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_dataset_name}{}}}
\subsection{Method \code{get_dataset_name()}}{
Get dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_dataset_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_ids"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_ids}{}}}
\subsection{Method \code{get_ids()}}{
Get names of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_ids(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_list}{}}}
\subsection{Method \code{get_list()}}{
Get data frame containing sequence bin assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_list(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  dataset$assign_bins(bin_ids, sequence_abundances, seq_ids = seq_ids)

  # (list) bins would look like:
  # bin1             bin2        bin3
  # seq1,seq2,seq4   seq3,seq6   seq5

  list <- dataset$get_list()

  #  list$bin_id  list$seq_id
  #  bin1          seq1
  #  bin1          seq2
  #  bin1          seq4
  #  bin2          seq3
  #  bin2          seq6
  #  bin3          seq5

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get data.frame containing metadata for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame()
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)
  dataset$add_metadata(metadata)
  dataset$get_metadata()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_bins"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_bins}{}}}
\subsection{Method \code{get_num_bins()}}{
Get the number of bins in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_bins(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_samples}{}}}
\subsection{Method \code{get_num_samples()}}{
Get number of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_sequences}{}}}
\subsection{Method \code{get_num_sequences()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_sequences(distinct = FALSE, sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distinct}}{Boolean. When distinct is TRUE the number of unique
sequence is returned.}

\item{\code{sample}}{The name of the sample you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_treatments}{}}}
\subsection{Method \code{get_num_treatments()}}{
Get the number of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_oligos"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_oligos}{}}}
\subsection{Method \code{get_oligos()}}{
Get data.frame containing the oligo data
tag, oligo, diffs, oligo(optional), diffs(optional), sample(optional)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_oligos()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_bin_taxonomy_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_bin_taxonomy_report}{}}}
\subsection{Method \code{get_bin_taxonomy_report()}}{
Get report containing the bin taxonomy table -
ids, taxonomy by levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_bin_taxonomy_report(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
bin_ids <- c("bin1", "bin2", "bin3", "bin4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_bins(bin_ids, abunds)
dataset$assign_bin_taxonomy(bin_ids, taxonomies)
dataset$get_bin_taxonomy_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_rabund"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_rabund}{}}}
\subsection{Method \code{get_rabund()}}{
Get data frame containing sequence bin assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_rabund(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin2", "bin3")
  bin_abundances <- c(111, 525, 80)
  dataset$assign_bins(bin_ids, bin_abundances)

  # (rabund) bins would look like:
  # bin1  bin2  bin3
  # 110   525   80

  rabund <- dataset$get_rabund()

  #  rabund$bin_id  rabund$abundance
  #  bin1           111
  #  bin2           525
  #  bin3           80

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_references"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_references}{}}}
\subsection{Method \code{get_references()}}{
Get data.frame containing information about the references used in the
analysis of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_references()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame()
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  reference <- readr::read_csv(rdataset_example("references.csv"))
  dataset$add_references(reference = reference)
  dataset$get_references()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_samples}{}}}
\subsection{Method \code{get_samples()}}{
Get names of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sample_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sample_summary}{}}}
\subsection{Method \code{get_sample_summary()}}{
Get summary of count table. sample, sample_totals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sample_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print sample summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_scrap_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_scrap_report}{}}}
\subsection{Method \code{get_scrap_report()}}{
Get report containing the scrapped sequences / bins -
ids, trash_codes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_scrap_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get data.frame sequence report data. Sequence report data includes: start
positions, end positions, number of bases, number of ambiguous bases,
length of longest homopolymer, and the number of N's.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_taxonomy_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_taxonomy_report}{}}}
\subsection{Method \code{get_sequence_taxonomy_report()}}{
Get report containing the sequence taxonomy table -
ids, taxonomy by levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_taxonomy_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{names <- c("seq1", "seq2", "seq3", "seq4")
tax1 <- "Bacteria(100);Bacteroidetes(95);Bacteroidia(90);"
tax2 <- "Bacteria(100);Proteobacteria(89);Betaproteobacteria(85);"
tax3 <- "Bacteria(100);Firmicutes(99);Bacilli(90);"
tax4 <- "Bacteria(100);Proteobacteria(87);Gammaproteobacteria(82);"
taxonomies <- c(tax1, tax2, tax3, tax4)

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)
dataset$get_sequence_taxonomy_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_summary}{}}}
\subsection{Method \code{get_sequence_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print summaries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequences}{}}}
\subsection{Method \code{get_sequences()}}{
Get vector containing sequence nucleotide strings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequences(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_shared"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_shared}{}}}
\subsection{Method \code{get_shared()}}{
Get data frame containing sequence bin assignments by sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_shared(type = "otu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string indicating the type of clusters. Options
include: "otu", "asv", or "phylotype". Default = "otu".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_bins(bin_ids, sample_abundances, samples)

  # (shared) bins would look like:
  # label  sample   bin1   bin2   bin3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  shared <- dataset$get_shared()

  # shared$bin_id  shared$sample  shared$abundance
  # bin1           sample1        10
  # bin1           sample2        100
  # bin1           sample5        1
  # bin2           sample1        500
  # bin2           sample3        25
  # bin3           sample1        80

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_treatments}{}}}
\subsection{Method \code{get_treatments()}}{
Get names of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-has_sample"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-has_sample}{}}}
\subsection{Method \code{has_sample()}}{
Determine if a sample is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$has_sample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-is_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-is_aligned}{}}}
\subsection{Method \code{is_aligned()}}{
Determine if the dataset is aligned
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$is_aligned()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
bool
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-remove_lineages"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-remove_lineages}{}}}
\subsection{Method \code{remove_lineages()}}{
Remove contaminants from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$remove_lineages(contaminants)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contaminants}}{vector of strings containing the taxonomies you would
like to remove}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
 "Eukaryota")

dataset$remove_lineages(contaminants)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-remove_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-remove_samples}{}}}
\subsection{Method \code{remove_samples()}}{
Remove samples from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$remove_samples(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{vector of strings containing the names of the samples to}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dataset <- read_mothur(fasta = rdataset_example("final.fasta"),
                       count = rdataset_example("final.count_table"),
                       taxonomy = rdataset_example("final.taxonomy"),
                       design = rdataset_example("mouse.time.design"),
                       list = rdataset_example("final.opti_mcc.list"),
                       dataset_name = "miseq_sop")

dataset$get_num_samples()

# To remove samples 'F3D0' and 'F3D1'

dataset$remove_samples(c("F3D0", "F3D1"))

dataset$get_num_samples()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
