% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_data.R
\name{sequence_data}
\alias{sequence_data}
\title{sequence_data}
\description{
'sequence_data' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `sequence_data$new`
## ------------------------------------------------


# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))


## ------------------------------------------------
## Method `sequence_data$add_sequences`
## ------------------------------------------------


  dataset <- sequence_data$new("my_dataset")
  sequences <- read_fasta(rdataset_example("test.fasta"))
  dataset$add_sequences(sequences$names, sequences$sequences)


## ------------------------------------------------
## Method `sequence_data$assign_otus`
## ------------------------------------------------


  # otu_ids  seq_ids
  #  otu1     seq1
  #  otu1     seq2
  #  otu1     seq4
  #  otu2     seq3
  #  otu2     seq6
  #  otu3     seq5

  # To assign sequences to otus:

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  dataset$assign_otus(otu_ids, seq_ids = seq_ids)

  # otus would look like:
  #            otu1             otu2        otu3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5

  # To add abundance only otu assignments:

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu2", "otu3")
  abundances <- c(110, 525, 80)
  dataset$assign_otus(otu_ids, abundances)

  # otus would look like:
  #            otu1             otu2        otu3
  # (rabund)   110              525         80

  # To add abundance otu assignments parsed by sample:

  # otu_ids  samples  abundances
  #  otu1     sample1        10
  #  otu1     sample2        100
  #  otu1     sample5        1
  #  otu2     sample1        500
  #  otu2     sample3        25
  #  otu3     sample1        80

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otus(otu_ids, sample_abundances, samples)

  # (shared) otus would look like:
  # label  sample   otu1   otu2   otu3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  # To assign sequences to otus with their abundances parsed by sample:

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu1", "otu1", "otu1",
               "otu2", "otu2", "otu2",
               "otu3", "otu3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  dataset$assign_otus(otu_ids, abundances, samples, seq_ids)

  # otus would look like:
  #            otu1             otu2        otu3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5
  # (rabund)   716              85          65

  dataset$get_shared()


## ------------------------------------------------
## Method `sequence_data$assign_sequence_abundance`
## ------------------------------------------------


 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)


## ------------------------------------------------
## Method `sequence_data$assign_otu_taxonomy`
## ------------------------------------------------


otu_ids <- c("otu1", "otu2", "otu3", "otu4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_otus(otu_ids, abunds)
dataset$assign_otu_taxonomy(otu_ids, taxonomies)


## ------------------------------------------------
## Method `sequence_data$assign_sequence_taxonomy`
## ------------------------------------------------


names <- c("seq1", "seq2", "seq3", "seq4")
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)


## ------------------------------------------------
## Method `sequence_data$assign_treatments`
## ------------------------------------------------


names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

treatments <- c("early", "early", "late")
dataset$assign_treatments(unique(samples), treatments)


## ------------------------------------------------
## Method `sequence_data$get_list`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  dataset$assign_otus(otu_ids, sequence_abundances, seq_ids = seq_ids)

  # (list) otus would look like:
  # otu1             otu2        otu3
  # seq1,seq2,seq4   seq3,seq6   seq5

  list <- dataset$get_list()

  #  list$otu_id  list$seq_id
  #  otu1          seq1
  #  otu1          seq2
  #  otu1          seq4
  #  otu2          seq3
  #  otu2          seq6
  #  otu3          seq5


## ------------------------------------------------
## Method `sequence_data$get_otu_taxonomy_report`
## ------------------------------------------------


otu_ids <- c("otu1", "otu2", "otu3", "otu4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_otus(otu_ids, abunds)
dataset$assign_otu_taxonomy(otu_ids, taxonomies)
dataset$get_otu_taxonomy_report()


## ------------------------------------------------
## Method `sequence_data$get_rabund`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu2", "otu3")
  otu_abundances <- c(111, 525, 80)
  dataset$assign_otus(otu_ids, otu_abundances)

  # (rabund) otus would look like:
  # otu1  otu2  otu3
  # 110   525   80

  rabund <- dataset$get_rabund()

  #  rabund$otu_id  rabund$abundance
  #  otu1           111
  #  otu2           525
  #  otu3           80


## ------------------------------------------------
## Method `sequence_data$get_sequence_taxonomy_report`
## ------------------------------------------------

names <- c("seq1", "seq2", "seq3", "seq4")
tax1 <- "Bacteria(100);Bacteroidetes(95);Bacteroidia(90);"
tax2 <- "Bacteria(100);Proteobacteria(89);Betaproteobacteria(85);"
tax3 <- "Bacteria(100);Firmicutes(99);Bacilli(90);"
tax4 <- "Bacteria(100);Proteobacteria(87);Gammaproteobacteria(82);"
taxonomies <- c(tax1, tax2, tax3, tax4)

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)
dataset$get_sequence_taxonomy_report()


## ------------------------------------------------
## Method `sequence_data$get_shared`
## ------------------------------------------------

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otus(otu_ids, sample_abundances, samples)

  # (shared) otus would look like:
  # label  sample   otu1   otu2   otu3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  shared <- dataset$get_shared()

  # shared$otu_id  shared$sample  shared$abundance
  # otu1           sample1        10
  # otu1           sample2        100
  # otu1           sample5        1
  # otu2           sample1        500
  # otu2           sample3        25
  # otu3           sample1        80

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{pointer to 'Dataset' (Rcpp Module). This allows package
developers an easy access point to the underlying C++ code with
additional functionality.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_data-new}{\code{sequence_data$new()}}
\item \href{#method-sequence_data-print}{\code{sequence_data$print()}}
\item \href{#method-sequence_data-add_sequences}{\code{sequence_data$add_sequences()}}
\item \href{#method-sequence_data-assign_otus}{\code{sequence_data$assign_otus()}}
\item \href{#method-sequence_data-assign_sequence_abundance}{\code{sequence_data$assign_sequence_abundance()}}
\item \href{#method-sequence_data-assign_otu_taxonomy}{\code{sequence_data$assign_otu_taxonomy()}}
\item \href{#method-sequence_data-assign_sequence_taxonomy}{\code{sequence_data$assign_sequence_taxonomy()}}
\item \href{#method-sequence_data-assign_treatments}{\code{sequence_data$assign_treatments()}}
\item \href{#method-sequence_data-clear}{\code{sequence_data$clear()}}
\item \href{#method-sequence_data-export}{\code{sequence_data$export()}}
\item \href{#method-sequence_data-get_align_report}{\code{sequence_data$get_align_report()}}
\item \href{#method-sequence_data-get_contigs_report}{\code{sequence_data$get_contigs_report()}}
\item \href{#method-sequence_data-get_count_table}{\code{sequence_data$get_count_table()}}
\item \href{#method-sequence_data-get_dataset_name}{\code{sequence_data$get_dataset_name()}}
\item \href{#method-sequence_data-get_ids}{\code{sequence_data$get_ids()}}
\item \href{#method-sequence_data-get_list}{\code{sequence_data$get_list()}}
\item \href{#method-sequence_data-get_num_otus}{\code{sequence_data$get_num_otus()}}
\item \href{#method-sequence_data-get_num_samples}{\code{sequence_data$get_num_samples()}}
\item \href{#method-sequence_data-get_num_sequences}{\code{sequence_data$get_num_sequences()}}
\item \href{#method-sequence_data-get_num_treatments}{\code{sequence_data$get_num_treatments()}}
\item \href{#method-sequence_data-get_oligos}{\code{sequence_data$get_oligos()}}
\item \href{#method-sequence_data-get_otu_taxonomy_report}{\code{sequence_data$get_otu_taxonomy_report()}}
\item \href{#method-sequence_data-get_rabund}{\code{sequence_data$get_rabund()}}
\item \href{#method-sequence_data-get_samples}{\code{sequence_data$get_samples()}}
\item \href{#method-sequence_data-get_sample_summary}{\code{sequence_data$get_sample_summary()}}
\item \href{#method-sequence_data-get_scrap_report}{\code{sequence_data$get_scrap_report()}}
\item \href{#method-sequence_data-get_sequence_report}{\code{sequence_data$get_sequence_report()}}
\item \href{#method-sequence_data-get_sequence_taxonomy_report}{\code{sequence_data$get_sequence_taxonomy_report()}}
\item \href{#method-sequence_data-get_sequence_summary}{\code{sequence_data$get_sequence_summary()}}
\item \href{#method-sequence_data-get_sequences}{\code{sequence_data$get_sequences()}}
\item \href{#method-sequence_data-get_shared}{\code{sequence_data$get_shared()}}
\item \href{#method-sequence_data-get_treatments}{\code{sequence_data$get_treatments()}}
\item \href{#method-sequence_data-has_sample}{\code{sequence_data$has_sample()}}
\item \href{#method-sequence_data-is_aligned}{\code{sequence_data$is_aligned()}}
\item \href{#method-sequence_data-clone}{\code{sequence_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new sequence dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$new(
  name = "",
  fasta = NULL,
  processors = parallelly::availableCores()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of dataset (optional)}

\item{\code{fasta}}{String name of FASTA file}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `sequence_data` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty dataset, run the following:

dataset <- sequence_data$new("soil")

# to create a dataset containing sequences from your fasta file,
#  run the following:

dataset <- sequence_data$new(name = "soil",
                             fasta = rdataset_example("test.fasta"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-add_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-add_sequences}{}}}
\subsection{Method \code{add_sequences()}}{
Add new sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$add_sequences(names, sequences = NULL, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data$new("my_dataset")
  sequences <- read_fasta(rdataset_example("test.fasta"))
  dataset$add_sequences(sequences$names, sequences$sequences)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_otus"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_otus}{}}}
\subsection{Method \code{assign_otus()}}{
Add otu data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_otus(
  otu_ids,
  abundances = NULL,
  samples = NULL,
  seq_ids = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{otu_ids}}{a vector of otu labels}

\item{\code{abundances}}{a vector of abundances (optional). You must provide
either abundances or seq_ids.}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{seq_ids}}{a vector of sequence names (optional) You must provide
either abundances or seq_ids.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  # otu_ids  seq_ids
  #  otu1     seq1
  #  otu1     seq2
  #  otu1     seq4
  #  otu2     seq3
  #  otu2     seq6
  #  otu3     seq5

  # To assign sequences to otus:

  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  dataset$assign_otus(otu_ids, seq_ids = seq_ids)

  # otus would look like:
  #            otu1             otu2        otu3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5

  # To add abundance only otu assignments:

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu2", "otu3")
  abundances <- c(110, 525, 80)
  dataset$assign_otus(otu_ids, abundances)

  # otus would look like:
  #            otu1             otu2        otu3
  # (rabund)   110              525         80

  # To add abundance otu assignments parsed by sample:

  # otu_ids  samples  abundances
  #  otu1     sample1        10
  #  otu1     sample2        100
  #  otu1     sample5        1
  #  otu2     sample1        500
  #  otu2     sample3        25
  #  otu3     sample1        80

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otus(otu_ids, sample_abundances, samples)

  # (shared) otus would look like:
  # label  sample   otu1   otu2   otu3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  # To assign sequences to otus with their abundances parsed by sample:

  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu1", "otu1", "otu1",
               "otu2", "otu2", "otu2",
               "otu3", "otu3")
  seq_ids <- c("seq1", "seq1", "seq1", "seq2", "seq4", "seq4",
               "seq3", "seq3", "seq6",
               "seq5", "seq5")
  samples <- c("sample1", "sample2", "sample5",
               "sample1", "sample3", "sample4",
               "sample2", "sample3", "sample1",
               "sample1", "sample6")
  abundances <- c(10, 100, 1, 500, 25, 80, 20, 5, 60, 15, 50)
  dataset$assign_otus(otu_ids, abundances, samples, seq_ids)

  # otus would look like:
  #            otu1             otu2        otu3
  # (list)     seq1,seq2,seq4   seq3,seq6   seq5
  # (rabund)   716              85          65

  dataset$get_shared()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sequence_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sequence_abundance}{}}}
\subsection{Method \code{assign_sequence_abundance()}}{
Add sequence abundance data with optional sample / treatment assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sequence_abundance(
  names,
  abundances,
  samples = NULL,
  treatments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{abundances}}{a vector of sequence abundances}

\item{\code{samples}}{a vector of sample assignments (optional)}

\item{\code{treatments}}{a vector of treatment assignments (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_otu_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_otu_taxonomy}{}}}
\subsection{Method \code{assign_otu_taxonomy()}}{
Assign otu classification.

Note, if you assign sequence taxonomies and assign otus, 'sequence_data'
will find the concensus taxonomy for each otu for you.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_otu_taxonomy(otu_ids, taxonomies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{otu_ids}}{a vector of otu names}

\item{\code{taxonomies}}{a vector of otu classifications}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
otu_ids <- c("otu1", "otu2", "otu3", "otu4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_otus(otu_ids, abunds)
dataset$assign_otu_taxonomy(otu_ids, taxonomies)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_sequence_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_sequence_taxonomy}{}}}
\subsection{Method \code{assign_sequence_taxonomy()}}{
Assign sequence classification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_sequence_taxonomy(names, taxonomies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{taxonomies}}{a vector of sequence classifications}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
names <- c("seq1", "seq2", "seq3", "seq4")
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-assign_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-assign_treatments}{}}}
\subsection{Method \code{assign_treatments()}}{
Assign samples to treatments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$assign_treatments(samples, treatments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{a vector of sample names}

\item{\code{treatments}}{a vector of treatment names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
names <- c("seq1", "seq1", "seq1", "seq2", "seq2",
             "seq2", "seq3", "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4",
            "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data$new("my_dataset")
unique_names <- unique(names)
dataset$add_sequences(unique_names)
dataset$assign_sequence_abundance(names, abundances, samples)

treatments <- c("early", "early", "late")
dataset$assign_treatments(unique(samples), treatments)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all data from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-export"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-export}{}}}
\subsection{Method \code{export()}}{
Get List containing dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_align_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_align_report}{}}}
\subsection{Method \code{get_align_report()}}{
Get report containing align report table -
ids, search_scores, sim_scores, longest_insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_align_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_contigs_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_contigs_report}{}}}
\subsection{Method \code{get_contigs_report()}}{
Get report containing contigs report table -
ids, lengths, ostarts, oends, olengths, mismatches, numns, ee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_contigs_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_count_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_count_table}{}}}
\subsection{Method \code{get_count_table()}}{
Get count table returns data.frame containing:
ids, abundances, sample(optional), treatment(optional)
This table represents mothur's count and design files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_count_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_dataset_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_dataset_name}{}}}
\subsection{Method \code{get_dataset_name()}}{
Get dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_dataset_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_ids"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_ids}{}}}
\subsection{Method \code{get_ids()}}{
Get names of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_ids(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_list}{}}}
\subsection{Method \code{get_list()}}{
Get data frame containing sequence otu assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  seq_ids <- c("seq1", "seq2", "seq4", "seq3", "seq6", "seq5")
  sequence_abundances <- c(10, 100, 1, 500, 25, 80)
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  dataset$assign_otus(otu_ids, sequence_abundances, seq_ids = seq_ids)

  # (list) otus would look like:
  # otu1             otu2        otu3
  # seq1,seq2,seq4   seq3,seq6   seq5

  list <- dataset$get_list()

  #  list$otu_id  list$seq_id
  #  otu1          seq1
  #  otu1          seq2
  #  otu1          seq4
  #  otu2          seq3
  #  otu2          seq6
  #  otu3          seq5

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_otus"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_otus}{}}}
\subsection{Method \code{get_num_otus()}}{
Get the number of otus in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_otus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_samples}{}}}
\subsection{Method \code{get_num_samples()}}{
Get number of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_sequences}{}}}
\subsection{Method \code{get_num_sequences()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_sequences(distinct = FALSE, sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distinct}}{Boolean. When distinct is TRUE the number of unique
sequence is returned.}

\item{\code{sample}}{The name of the sample you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_num_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_num_treatments}{}}}
\subsection{Method \code{get_num_treatments()}}{
Get the number of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_num_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_oligos"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_oligos}{}}}
\subsection{Method \code{get_oligos()}}{
Get data.frame containing the oligo data
tag, oligo, diffs, oligo(optional), diffs(optional), sample(optional)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_oligos()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_otu_taxonomy_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_otu_taxonomy_report}{}}}
\subsection{Method \code{get_otu_taxonomy_report()}}{
Get report containing the sequence taxonomy table -
ids, taxonomy by levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_otu_taxonomy_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
otu_ids <- c("otu1", "otu2", "otu3", "otu4")
abunds <- c(200, 40, 100, 5)
taxonomies <- c("Bacteria;Bacteroidetes;Bacteroidia;Bacteroidales;",
              "Bacteria;Proteobacteria;Betaproteobacteria;Neisseriales;",
               "Bacteria;Firmicutes;Bacilli;Lactobacillales;",
           "Bacteria;Proteobacteria;Gammaproteobacteria;Pasteurellales;")

dataset <- sequence_data$new("my_dataset")
dataset$assign_otus(otu_ids, abunds)
dataset$assign_otu_taxonomy(otu_ids, taxonomies)
dataset$get_otu_taxonomy_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_rabund"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_rabund}{}}}
\subsection{Method \code{get_rabund()}}{
Get data frame containing sequence otu assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_rabund()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu2", "otu3")
  otu_abundances <- c(111, 525, 80)
  dataset$assign_otus(otu_ids, otu_abundances)

  # (rabund) otus would look like:
  # otu1  otu2  otu3
  # 110   525   80

  rabund <- dataset$get_rabund()

  #  rabund$otu_id  rabund$abundance
  #  otu1           111
  #  otu2           525
  #  otu3           80

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_samples"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_samples}{}}}
\subsection{Method \code{get_samples()}}{
Get names of samples in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sample_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sample_summary}{}}}
\subsection{Method \code{get_sample_summary()}}{
Get summary of count table. sample, sample_totals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sample_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print sample summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_scrap_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_scrap_report}{}}}
\subsection{Method \code{get_scrap_report()}}{
Get report containing the scrapped sequences / otus -
ids, trash_codes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_scrap_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get data.frame sequence report data. Sequence report data includes: start
positions, end positions, number of bases, number of ambiguous bases,
length of longest homopolymer, and the number of N's.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_taxonomy_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_taxonomy_report}{}}}
\subsection{Method \code{get_sequence_taxonomy_report()}}{
Get report containing the sequence taxonomy table -
ids, taxonomy by levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_taxonomy_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{names <- c("seq1", "seq2", "seq3", "seq4")
tax1 <- "Bacteria(100);Bacteroidetes(95);Bacteroidia(90);"
tax2 <- "Bacteria(100);Proteobacteria(89);Betaproteobacteria(85);"
tax3 <- "Bacteria(100);Firmicutes(99);Bacilli(90);"
tax4 <- "Bacteria(100);Proteobacteria(87);Gammaproteobacteria(82);"
taxonomies <- c(tax1, tax2, tax3, tax4)

dataset <- sequence_data$new("my_dataset")
dataset$assign_sequence_taxonomy(names, taxonomies)
dataset$get_sequence_taxonomy_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequence_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequence_summary}{}}}
\subsection{Method \code{get_sequence_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequence_summary(silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Default = FALSE, meaning print summaries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_sequences"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_sequences}{}}}
\subsection{Method \code{get_sequences()}}{
Get vector containing sequence nucleotide strings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_sequences(sample = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_shared"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_shared}{}}}
\subsection{Method \code{get_shared()}}{
Get data frame containing sequence otu assignments by sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_shared()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data$new("my_dataset")
  otu_ids <- c("otu1", "otu1", "otu1", "otu2", "otu2", "otu3")
  samples <- c("sample1", "sample2", "sample5",
   "sample1", "sample3", "sample1")
  sample_abundances <- c(10, 100, 1, 500, 25, 80)
  dataset$assign_otus(otu_ids, sample_abundances, samples)

  # (shared) otus would look like:
  # label  sample   otu1   otu2   otu3
  # 0.03   sample1  10     500    80
  # 0.03   sample2  100    0      0
  # 0.03   sample3  0      25     0
  # 0.03   sample5  1      0      0

  shared <- dataset$get_shared()

  # shared$otu_id  shared$sample  shared$abundance
  # otu1           sample1        10
  # otu1           sample2        100
  # otu1           sample5        1
  # otu2           sample1        500
  # otu2           sample3        25
  # otu3           sample1        80

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-get_treatments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-get_treatments}{}}}
\subsection{Method \code{get_treatments()}}{
Get names of treatments in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$get_treatments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-has_sample"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-has_sample}{}}}
\subsection{Method \code{has_sample()}}{
Determine if a sample is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$has_sample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-is_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-is_aligned}{}}}
\subsection{Method \code{is_aligned()}}{
Determine if the dataset is aligned
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$is_aligned()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
bool
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
