% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_bin_abundance}
\alias{set_bin_abundance}
\title{set_bin_abundance}
\usage{
set_bin_abundance(data, bin_names, abunds, reason = "update", type = "otu")
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{bin_names}{a vector strings containing of bin names to set the
abundances for.}

\item{abunds}{vector of integers containing the abundances of each bin.}

\item{reason}{a string containing the trash tag to be applied to any bins
set to 0 abundance. Default = "update".}

\item{type}{a string indicating the type of clusters. Default = "otu".}
}
\description{
Set bin abundances for an instance of the 'Dataset' class without sample data
}
\examples{

  dataset <- new_dataset("my_dataset", 4)
  bin_ids <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)
  assign_bins(dataset, bin_ids, abundances, "", "")

  bins <- c("bin1", "bin2")
  new_abunds <- c(300, 250)

  set_bin_abundance(dataset, bins, new_abunds)
  get_bin_abundance(dataset, "bin1")

}
