% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_dataset_name}
\alias{set_dataset_name}
\title{set_dataset_name}
\usage{
set_dataset_name(data, dataset_name)
}
\arguments{
\item{data}{an Rcpp::XPtr<Dataset> pointer to an instance of the
'Dataset' c++ class.}

\item{dataset_name, }{a string containing the desired name}
}
\description{
Set the name of an instance of the 'Dataset' class.
}
\examples{
dataset <- new_dataset("my_dataset", 4)
set_dataset_name(dataset, "new_dataset_name")

}
