% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_num_processors}
\alias{set_num_processors}
\title{set_num_processors}
\usage{
set_num_processors(data, processors)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{processors, }{a integer containing the desired number of processors}
}
\description{
Set the number of processors used to summarize a \link{dataset} object
}
\examples{

data <- new_dataset("my_dataset", 2)
set_num_processors(data, 1)

}
