% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strollur-package.R
\docType{package}
\name{strollur-package}
\alias{strollur}
\alias{strollur-package}
\title{strollur: Store and transfer data used in the analysis of microbial RNA}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'strollur' package stores the data associated with your microbial RNA analysis. This includes nucleotide sequences, abundance, sample and treatment assignments, taxonomic classifications, asv, otu and phylotype clusters, metadata, trees and various reports. It is designed to facilitate data analysis across multiple R packages with utility functions to read / write from [mothur](https://mothur.org), [qiime2](https://qiime2.org), [dada2](https://benjjneb.github.io/dada2/) and [phyloseq](https://www.bioconductor.org/packages/release/bioc/html/phyloseq.html).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mothur/strollur}
  \item \url{https://mothur.org/strollur/}
  \item Report bugs at \url{https://github.com/mothur/strollur/issues}
}

}
\author{
\strong{Maintainer}: Sarah Westcott \email{swestcot@umich.edu}

}
\keyword{internal}
