% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_strollur.R
\name{strollur}
\alias{strollur}
\title{The \code{strollur} object stores the data associated with your amplicon
sequence analysis.}
\description{
'strollur' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `strollur$new()`
## ------------------------------------------------


# to create an empty strollur object, run the following:

data <- new_dataset("soil")


## ------------------------------------------------
## Method `strollur$print()`
## ------------------------------------------------

miseq <- load_dataset(strollur_example("miseq_sop.rds"))
miseq


## ------------------------------------------------
## Method `strollur$abundance()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To the total abundance for each sequence
miseq$abundance(type = "sequence") |> head(n = 5)

# To the total abundance for each sequence parsed by sample
miseq$abundance(type = "sequence", by_sample = TRUE) |> head(n = 5)

# To the total abundance for each "otu" bin
miseq$abundance(type = "bin", bin_type = "otu") |> head(n = 5)

# To the total abundance for each "otu" bin parsed by sample
miseq$abundance(type = "bin", bin_type = "otu", by_sample = TRUE) |>
head(n = 5)

# To the total abundance for each "asv" bin
miseq$abundance(type = "bin", bin_type = "asv") |> head(n = 5)

# To the total abundance for each "asv" bin parsed by sample
miseq$abundance(type = "bin", bin_type = "asv", by_sample = TRUE) |>
head(n = 5)

# To the total abundance for each sample
miseq$abundance(type = "sample") |> head(n = 5)

# To the total abundance for each treatment
miseq$abundance(type = "treatment")


## ------------------------------------------------
## Method `strollur$add()`
## ------------------------------------------------


fasta_data <- read_fasta(fasta = strollur_example("final.fasta.gz"))
contigs_report <- readRDS(strollur_example("miseq_contigs_report.rds"))

# Create a new empty `strollur` object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

data$add(table = fasta_data, type = "sequence")
data$add(
  table = contigs_report, type = "report",
  report_type = "contigs_report", list(sequence_name = "Name")
)

# To add metadata related to your study

metadata <- readRDS(strollur_example("miseq_metadata.rds"))

data$add(table = metadata, type = "report", report_type = "metadata")

# To add a tree relating to your sequences

tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))

data$add(table = tree, type = "sequence_tree")

# To add a tree relating to your samples
data <- new_dataset(dataset_name = "example_dataset")

df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
data$assign(table = df, type = "bin", bin_type = "otu")

tree <- ape::read.tree(strollur_example("final.opti_mcc.jclass.ave.tre"))

data$add(table = tree, type = "sample_tree")


## ------------------------------------------------
## Method `strollur$alignment_length()`
## ------------------------------------------------


data <- miseq_sop_example()
data$alignment_length()


## ------------------------------------------------
## Method `strollur$assign()`
## ------------------------------------------------


# create a new empty strollur object named 'example_dataset'

data <- new_dataset(dataset_name = "example_dataset")

# Assign sequence abundances

abundance_by_sample <- read_mothur_count(strollur_example(
  "final.count_table.gz"
))

data$assign(table = abundance_by_sample, type = "sequence_abundance")

# Assign sequence classifications

sequence_classifications <- read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

data$assign(table = sequence_classifications, type = "sequence_taxonomy")

# Assigning bins

# read mothur's otu list file into data.frame
otu_data <- read_mothur_list(list = strollur_example(
  "final.opti_mcc.list.gz"
))

# read mothur's asv list file into data.frame
asv_data <- read_mothur_list(list = strollur_example(
  "final.asv.list.gz"
))

# read mothur's phylotype list file into data.frame
phylo_data <- read_mothur_list(list = strollur_example(
  "final.tx.list.gz"
))

# read otu bin representative sequences into a data.frame
bin_reps <- readRDS(strollur_example(
                        "miseq_representative_sequences.rds"))

# assign 'otu' bins using sequence names
data$assign(table = otu_data, bin_type = "otu")

# assign 'asv' bins using sequence names
data$assign(table = asv_data, bin_type = "asv")

# assign 'phylotype' bins using sequence names
data$assign(table = phylo_data, bin_type = "phylotype")

# assign 'otu' bin representative sequences
data$assign(table = bin_reps, type = "bin_representative")

# To assign abundance only bins

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# read mothur's shared file
otu_data <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# assign abundance only otus parsed by sample
data$assign(table = otu_data, bin_type = "otu")

# Assigning bin classifications

# read bin taxonomies
otu_data <- read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

# assign otu consensus taxonomies
data$assign(
  table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

# Assign treatments

sample_assignments <- readRDS(
   strollur_example("miseq_sample_design.rds"))

data$assign(table = sample_assignments, type = "treatment")


## ------------------------------------------------
## Method `strollur$clear()`
## ------------------------------------------------

miseq <- load_dataset(strollur_example("miseq_sop.rds"))
miseq
miseq$clear()
miseq

## ------------------------------------------------
## Method `strollur$count()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the total number of sequences
miseq$count(type = "sequence")

# To get number of unique sequences
miseq$count(type = "sequence", distinct = TRUE)

# To get number of unique sequences from samples 'F3D0' and 'F3D1'
# Note these sequences will be present in both samples but may be
# be present in other samples as well
miseq$count(type = "sequence", samples = c("F3D0", "F3D1"))

# To get number of unique sequences exclusive to samples 'F3D0' and
# 'F3D1'. Note sequences are present in both samples and NOT present in
# any other samples.

miseq$count(type = "sequence",
            samples = c("F3D0", "F3D1"), distinct = TRUE )

# To get the number of samples in the dataset
miseq$count(type = "sample")

# To get the number of treatments in the dataset
miseq$count(type = "treatment")

# To get the number of "otu" bins in the dataset
miseq$count(type = "bin", bin_type = "otu")

# To get the number of "asv" bins in the dataset
miseq$count(type = "bin", bin_type = "asv")

# To get the number of "phylotype" bins in the dataset
miseq$count(type = "bin", bin_type = "phylotype")

# To get number of "otu" bins from samples 'F3D0' and 'F3D1'
# Note these bins will have sequences from both samples but there may be
# other samples present as well
miseq$count(
  type = "bin", bin_type = "otu", samples = c("F3D0", "F3D1")
)

# To get number of "otu" bins unique to samples 'F3D0' and 'F3D1'
# Note these bins will have sequences from both samples and NO other
# samples will be present in the bins.

miseq$count(
  type = "bin", bin_type = "otu",
  samples = c("F3D0", "F3D1"), distinct = TRUE
)


## ------------------------------------------------
## Method `strollur$get_bin_types()`
## ------------------------------------------------


data <- miseq_sop_example()
data$get_bin_types()


## ------------------------------------------------
## Method `strollur$get_version()`
## ------------------------------------------------


data <- new_dataset("test")

data$get_version()


## ------------------------------------------------
## Method `strollur$has_bin_taxonomy()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

miseq$has_bin_taxonomy(bin_type = "otu")


## ------------------------------------------------
## Method `strollur$has_sequence_taxonomy()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

miseq$has_sequence_taxonomy()


## ------------------------------------------------
## Method `strollur$is_equal()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

data <- copy_dataset(miseq)

miseq$is_equal(data)


## ------------------------------------------------
## Method `strollur$names()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the name of the dataset
miseq$names(type = "dataset")

# To get the names of the sequences
miseq$names(type = "sequence")

# To get the names of the sequences present sample 'F3D0'
miseq$names(type = "sequence", samples = c("F3D0"))

#' # To get the names of the sequences unique to sample 'F3D0'
miseq$names(type = "sequence", samples = c("F3D0"), distinct = TRUE)

# To get the names of the samples
miseq$names(type = "sample")

# To get the names of the treatments
miseq$names(type = "treatment")

# To get the names of the bins
miseq$names(type = "bin")

# To get the names of the bins that are unique to 'F3D0'
miseq$names(type = "bin", samples = c("F3D0"), distinct = TRUE)

# To get the names of the bins that include sequences from 'F3D0'
miseq$names(type = "bin", samples = c("F3D0"), distinct = FALSE)

# To get the names of the reports
miseq$names(type = "report")


## ------------------------------------------------
## Method `strollur$report()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the FASTA data

miseq$report(type = "fasta") |> head(n = 5)

# To get a report about the FASTA data

miseq$report(type = "sequence") |> head(n = 5)

# To get the sequence bin assignments

miseq$report(type = "sequence_bin_assignment", bin_type = "otu") |>
head(n = 5)

# To get the sample treatment assignments

miseq$report(type = "sample_assignment") |> head(n = 5)

# To get a report about sequence classifications

miseq$report(type = "sequence_taxonomy") |> head(n = 5)

# To get a report about bin classifications for 'otu' data

miseq$report(type = "bin_taxonomy", bin_type = "otu") |> head(n = 5)

# To get the 'otu' bin representative sequences

miseq$report(type = "bin_representative", bin_type = "otu") |>
head(n = 5)

# To get a report about the sequences removed during your analysis:

miseq$report(type = "sequence_scrap")

# To get a report about the "otu" bins removed during your analysis:

miseq$report(type = "bin_scrap", bin_type = "otu")

# To get the metadata associated with your data:

metadata <- miseq$report(type = "metadata") |> head(n = 5)

# To get the resource references associated with your data:

references <- miseq$report(type = "resource_reference")

# To get our custom report containing the contigs assembly data:

miseq$report(type = "contigs_report") |> head(n = 5)

# To get the tree that relates your sequences:

sequence_tree <- miseq$report(type = "sequence_tree")

# To get the tree that relates your sequences:

sample_tree <- miseq$report(type = "sample_tree")


## ------------------------------------------------
## Method `strollur$summary()`
## ------------------------------------------------


miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the summary of your FASTA data
miseq$summary(type = "sequence")

# summarize contigs_report
miseq$summary(type = "report", report_type = "contigs_report")

# remove sample 'F3D0' to produce a scrap report
xdev_remove_samples(data = miseq, samples = c("F3D0"))

# summarize scrapped data -
# sequences and bins scrapped by removing the sample "F3D0"
miseq$summary(type = "scrap")

}
\seealso{
\link{taxonomy_table_format}
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{data}}{Rcpp::XPtr\if{html}{\out{<Dataset>}} pointer to 'Dataset' c++ class. This
allows package developers an easy access point to the underlying C++ code
with additional functionality.}

    \item{\code{raw}}{Rcpp::RawVector containing the serialized data of the
'Dataset' c++ class. This allows the load and save functions to work with
the class.}

    \item{\code{sequence_tree}}{a tree that relates sequences to each other}

    \item{\code{sample_tree}}{a tree that relates samples to each other}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-strollur-initialize}{\code{strollur$new()}}
    \item \href{#method-strollur-print}{\code{strollur$print()}}
    \item \href{#method-strollur-abundance}{\code{strollur$abundance()}}
    \item \href{#method-strollur-add}{\code{strollur$add()}}
    \item \href{#method-strollur-alignment_length}{\code{strollur$alignment_length()}}
    \item \href{#method-strollur-assign}{\code{strollur$assign()}}
    \item \href{#method-strollur-clear}{\code{strollur$clear()}}
    \item \href{#method-strollur-count}{\code{strollur$count()}}
    \item \href{#method-strollur-get_bin_types}{\code{strollur$get_bin_types()}}
    \item \href{#method-strollur-get_version}{\code{strollur$get_version()}}
    \item \href{#method-strollur-has_bin_taxonomy}{\code{strollur$has_bin_taxonomy()}}
    \item \href{#method-strollur-has_sequence_taxonomy}{\code{strollur$has_sequence_taxonomy()}}
    \item \href{#method-strollur-is_equal}{\code{strollur$is_equal()}}
    \item \href{#method-strollur-names}{\code{strollur$names()}}
    \item \href{#method-strollur-report}{\code{strollur$report()}}
    \item \href{#method-strollur-summary}{\code{strollur$summary()}}
    \item \href{#method-strollur-clone}{\code{strollur$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-initialize}{}}}
\subsection{\code{strollur$new()}}{
  Create a new strollur dataset
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$new(name = "", dataset = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{String, name of dataset (optional)}
      \item{\code{dataset}}{a \code{strollur} object.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{strollur} object.
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# to create an empty strollur object, run the following:

data <- new_dataset("soil")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-print"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-print}{}}}
\subsection{\code{strollur$print()}}{
  Print summary of \code{strollur} object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$print()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    No return value, called for side effects.
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))
miseq
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-abundance"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-abundance}{}}}
\subsection{\code{strollur$abundance()}}{
  Get the abundance data for sequences, bins, samples, and treatments.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$abundance(type = "sequence", bin_type = "otu", by_sample = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{string containing the type of data you want the number of.
Options include: "sequence", "bin", "sample" and "treatment".
Default = "sequence".}
      \item{\code{bin_type}}{string containing the bin type you would like the
abundance data for. Default = "otu".}
      \item{\code{by_sample}}{Boolean. When by_sample is TRUE, the abundance data
will be parsed by sample. Default = FALSE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    data.frame
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To the total abundance for each sequence
miseq$abundance(type = "sequence") |> head(n = 5)

# To the total abundance for each sequence parsed by sample
miseq$abundance(type = "sequence", by_sample = TRUE) |> head(n = 5)

# To the total abundance for each "otu" bin
miseq$abundance(type = "bin", bin_type = "otu") |> head(n = 5)

# To the total abundance for each "otu" bin parsed by sample
miseq$abundance(type = "bin", bin_type = "otu", by_sample = TRUE) |>
head(n = 5)

# To the total abundance for each "asv" bin
miseq$abundance(type = "bin", bin_type = "asv") |> head(n = 5)

# To the total abundance for each "asv" bin parsed by sample
miseq$abundance(type = "bin", bin_type = "asv", by_sample = TRUE) |>
head(n = 5)

# To the total abundance for each sample
miseq$abundance(type = "sample") |> head(n = 5)

# To the total abundance for each treatment
miseq$abundance(type = "treatment")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add}{}}}
\subsection{\code{strollur$add()}}{
  Add sequences, reports or resource references
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$add(
  table,
  type = "sequence",
  report_type = NULL,
  table_names = list(sequence_name = "sequence_name", sequence = "sequence", comment =
    "comment", reference_vendor = "vendor", reference_name = "name", reference_version =
    "version", reference_usage = "usage", reference_note = "note", reference_method_url =
    "method_url", reference_documentation_url = "documentation_url", reference_parameter
    = "parameter", reference_citation = "citation"),
  reference = NULL,
  verbose = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{table}}{a data.frame or tree containing the data you wish to add.}
      \item{\code{type}}{a string containing the type of data. Options include:
'sequence', 'sample_tree', 'sequence_tree', 'resource_reference' and
'report'.}
      \item{\code{report_type}}{a string containing the type of report you are
adding.}
      \item{\code{table_names}}{named list used to indicate the names of the columns
in the table. By default:

table_names <- list(sequence_name = "sequence_name",
comment = "comment",
sequence = "sequence",
reference_name = "name",
reference_vendor = "vendor",
reference_version = "version",
reference_usage = "usage",
reference_note = "note",
reference_documentation_url = "documentation_url",
reference_method_url = "method_url",
reference_parameter = "parameter",
reference_citation = "citation")

In table_names, 'sequence_name' is a string containing the name of the
column in 'table' that contains the sequence names. It is used when you
are adding FASTA data. Default column name is 'sequence_name'.

In table_names, 'sequence' is a string containing the name of the column
in 'table' that contains the sequence nucleotide strings. It is used when
you are adding FASTA data. Default column name is 'sequence'.

In table_names, 'comment' is a string containing the name of the column
in 'table' that contains the sequence comments. It is used when you are
adding FASTA data. Default column name is 'comment'.

In table_names, 'reference_vendor' is a string containing the name of the
column in 'table' that contains the reference vendor names. It is used
when ' you are adding reference data. Default column name is 'vendor'.

In table_names, 'reference_name' is a string containing the name of the '
column in 'table' that contains the reference names. It is used when you
are ' adding reference data. Default column name is 'name'.

In table_names, 'reference_version' is a string containing the name of
the ' column in 'table' that contains the reference versions. Default
column name is 'version'.

In table_names, 'reference_usage' is a string containing the name of the
column in 'table' that contains the reference usages. Default column name
is 'usage'.

In table_names, 'reference_note' is a string containing the name of the
column in 'table' that contains the reference notes. Default column name
is 'note'.

In table_names, 'reference_method_url' is a string containing the name of
the column in 'table' that contains the reference method urls. Default
column name is 'method_url'.

In table_names, 'reference_documentation_url' is a string containing the
name of the column in 'table' that contains the reference urls. Default
column name is 'documentation_url'.

In table_names, 'reference_parameter' is a string containing the name of
the column in 'table' that contains the reference parameters. Default
column name is 'parameter'.

In table_names, 'reference_citation' is a string containing the name of
the column in 'table' that contains the reference citations. Default
column name is 'citation'.}
      \item{\code{reference}}{a list created by the function \link{new_reference}.
Optional.}
      \item{\code{verbose}}{boolean indicating whether or not you want progress
messages. Default = TRUE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Updated \code{strollur} object - invisible(self)
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{fasta_data <- read_fasta(fasta = strollur_example("final.fasta.gz"))
contigs_report <- readRDS(strollur_example("miseq_contigs_report.rds"))

# Create a new empty `strollur` object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

data$add(table = fasta_data, type = "sequence")
data$add(
  table = contigs_report, type = "report",
  report_type = "contigs_report", list(sequence_name = "Name")
)

# To add metadata related to your study

metadata <- readRDS(strollur_example("miseq_metadata.rds"))

data$add(table = metadata, type = "report", report_type = "metadata")

# To add a tree relating to your sequences

tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))

data$add(table = tree, type = "sequence_tree")

# To add a tree relating to your samples
data <- new_dataset(dataset_name = "example_dataset")

df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
data$assign(table = df, type = "bin", bin_type = "otu")

tree <- ape::read.tree(strollur_example("final.opti_mcc.jclass.ave.tre"))

data$add(table = tree, type = "sample_tree")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-alignment_length"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-alignment_length}{}}}
\subsection{\code{strollur$alignment_length()}}{
  Get length of alignment
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$alignment_length()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer containing the length of the aligned sequences or -1 if
unaligned.
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{data <- miseq_sop_example()
data$alignment_length()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-assign"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-assign}{}}}
\subsection{\code{strollur$assign()}}{
  Assign sequence abundances, sequence classifications, bins, bin
representative sequences, bin classifications or treatments.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$assign(
  table,
  type = "bin",
  bin_type = "otu",
  table_names = list(sequence_name = "sequence_name", abundance = "abundance", sample =
    "sample", treatment = "treatment", taxonomy = "taxonomy", level = "level", confidence
    = "confidence", bin_name = "bin_name"),
  reference = NULL,
  verbose = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{table}}{a data.frame containing the data you wish to assign}
      \item{\code{type}}{a string containing the type of data. Options include:
'sequence_abundance', 'sequence_taxonomy', 'bin',
'bin_representative', 'bin_taxonomy' and 'treatment'.
Default = "bin".}
      \item{\code{bin_type}}{string containing the bin type you would like the number
of bins for. Default = "otu".}
      \item{\code{table_names}}{named list used to indicate the names of the columns
in the table. By default:

table_names <- list(sequence_name = "sequence_name", abundance =
"abundance", sample = "sample", treatment = "treatment", taxonomy =
"taxonomy", bin_name = "bin_name")

In table_names, 'sequence_name' is a string containing the name of the
column in 'table' that contains the sequence names. Default column name
is 'sequence_name'.

In table_names, 'abundance' is a string containing the name of the
column in 'table' that contains the abundances. Default column name is
'abundance'.

In table_names, 'sample' is a string containing the name of the column
in 'table' that contains the samples. Default column name is 'sample'.

In table_names, 'treatment' is a string containing the name of the
column in 'table' that contains the treatment names. Default column
name is 'treatment'.

In table_names, 'taxonomy' is a string containing the name of the
column in 'table' that contains the classifications. Default column
name is 'taxonomy'.

In table_names, 'level' is a string containing the name of the column in
the data.frame that contains the classifications. Default column name is
'level'.

In table_names, 'confidence' is a string containing the name of the
column in the data.frame that contains the classifications confidence
scores. Default column name is 'confidence'.

In table_names, 'bin_name' is a string containing the name of the
column in 'table' that contains the bin names. Default column name is
'bin_name'.}
      \item{\code{reference}}{a list created by the function \link{new_reference}.
Optional.}
      \item{\code{verbose}}{boolean indicating whether or not you want progress
messages. Default = TRUE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Updated \code{strollur} object
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# create a new empty strollur object named 'example_dataset'

data <- new_dataset(dataset_name = "example_dataset")

# Assign sequence abundances

abundance_by_sample <- read_mothur_count(strollur_example(
  "final.count_table.gz"
))

data$assign(table = abundance_by_sample, type = "sequence_abundance")

# Assign sequence classifications

sequence_classifications <- read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

data$assign(table = sequence_classifications, type = "sequence_taxonomy")

# Assigning bins

# read mothur's otu list file into data.frame
otu_data <- read_mothur_list(list = strollur_example(
  "final.opti_mcc.list.gz"
))

# read mothur's asv list file into data.frame
asv_data <- read_mothur_list(list = strollur_example(
  "final.asv.list.gz"
))

# read mothur's phylotype list file into data.frame
phylo_data <- read_mothur_list(list = strollur_example(
  "final.tx.list.gz"
))

# read otu bin representative sequences into a data.frame
bin_reps <- readRDS(strollur_example(
                        "miseq_representative_sequences.rds"))

# assign 'otu' bins using sequence names
data$assign(table = otu_data, bin_type = "otu")

# assign 'asv' bins using sequence names
data$assign(table = asv_data, bin_type = "asv")

# assign 'phylotype' bins using sequence names
data$assign(table = phylo_data, bin_type = "phylotype")

# assign 'otu' bin representative sequences
data$assign(table = bin_reps, type = "bin_representative")

# To assign abundance only bins

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# read mothur's shared file
otu_data <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# assign abundance only otus parsed by sample
data$assign(table = otu_data, bin_type = "otu")

# Assigning bin classifications

# read bin taxonomies
otu_data <- read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

# assign otu consensus taxonomies
data$assign(
  table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

# Assign treatments

sample_assignments <- readRDS(
   strollur_example("miseq_sample_design.rds"))

data$assign(table = sample_assignments, type = "treatment")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clear"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clear}{}}}
\subsection{\code{strollur$clear()}}{
  Clear data from datasest
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$clear()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Updated \code{strollur} object
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))
miseq
miseq$clear()
miseq
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-count"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-count}{}}}
\subsection{\code{strollur$count()}}{
  Find the number of sequences, samples, treatments or bins of a given type
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$count(
  type = "sequence",
  bin_type = "otu",
  samples = NULL,
  distinct = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{string containing the type of data you want the number of.
Options include: "sequence", "sample", "treatment", "bin", and
"resource_reference". Default = "sequence".}
      \item{\code{bin_type}}{string containing the bin type you would like the number
of bins for. Default = "otu".}
      \item{\code{samples}}{vector of strings. samples is only used when 'type' =
"sequence" or 'type' = "bin" . samples should contain the names of
the samples you want the count for. Default = NULL.}
      \item{\code{distinct}}{Boolean. distinct is used when 'type' = "sequence" or
'type' = "bin". When 'type' = "sequence" and distinct is TRUE the
number of unique sequences is returned. When 'type' = "sequence" and
distinct is FALSE the total number of sequences is returned. This can
also be combined with samples to find the number of unique sequences
found ONLY in a given set of samples, or to find the number of unique
sequences in given set of samples that may also be present in other
samples. When 'type' = "bin", you can set distinct = TRUE to return
the number of bins that ONLY contain sequences from the given samples.
When distinct is FALSE the count returned contains bins with sequences
from a given samples, but those bins may also contain other samples.
Default = FALSE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    double
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the total number of sequences
miseq$count(type = "sequence")

# To get number of unique sequences
miseq$count(type = "sequence", distinct = TRUE)

# To get number of unique sequences from samples 'F3D0' and 'F3D1'
# Note these sequences will be present in both samples but may be
# be present in other samples as well
miseq$count(type = "sequence", samples = c("F3D0", "F3D1"))

# To get number of unique sequences exclusive to samples 'F3D0' and
# 'F3D1'. Note sequences are present in both samples and NOT present in
# any other samples.

miseq$count(type = "sequence",
            samples = c("F3D0", "F3D1"), distinct = TRUE )

# To get the number of samples in the dataset
miseq$count(type = "sample")

# To get the number of treatments in the dataset
miseq$count(type = "treatment")

# To get the number of "otu" bins in the dataset
miseq$count(type = "bin", bin_type = "otu")

# To get the number of "asv" bins in the dataset
miseq$count(type = "bin", bin_type = "asv")

# To get the number of "phylotype" bins in the dataset
miseq$count(type = "bin", bin_type = "phylotype")

# To get number of "otu" bins from samples 'F3D0' and 'F3D1'
# Note these bins will have sequences from both samples but there may be
# other samples present as well
miseq$count(
  type = "bin", bin_type = "otu", samples = c("F3D0", "F3D1")
)

# To get number of "otu" bins unique to samples 'F3D0' and 'F3D1'
# Note these bins will have sequences from both samples and NO other
# samples will be present in the bins.

miseq$count(
  type = "bin", bin_type = "otu",
  samples = c("F3D0", "F3D1"), distinct = TRUE
)
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_bin_types"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_bin_types}{}}}
\subsection{\code{strollur$get_bin_types()}}{
  Get bin table types
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$get_bin_types()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    vector of strings
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{data <- miseq_sop_example()
data$get_bin_types()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_version}{}}}
\subsection{\code{strollur$get_version()}}{
  Get the version of the
\href{https://mothur.org/strollur/reference/strollur.html}{strollur}
object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$get_version()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a logical
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{data <- new_dataset("test")

data$get_version()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-has_bin_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-has_bin_taxonomy}{}}}
\subsection{\code{strollur$has_bin_taxonomy()}}{
  Determine if your dataset contains bin classifications
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$has_bin_taxonomy(bin_type = "otu")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{bin_type}}{string containing the bin type you would like to check
for classifaction data. Default = "otu".}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a logical
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

miseq$has_bin_taxonomy(bin_type = "otu")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-has_sequence_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-has_sequence_taxonomy}{}}}
\subsection{\code{strollur$has_sequence_taxonomy()}}{
  Determine if your dataset contains sequence classifications
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$has_sequence_taxonomy()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a logical
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

miseq$has_sequence_taxonomy()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-is_equal"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-is_equal}{}}}
\subsection{\code{strollur$is_equal()}}{
  Determine if two
\href{https://mothur.org/strollur/reference/strollur.html}{strollur}
objects are equal.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$is_equal(data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data, }}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur}
object}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a logical
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

data <- copy_dataset(miseq)

miseq$is_equal(data)
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-names"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-names}{}}}
\subsection{\code{strollur$names()}}{
  Get the names of a given type of data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$names(
  type = "sequence",
  bin_type = "otu",
  samples = NULL,
  distinct = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{string containing the type of data you would like. Options
include: "dataset", "sequence", "bin", "sample", "treatment",
"report". Default = "sequence".}
      \item{\code{bin_type}}{string containing the bin type you would like the names
for. Default = "otu".}
      \item{\code{samples}}{vector of strings. samples is only used when 'type' =
"sequence" or 'type' = "bin" . samples should contain the names of
the samples you want names for. Default = NULL.}
      \item{\code{distinct}}{Boolean. distinct is used when 'type' = "sequence" or
'type' = "bin" and the samples parameter is used. The distinct
parameter allows you to get the names that present given set of
samples. When distinct is TRUE, the names function will return the
names that ONLY contain data from the given samples. When distinct is
FALSE the data returned contains data from a given samples, but may
ALSO contain data from other samples. Default = FALSE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    vector of strings, containing the names requested
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the name of the dataset
miseq$names(type = "dataset")

# To get the names of the sequences
miseq$names(type = "sequence")

# To get the names of the sequences present sample 'F3D0'
miseq$names(type = "sequence", samples = c("F3D0"))

#' # To get the names of the sequences unique to sample 'F3D0'
miseq$names(type = "sequence", samples = c("F3D0"), distinct = TRUE)

# To get the names of the samples
miseq$names(type = "sample")

# To get the names of the treatments
miseq$names(type = "treatment")

# To get the names of the bins
miseq$names(type = "bin")

# To get the names of the bins that are unique to 'F3D0'
miseq$names(type = "bin", samples = c("F3D0"), distinct = TRUE)

# To get the names of the bins that include sequences from 'F3D0'
miseq$names(type = "bin", samples = c("F3D0"), distinct = FALSE)

# To get the names of the reports
miseq$names(type = "report")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-report"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-report}{}}}
\subsection{\code{strollur$report()}}{
  Get a data.frame containing the given report, or the phylo tree
requested.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$report(type = "sequence", bin_type = "otu")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{string containing the type of report you would like. Options
include: "fasta", "sequence", "sequence_bin_assignment",
"sequence_taxonomy", "sequence_tree", "bin_taxonomy",
"bin_representative", "sample_assignment", "sample_tree",
"resource_reference", "sequence_scrap", "bin_scrap". If you have added
custom reports for alignment, contigs_assembly or chimeras, you can get
those as well. Default = "sequence".}
      \item{\code{bin_type}}{string containing the bin type you would like a
bin_taxonomy report for. Default = "otu".}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    data.frame or tree
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the FASTA data

miseq$report(type = "fasta") |> head(n = 5)

# To get a report about the FASTA data

miseq$report(type = "sequence") |> head(n = 5)

# To get the sequence bin assignments

miseq$report(type = "sequence_bin_assignment", bin_type = "otu") |>
head(n = 5)

# To get the sample treatment assignments

miseq$report(type = "sample_assignment") |> head(n = 5)

# To get a report about sequence classifications

miseq$report(type = "sequence_taxonomy") |> head(n = 5)

# To get a report about bin classifications for 'otu' data

miseq$report(type = "bin_taxonomy", bin_type = "otu") |> head(n = 5)

# To get the 'otu' bin representative sequences

miseq$report(type = "bin_representative", bin_type = "otu") |>
head(n = 5)

# To get a report about the sequences removed during your analysis:

miseq$report(type = "sequence_scrap")

# To get a report about the "otu" bins removed during your analysis:

miseq$report(type = "bin_scrap", bin_type = "otu")

# To get the metadata associated with your data:

metadata <- miseq$report(type = "metadata") |> head(n = 5)

# To get the resource references associated with your data:

references <- miseq$report(type = "resource_reference")

# To get our custom report containing the contigs assembly data:

miseq$report(type = "contigs_report") |> head(n = 5)

# To get the tree that relates your sequences:

sequence_tree <- miseq$report(type = "sequence_tree")

# To get the tree that relates your sequences:

sample_tree <- miseq$report(type = "sample_tree")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-summary"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-summary}{}}}
\subsection{\code{strollur$summary()}}{
  Summarize the sequences data, custom reports, and scrapped data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$summary(type = "sequence", report_type = NULL, verbose = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{string containing the type of data you want the number of.
Options include: "sequence", "report" and "scrap".
Default = "sequence".}
      \item{\code{report_type}}{string containing the report type you would
summarized. For example, the miseq_sop_example includes contigs assembly
data and can be accessed with report_type = "contigs_report".
Default = NULL.}
      \item{\code{verbose}}{boolean indicating whether or not you want progress
messages. Default = TRUE.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    data.frame
Get summary of the sequence reports
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{miseq <- load_dataset(strollur_example("miseq_sop.rds"))

# To get the summary of your FASTA data
miseq$summary(type = "sequence")

# summarize contigs_report
miseq$summary(type = "report", report_type = "contigs_report")

# remove sample 'F3D0' to produce a scrap report
xdev_remove_samples(data = miseq, samples = c("F3D0"))

# summarize scrapped data -
# sequences and bins scrapped by removing the sample "F3D0"
miseq$summary(type = "scrap")
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clone"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clone}{}}}
\subsection{\code{strollur$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{strollur$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
