% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_dataset.R, R/r6_strollur.R
\name{strollur}
\alias{strollur}
\title{strollur}
\description{
'strollur' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.

'strollur' is an R6 class that stores nucleotide sequences,
abundance, sample and treatment assignments, taxonomic classifications,
asv / otu clusters and various reports. It is designed to facilitate data
analysis across multiple R packages.
}
\examples{

## ------------------------------------------------
## Method `strollur$new`
## ------------------------------------------------


# to create an empty strollur object, run the following:

data <- new_dataset("soil")


## ------------------------------------------------
## Method `strollur$add`
## ------------------------------------------------


# Create a new empty `strollur` object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# Read FASTA data into data.frame
fasta_data <- read_fasta(fasta = strollur_example("final.fasta.gz"))

# Add FASTA sequence data
data$add(table = fasta_data, type = "sequences")

# To add FASTA data with a resource reference

# Create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# Create a resource reference for the FASTA data
resource_url <- "https://mothur.org/wiki/silva_reference_files/"
resource_reference <-
  new_reference(
    reference_name = "silva.bacteria.fasta",
    reference_version = "1.38.1",
    reference_usage = "alignment by mothur2 v1.0",
    reference_note = "default options",
    reference_url = resource_url
  )

# Add FASTA data with a resource reference

data$add(
  table = fasta_data,
  type = "sequences",
  reference = resource_reference
)

# Add contigs assembly report with a 'sequence_name' column named 'Name'

contigs_report <- readr::read_tsv(
  strollur_example(
    "final.contigs_report.gz"
  ),
  col_names = TRUE, show_col_types = FALSE
)

data$add(
  table = contigs_report, type = "reports",
  report_type = "contigs_report", list(sequence_name = "Name")
)

# To add metadata related to your study

metadata <- readr::read_tsv(strollur_example("mouse.dpw.metadata"),
  col_names = TRUE, show_col_types = FALSE
)

data$add(table = metadata, type = "metadata")


## ------------------------------------------------
## Method `strollur$add_sample_tree`
## ------------------------------------------------


 data <- strollur$new("my_dataset")

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 tree <- ape::read.tree(strollur_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)


## ------------------------------------------------
## Method `strollur$add_sequence_tree`
## ------------------------------------------------


 data <- strollur$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)


## ------------------------------------------------
## Method `strollur$assign`
## ------------------------------------------------


# Assign sequence classifications

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

sequence_classifications <- read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

data$assign(table = sequence_classifications, type = "sequence_taxonomy")

# Assigning bins

# read mothur's otu list file into data.frame
otu_data <- read_mothur_list(list = strollur_example(
  "final.opti_mcc.list.gz"
))

# read mothur's asv list file into data.frame
asv_data <- read_mothur_list(list = strollur_example(
  "final.asv.list.gz"
))

# read mothur's phylotype list file into data.frame
phylo_data <- read_mothur_list(list = strollur_example(
  "final.tx.list.gz"
))

# read otu bin representative sequences into a data.frame
bin_reps <- readr::read_tsv(
  strollur_example(
    "otu_representative_sequences.tsv"
  ),
  show_col_types = FALSE
)

# assign 'otu' bins using sequence names
data$assign(table = otu_data, bin_type = "otu")

# assign 'asv' bins using sequence names
data$assign(table = asv_data, bin_type = "asv")

# assign 'phylotype' bins using sequence names
data$assign(table = phylo_data, bin_type = "phylotype")

# assign 'otu' bin representative sequences
data$assign(table = bin_reps, type = "bin_representatives")

# To assign abundance only bins

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# read mothur's shared file
otu_data <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# assign abundance only otus parsed by sample
data$assign(table = otu_data, bin_type = "otu")

# Assigning bin classifications

# read bin taxonomies
otu_data <- read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

# assign otu consensus taxonomies
data$assign(
  table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

# Assign treatments

sample_assignments <- readr::read_table(
  strollur_example("mouse.time.design"),
  col_names = TRUE, show_col_types = FALSE
)

data$assign(table = sample_assignments, type = "treatments")


## ------------------------------------------------
## Method `strollur$get_bin_types`
## ------------------------------------------------


data <- miseq_sop_example()
data$get_bin_types()


## ------------------------------------------------
## Method `strollur$get_metadata`
## ------------------------------------------------

  data <- strollur$new("my_dataset")

  metadata <- readr::read_tsv(strollur_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add(data = data, table = metadata, type = "metadata")

  data$get_metadata()


## ------------------------------------------------
## Method `strollur$get_sample_tree`
## ------------------------------------------------


 tree <- ape::read.tree(strollur_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

 data <- strollur$new("my_dataset")

 # assign abundance 'otu' bins
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 data$add_sample_tree(tree)
 data$get_sample_tree()


## ------------------------------------------------
## Method `strollur$get_sequence_tree`
## ------------------------------------------------


 data <- strollur$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Rcpp::XPtr<Dataset> pointer to 'Dataset' c++ class. This
allows package developers an easy access point to the underlying C++ code
with additional functionality.}

\item{\code{raw}}{Rcpp::RawVector containing the serialized data of the
'Dataset' c++ class. This allows the load and save functions to work with
the class.}

\item{\code{sequence_tree}}{a tree that relates sequences to eachother}

\item{\code{sample_tree}}{a tree that relates samples to eachother}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-strollur-new}{\code{strollur$new()}}
\item \href{#method-strollur-print}{\code{strollur$print()}}
\item \href{#method-strollur-add}{\code{strollur$add()}}
\item \href{#method-strollur-add_sample_tree}{\code{strollur$add_sample_tree()}}
\item \href{#method-strollur-add_sequence_tree}{\code{strollur$add_sequence_tree()}}
\item \href{#method-strollur-assign}{\code{strollur$assign()}}
\item \href{#method-strollur-clear}{\code{strollur$clear()}}
\item \href{#method-strollur-get_bin_types}{\code{strollur$get_bin_types()}}
\item \href{#method-strollur-get_metadata}{\code{strollur$get_metadata()}}
\item \href{#method-strollur-get_sample_tree}{\code{strollur$get_sample_tree()}}
\item \href{#method-strollur-get_sequence_report}{\code{strollur$get_sequence_report()}}
\item \href{#method-strollur-get_summary}{\code{strollur$get_summary()}}
\item \href{#method-strollur-get_sequence_tree}{\code{strollur$get_sequence_tree()}}
\item \href{#method-strollur-clone}{\code{strollur$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-new"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-print"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add(
  table,
  type = "sequences",
  report_type = NULL,
  table_names = list(sequence_name = "sequence_names", sequence = "sequences", comment =
    "comments", reference_name = "reference_names", reference_version =
    "reference_versions", reference_usage = "reference_usages", reference_note =
    "reference_notes", reference_url = "reference_urls"),
  reference = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add_sample_tree}{}}}
\subsection{Method \code{add_sample_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add_sample_tree(tree)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add_sequence_tree}{}}}
\subsection{Method \code{add_sequence_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add_sequence_tree(tree)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-assign"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$assign(
  table,
  type = "bins",
  bin_type = "otu",
  table_names = list(sequence_name = "sequence_names", abundance = "abundances", sample =
    "samples", treatment = "treatments", taxonomy = "taxonomies", bin_name = "bin_names"),
  reference = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clear"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_bin_types"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_bin_types}{}}}
\subsection{Method \code{get_bin_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_bin_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sample_tree}{}}}
\subsection{Method \code{get_sample_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sample_tree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sequence_report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sequence_tree}{}}}
\subsection{Method \code{get_sequence_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sequence_tree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clone"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Rcpp::XPtr<Dataset> pointer to 'Dataset' c++ class. This
allows package developers an easy access point to the underlying C++ code
with additional functionality.}

\item{\code{raw}}{Rcpp::RawVector containing the serialized data of the
'Dataset' c++ class. This allows the load and save functions to work with
the class.}

\item{\code{sequence_tree}}{a tree that relates sequences to eachother}

\item{\code{sample_tree}}{a tree that relates samples to eachother}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-strollur-new}{\code{strollur$new()}}
\item \href{#method-strollur-print}{\code{strollur$print()}}
\item \href{#method-strollur-add}{\code{strollur$add()}}
\item \href{#method-strollur-add_sample_tree}{\code{strollur$add_sample_tree()}}
\item \href{#method-strollur-add_sequence_tree}{\code{strollur$add_sequence_tree()}}
\item \href{#method-strollur-assign}{\code{strollur$assign()}}
\item \href{#method-strollur-clear}{\code{strollur$clear()}}
\item \href{#method-strollur-get_bin_types}{\code{strollur$get_bin_types()}}
\item \href{#method-strollur-get_metadata}{\code{strollur$get_metadata()}}
\item \href{#method-strollur-get_sample_tree}{\code{strollur$get_sample_tree()}}
\item \href{#method-strollur-get_sequence_report}{\code{strollur$get_sequence_report()}}
\item \href{#method-strollur-get_summary}{\code{strollur$get_summary()}}
\item \href{#method-strollur-get_sequence_tree}{\code{strollur$get_sequence_tree()}}
\item \href{#method-strollur-clone}{\code{strollur$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-new"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-new}{}}}
\subsection{Method \code{new()}}{
Create a new strollur dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of dataset (optional)}

\item{\code{processors}}{Integer, number of cores to use.
Default = all available}

\item{\code{dataset}}{a `strollur` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `strollur` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# to create an empty strollur object, run the following:

data <- new_dataset("soil")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-print"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add}{}}}
\subsection{Method \code{add()}}{
Add sequences, reports, metadata or resource references
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add(
  table,
  type = "sequences",
  report_type = NULL,
  table_names = list(sequence_name = "sequence_names", sequence = "sequences", comment =
    "comments", reference_name = "reference_names", reference_version =
    "reference_versions", reference_usage = "reference_usages", reference_note =
    "reference_notes", reference_url = "reference_urls"),
  reference = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table, }}{a data.frame containing the data you wish to add.}

\item{\code{type, }}{a string containing the type of data. Options include:
'sequences', 'references' 'metadata' and 'reports'.}

\item{\code{report_type, }}{a string containing the type of report you are adding.
Options include: 'metadata' and custom reports.}

\item{\code{table_names, }}{named list used to indicate the names of the columns in
the table. By default:

table_names <- list(sequence_name = "sequence_names",
                    comment = "comments",
                    sequence = "sequences",
                    reference_name = "reference_names",
                    reference_version = "reference_versions",
                    reference_usage = "reference_usages",
                    reference_note = "reference_notes",
                    reference_url = "reference_urls")

In table_names, 'sequence_name' is a string containing the name of the column
in 'table' that contains the sequence names. It is used when you are adding
FASTA data. Default column name is 'sequence_names'.

In table_names, 'sequence' is a string containing the name of the column in
'table' that contains the sequence nucleotide strings. It is used when you
are adding FASTA data. Default column name is 'sequences'.

In table_names, 'comment' is a string containing the name of the column in
'table' that contains the sequence comments. It is used when you are adding
FASTA data. Default column name is 'comments'.

In table_names, 'reference_name' is a string containing the name of the
column in 'table' that contains the reference names. It is used when you are
adding reference data. Default column name is 'reference_names'.

In table_names, 'reference_version' is a string containing the name of the
column in 'table' that contains the reference versions. Default column name
is 'reference_versions'.

In table_names, 'reference_usage' is a string containing the name of the
column in 'table' that contains the reference usages. Default column name is
'reference_usages'.

In table_names, 'reference_note' is a string containing the name of the
column in 'table' that contains the reference notes. Default column name is
 'reference_notes'.

In table_names, 'reference_url' is a string containing the name of the column
in 'table' that contains the reference urls. Default column name is
 'reference_urls'.}

\item{\code{reference, }}{a list created by the function [new_reference]. Optional.}

\item{\code{verbose, }}{boolean indicating whether or not you want progress messages.
Default = TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
double - The number of items added
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# Create a new empty `strollur` object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# Read FASTA data into data.frame
fasta_data <- read_fasta(fasta = strollur_example("final.fasta.gz"))

# Add FASTA sequence data
data$add(table = fasta_data, type = "sequences")

# To add FASTA data with a resource reference

# Create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# Create a resource reference for the FASTA data
resource_url <- "https://mothur.org/wiki/silva_reference_files/"
resource_reference <-
  new_reference(
    reference_name = "silva.bacteria.fasta",
    reference_version = "1.38.1",
    reference_usage = "alignment by mothur2 v1.0",
    reference_note = "default options",
    reference_url = resource_url
  )

# Add FASTA data with a resource reference

data$add(
  table = fasta_data,
  type = "sequences",
  reference = resource_reference
)

# Add contigs assembly report with a 'sequence_name' column named 'Name'

contigs_report <- readr::read_tsv(
  strollur_example(
    "final.contigs_report.gz"
  ),
  col_names = TRUE, show_col_types = FALSE
)

data$add(
  table = contigs_report, type = "reports",
  report_type = "contigs_report", list(sequence_name = "Name")
)

# To add metadata related to your study

metadata <- readr::read_tsv(strollur_example("mouse.dpw.metadata"),
  col_names = TRUE, show_col_types = FALSE
)

data$add(table = metadata, type = "metadata")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add_sample_tree}{}}}
\subsection{Method \code{add_sample_tree()}}{
Add phylo tree relating the samples in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add_sample_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- strollur$new("my_dataset")

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 tree <- ape::read.tree(strollur_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-add_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-add_sequence_tree}{}}}
\subsection{Method \code{add_sequence_tree()}}{
Add phylo tree relating the sequences in your dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$add_sequence_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{a phylo tree object created by ape::read.tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- strollur$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-assign"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-assign}{}}}
\subsection{Method \code{assign()}}{
Assign sequence abundances, sequence classifications, bins, bin
representative sequences, bin classifications or treatments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$assign(
  table,
  type = "bins",
  bin_type = "otu",
  table_names = list(sequence_name = "sequence_names", abundance = "abundances", sample =
    "samples", treatment = "treatments", taxonomy = "taxonomies", bin_name = "bin_names"),
  reference = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table, }}{a data.frame containing the data you wish to assign}

\item{\code{type, }}{a string containing the type of data. Options include:
'sequence_abundance', 'sequence_taxonomy', 'bins',
 'bin_representatives', 'bin_taxonomy' and 'treatments'.
 Default = "bins".}

\item{\code{bin_type, }}{string containing the bin type you would like the number of
bins for. Default = "otu".}

\item{\code{table_names, }}{named list used to indicate the names of the columns in
the table. By default:

table_names <- list(sequence_name = "sequence_names",
                    abundance = "abundances",
                    sample = "samples",
                    treatment = "treatments",
                    taxonomy = "taxonomies",
                    bin_name = "bin_names")

In table_names, 'sequence_name' is a string containing the name of the column
in 'table' that contains the sequence names. Default column name is
'sequence_names'.

In table_names, 'abundance' is a string containing the name of the column in
'table' that contains the abundances. Default column name is 'abundances'.

In table_names, 'sample' is a string containing the name of the column in
'table' that contains the samples. Default column name is 'samples'.

In table_names, 'treatment' is a string containing the name of the
column in 'table' that contains the treatment names. Default column name is
 'treatments'.

In table_names, 'taxonomy' is a string containing the name of the
column in 'table' that contains the classifications. Default column name
is 'taxonomies'.

In table_names, 'bin_name' is a string containing the name of the
column in 'table' that contains the bin names. Default column name is
'bin_names'.}

\item{\code{reference, }}{a list created by the function [new_reference]. Optional.}

\item{\code{verbose, }}{boolean indicating whether or not you want progress messages.
Default = TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
double - The number of items assigned
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
# Assign sequence classifications

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

sequence_classifications <- read_mothur_taxonomy(strollur_example(
  "final.taxonomy.gz"
))

data$assign(table = sequence_classifications, type = "sequence_taxonomy")

# Assigning bins

# read mothur's otu list file into data.frame
otu_data <- read_mothur_list(list = strollur_example(
  "final.opti_mcc.list.gz"
))

# read mothur's asv list file into data.frame
asv_data <- read_mothur_list(list = strollur_example(
  "final.asv.list.gz"
))

# read mothur's phylotype list file into data.frame
phylo_data <- read_mothur_list(list = strollur_example(
  "final.tx.list.gz"
))

# read otu bin representative sequences into a data.frame
bin_reps <- readr::read_tsv(
  strollur_example(
    "otu_representative_sequences.tsv"
  ),
  show_col_types = FALSE
)

# assign 'otu' bins using sequence names
data$assign(table = otu_data, bin_type = "otu")

# assign 'asv' bins using sequence names
data$assign(table = asv_data, bin_type = "asv")

# assign 'phylotype' bins using sequence names
data$assign(table = phylo_data, bin_type = "phylotype")

# assign 'otu' bin representative sequences
data$assign(table = bin_reps, type = "bin_representatives")

# To assign abundance only bins

# create a new empty strollur object named 'example_dataset'
data <- new_dataset(dataset_name = "example_dataset")

# read mothur's shared file
otu_data <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

# assign abundance only otus parsed by sample
data$assign(table = otu_data, bin_type = "otu")

# Assigning bin classifications

# read bin taxonomies
otu_data <- read_mothur_cons_taxonomy(strollur_example(
  "final.cons.taxonomy"
))

# assign otu consensus taxonomies
data$assign(
  table = otu_data,
  type = "bin_taxonomy", bin_type = "otu"
)

# Assign treatments

sample_assignments <- readr::read_table(
  strollur_example("mouse.time.design"),
  col_names = TRUE, show_col_types = FALSE
)

data$assign(table = sample_assignments, type = "treatments")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clear"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clear}{}}}
\subsection{Method \code{clear()}}{
Clear data from datasest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_bin_types"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_bin_types}{}}}
\subsection{Method \code{get_bin_types()}}{
Get bin table types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_bin_types()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of strings
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
data <- miseq_sop_example()
data$get_bin_types()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get data.frame containing metadata for the strollur object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame()
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  data <- strollur$new("my_dataset")

  metadata <- readr::read_tsv(strollur_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add(data = data, table = metadata, type = "metadata")

  data$get_metadata()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sample_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sample_tree}{}}}
\subsection{Method \code{get_sample_tree()}}{
Get phylo tree relating the samples in your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sample_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 tree <- ape::read.tree(strollur_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(strollur_example("final.opti_mcc.shared"))

 data <- strollur$new("my_dataset")

 # assign abundance 'otu' bins
 assign(data = data, table = df, type = "bins", bin_type = "otu")

 data$add_sample_tree(tree)
 data$get_sample_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sequence_report"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sequence_report}{}}}
\subsection{Method \code{get_sequence_report()}}{
Get data.frame sequence report data. Sequence report data includes: start
positions, end positions, number of bases, number of ambiguous bases,
length of longest homopolymer, and the number of N's.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sequence_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
Get summary of the sequence reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-get_sequence_tree"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-get_sequence_tree}{}}}
\subsection{Method \code{get_sequence_tree()}}{
Get phylo tree relating the sequences in your strollur object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$get_sequence_tree()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 data <- strollur$new("my_dataset")
 tree <- ape::read.tree(strollur_example("final.phylip.tre.gz"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strollur-clone"></a>}}
\if{latex}{\out{\hypertarget{method-strollur-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{strollur$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
