% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strollur_example.R
\name{strollur_example}
\alias{strollur_example}
\title{strollur_example}
\usage{
strollur_example(file = NULL)
}
\arguments{
\item{file}{Name of file.}
}
\description{
strollur comes bundled with some example files in its `inst/extdata`
directory. This function make them easy to access them.
}
\details{
Get path to strollur example
}
\examples{
strollur_example()
strollur_example("final.fasta.gz")
}
