% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_qiime2_artifact.R
\name{unpack_qiime2_artifact}
\alias{unpack_qiime2_artifact}
\title{unpack_qiime2_artifact}
\usage{
unpack_qiime2_artifact(qza, dir_path = NULL)
}
\arguments{
\item{qza}{filename, a .qza file containing artifact}

\item{dir_path}{a string containing the name of directory where the artifacts
files should be written. Default = current working directory.}
}
\value{
A unpacked qza artifact
}
\description{
The unpack_qiime2_artifact function reads .qza files created by
\href{https://qiime2.org}{qiime2}, and returns the artifact.

To generate the various input files you can follow \href{https://amplicon-docs.qiime2.org/en/latest/tutorials/moving-pictures.html}{qiime2 moving-pictures}.
}
\examples{

# Using the example files from moving-pictures

artifact <- unpack_qiime2_artifact(
  qza = strollur_example("table.qza"),
)

}
