% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biom.R
\name{write_biom}
\alias{write_biom}
\title{Write a \href{https://biom-format.org}{BIOM} formatted file containing
your \href{https://mothur.org/strollur/reference/strollur.html}{strollur}
objects data.}
\usage{
write_biom(
  data,
  file_root = NULL,
  path = NULL,
  taxonomy = TRUE,
  sequence = TRUE
)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{file_root}{a string containing the root name of the output file.
Default = \code{{dataset_name}}.\code{{bin_type}}.biom}

\item{path}{string containing the name of directory where the files
should be written. Default = current working directory.}

\item{taxonomy}{logical, when \code{TRUE} if consensus taxonomies are available
include them in the biom file. Default = \code{TRUE}.}

\item{sequence}{logical, when \code{TRUE} if bins have a representative sequence
or all bins contain a single feature sequence include the sequences in the
biom file. Default = \code{TRUE}.}
}
\value{
vector of BIOM file names. One for each bin type.
}
\description{
Write a \href{https://biom-format.org}{BIOM} formatted file containing your
strollur::strollur objects data.
}
\examples{

miseq <- strollur::miseq_sop_example()
strollur::write_biom(miseq, tempfile())

}
