% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(data, filename = NULL, degap = FALSE)
}
\arguments{
\item{data}{A \code{strollur} object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.fasta}

\item{degap}{a logical. Default = FALSE. When degap = \code{TRUE}, all gap
characters will be removed from the sequences.}
}
\value{
name of FASTA file
}
\description{
Write a \href{https://www.ncbi.nlm.nih.gov/genbank/fastaformat/}{FASTA}
formatted sequence file
}
\examples{

miseq <- miseq_sop_example()
write_fasta(miseq, tempfile())

}
