% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fastq.R
\name{write_fastq}
\alias{write_fastq}
\title{Write a
\href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2847217/}{FASTQ} formatted
file}
\usage{
write_fastq(data, filename = NULL)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.fastq}
}
\value{
name of FASTQ file
}
\description{
Write a \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2847217/}{FASTQ}
formatted file
}
\examples{

table <- strollur::read_fastq(strollur_example("tiny.fastq.gz"))

# Create `strollur::strollur` object
data <- strollur::new_dataset("example")

# Add FASTQ data `strollur::strollur` object
strollur::add(data, table, type = "fastq")

# Write `strollur::strollur` objects FASTQ data to file
strollur::write_fastq(data, tempfile())

}
