% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur.R
\name{write_mothur}
\alias{write_mothur}
\title{write_mothur}
\usage{
write_mothur(data, dir_path = NULL, compress = TRUE, tags = NULL)
}
\arguments{
\item{data}{A `strollur` object}

\item{dir_path}{a string containing the name of directory where the files
should be written. Default = current working directory.}

\item{compress}{boolean, Default = TRUE.}

\item{tags}{a vector of strings containing the items you wish to write
Options are 'sequence_data', 'bin_data', 'metadata',
'resource_reference', 'sequence_tree', 'sample_tree' and 'report'.
 By default, everything is written to files.}
}
\value{
a vector of file names
}
\description{
The write_mothur function will write various
\href{https://mothur.org/wiki/tags/#file_types}{file types} for use with
mothur.
}
\examples{

miseq <- miseq_sop_example()
files <- write_mothur(miseq, tempdir(), compress = FALSE)

}
\references{
Schloss,P.D., Westcott,S.L., Ryabin,T., Hall,J.R., Hartmann,M.,
 Hollister,E.B., Lesniewski,R.A., Oakley,B.B., Parks,D.H., Robinson,C.J.,
 Sahl,J.W., Stres,B., Thallinger,G.G., Van Horn,D.J. and Weber,C.F. (2009),
 Introducing mothur: Open-source, platform-independent, community-supported
 software for describing and comparing microbial communities. Applied and
 Environmental Microbiology 75:7537-7541. <doi:10.1128/AEM.01541-09>
}
