% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur.R
\name{write_mothur}
\alias{write_mothur}
\title{write_mothur}
\usage{
write_mothur(data, dir_path = NULL, compress = TRUE, tags = NULL)
}
\arguments{
\item{data}{A 'dataset' object}

\item{dir_path}{a string containing the name of directory where the files
should be written. Default = current working directory.}

\item{compress}{boolean, Default = TRUE.}

\item{tags}{a vector of strings containing the items you wish to write
Options are 'sequence_data', 'bin_data', 'metadata',
'references', 'sequence_tree', 'sample_tree', 'alignment_report',
'contigs_assembly_report' and 'chimera_report'. By default, everything is
 written to files.}
}
\value{
a vector of file names
}
\description{
The write_mothur function will write various
\href{https://mothur.org/wiki/tags/#file_types}{file types} for use with
mothur.
}
\examples{

miseq <- miseq_sop_example()
files <- write_mothur(miseq)

}
