% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_cons_taxonomy.R
\name{write_mothur_cons_taxonomy}
\alias{write_mothur_cons_taxonomy}
\title{write_mothur_cons_taxonomy}
\usage{
write_mothur_cons_taxonomy(data, file_root = NULL)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{file_root}{a string containing the root name of the output file.
Default = 'dataset_name'. Resulting in output files
'dataset_name'.bin_type'.cons.taxonomy.}
}
\value{
vector containing the names of the files created
}
\description{
Write a mothur formatted
\href{https://mothur.org/wiki/constaxonomy_file/}{cons_taxonomy file}
}
\examples{

miseq <- strollur::miseq_sop_example()
strollur::write_mothur_cons_taxonomy(miseq, tempfile())

}
