% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_cons_taxonomy.R
\name{write_mothur_cons_taxonomy}
\alias{write_mothur_cons_taxonomy}
\title{write_mothur_cons_taxonomy}
\usage{
write_mothur_cons_taxonomy(data, filename = NULL)
}
\arguments{
\item{data}{A 'dataset' object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.'bin_type'.cons.taxonomy}
}
\value{
vector containing the names of the files created
}
\description{
Write a mothur formatted
\href{https://mothur.org/wiki/constaxonomy_file/}{cons_taxonomy file}
}
\examples{

miseq <- miseq_sop_example()
write_mothur_cons_taxonomy(miseq)

}
