% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_design.R
\name{write_mothur_design}
\alias{write_mothur_design}
\title{write_mothur_design}
\usage{
write_mothur_design(data, filename = NULL)
}
\arguments{
\item{data}{A 'dataset' object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.design}
}
\value{
name of design file
}
\description{
Write a mothur formatted
\href{https://mothur.org/wiki/design_file/}{design file}
}
\examples{

miseq <- miseq_sop_example()
write_mothur_design(miseq)

}
