% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_list.R
\name{write_mothur_list}
\alias{write_mothur_list}
\title{write_mothur_list}
\usage{
write_mothur_list(data, file_root = NULL)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{file_root}{a string containing the root name of the output file.
Default = 'dataset_name'. Resulting in output files
\verb{\{dataset_name\}.\{bin_type\}.list}.}
}
\value{
vector containing the names of the files created
}
\description{
Write mothur formatted \href{https://mothur.org/wiki/list_file/}{list files}
}
\examples{

miseq <- strollur::miseq_sop_example()
strollur::write_mothur_list(miseq, tempfile())

}
