% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_rabund.R
\name{write_mothur_rabund}
\alias{write_mothur_rabund}
\title{write_mothur_rabund}
\usage{
write_mothur_rabund(data, file_root = NULL)
}
\arguments{
\item{data}{A \code{strollur} object}

\item{file_root}{a string containing the root name of the output file.
Default = 'dataset_name'. Resulting in output files
'dataset_name'.bin_type'.rabund.}
}
\value{
vector containing the names of the files created
}
\description{
Write mothur formatted
\href{https://mothur.org/wiki/rabund_file/}{rabund files}
}
\examples{

miseq <- miseq_sop_example()
write_mothur_rabund(miseq, tempfile())

}
