% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mothur_shared.R
\name{write_mothur_shared}
\alias{write_mothur_shared}
\title{write_mothur_shared}
\usage{
write_mothur_shared(data, file_root = NULL)
}
\arguments{
\item{data}{A 'dataset' object}

\item{file_root}{a string containing the root name of the output file.
Default = 'dataset_name'. Resulting in output files
'dataset_name'.bin_type'.shared.}
}
\value{
vector containing the names of the files created
}
\description{
Write mothur formatted
\href{https://mothur.org/wiki/shared_file/}{shared files}
}
\examples{

miseq <- miseq_sop_example()
write_mothur_shared(miseq, tempfile())

}
