% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{write_phyloseq}
\usage{
write_phyloseq(data)
}
\arguments{
\item{data}{the strollur object you created using one of the many
read functions in this package.}
}
\value{
returns a "phyloseq" object.
}
\description{
The \code{write_phyloseq()} function will take any strollur object and
return it as a "phyloseq" object.
}
\examples{
miseq <- miseq_sop_example()
if (requireNamespace("phyloseq", quietly = TRUE)) {
  phylo_obj <- write_phyloseq(miseq)
} else {
  message(paste(
    "To use this functionality you have to install the",
    "phyloseq package."
  ))
}

}
\references{
McMurdie,P.J. and Holmes,S. (2013), phyloseq: An R Package for
Reproducible Interactive Analysis and Graphics of Microbiome Census Data.
PLoS ONE 8:e61217. \url{doi:10.1371/journal.pone.0061217}
}
