% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{write_phyloseq}
\usage{
write_phyloseq(data)
}
\arguments{
\item{data}{the strollur object you created using one of the many 
read functions in this package.}
}
\value{
returns a "phyloseq" object.¬
}
\description{
The `write_phyloseq()` function will take any strollur object and
return it as a "phyloseq" object.
}
\examples{
 miseq <- miseq_sop_example()
 phylo_obj <- write_phyloseq(miseq)

}
