% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quality.R
\name{write_quality}
\alias{write_quality}
\title{write_quality}
\usage{
write_quality(data, filename = NULL)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.qual}
}
\value{
name of sequence quality file
}
\description{
Write a file containing sequence quality scores
}
\examples{

table <- strollur::read_fastq(strollur_example("tiny.fastq.gz"))

# Create `strollur::strollur` object
data <- strollur::new_dataset("example")

# Add FASTQ data `strollur::strollur` object
strollur::add(data, table, type = "fastq")

# Write `strollur::strollur` objects sequence quality data to file
strollur::write_quality(data, tempfile())

}
