% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxonomy.R
\name{write_taxonomy}
\alias{write_taxonomy}
\title{Write a 2 column
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}}
\usage{
write_taxonomy(data, filename = NULL)
}
\arguments{
\item{data}{a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{filename}{a string containing the name of the output file. Default =
'dataset_name'.taxonomy}
}
\value{
name of taxonomy file
}
\description{
Write a 2 column
\href{https://mothur.org/wiki/taxonomy_file/}{taxonomy file}
}
\examples{

miseq <- strollur::miseq_sop_example()
strollur::write_taxonomy(miseq, tempfile())

}
