% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_abundance}
\alias{xdev_abundance}
\title{xdev_abundance}
\usage{
xdev_abundance(data, type = "sequences", bin_type = "otu", by_sample = FALSE)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you want the number of.
Options include: "sequences", "bins".
Default = "sequences".}

\item{bin_type, }{string containing the bin type you would like the number of
bins for. Default = "otu".}

\item{by_sample, }{Boolean. When by_sample is TRUE, the abundance data will
be parsed by sample. Default = FALSE.}
}
\value{
data.frame
}
\description{
Get a table containing the requested abundance data in a \link{dataset}
object
}
\examples{

miseq <- miseq_sop_example()

# To the total abundance for each sequence
xdev_abundance(data = miseq, type = "sequences")

# To the total abundance for each sequence parsed by sample
xdev_abundance(data = miseq, type = "sequences", by_sample = TRUE)

# To the total abundance for each "otu" bin
xdev_abundance(data = miseq, type = "bins", bin_type = "otu")

# To the total abundance for each "otu" bin parsed by sample
xdev_abundance(data = miseq, type = "bins", bin_type = "otu", by_sample = TRUE)

# To the total abundance for each "asv" bin
xdev_abundance(data = miseq, type = "bins", bin_type = "asv")

# To the total abundance for each "asv" bin parsed by sample
xdev_abundance(data = miseq, type = "bins", bin_type = "asv", by_sample = TRUE)

# To the total abundance of each sample
xdev_abundance(data = miseq, type = "samples")

# To the total abundance of each treatment
xdev_abundance(data = miseq, type = "treatments")

}
