% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_add_references}
\alias{xdev_add_references}
\title{xdev_add_references}
\usage{
xdev_add_references(
  data,
  table,
  reference_name = "reference_names",
  reference_version = "reference_versions",
  reference_usage = "reference_usages",
  reference_note = "reference_notes",
  reference_url = "reference_urls",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing reference_names, reference_versions
(optional), reference_usages (optional), reference_notes (optional), and
reference_urls (optional).}

\item{reference_name, }{a string containing the name of the column in 'table'
that contains the reference names. Default column name is 'reference_names'.}

\item{reference_version, }{a string containing the name of the column in
'table' that contains the reference versions. Default column name is
'reference_versions'.}

\item{reference_usage, }{a string containing the name of the column in
'table' that contains the reference usages. Default column name is
 reference_usages'.}

\item{reference_note, }{a string containing the name of the column in
'table' that contains the reference notes. Default column name is
 reference_notes'.}

\item{reference_url, }{a string containing the name of the column in
'table' that contains the reference urls. Default column name is
 reference_urls'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
double containing the number of references added
}
\description{
Add resource references to a \link{dataset} object
}
\examples{

data <- new_dataset("just for fun", 2)
reference_table <- readr::read_csv(strollur_example("references.csv"),
                             col_names = TRUE, show_col_types = FALSE)
xdev_add_references(data, reference_table)

}
