% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_add_sequences}
\alias{xdev_add_sequences}
\title{Add sequence data to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_add_sequences(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_name",
  sequence = "sequence",
  comment = "comment",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing names, sequences(optional) and
comments(optional).}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.}

\item{sequence}{a string containing the name of the column in 'table' that
contains the sequence nucleotide strings. Default column name is
'sequence'.}

\item{comment}{a string containing the name of the column in
'table' that contains the sequence comments. Default column name is
'comment'.}

\item{verbose}{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Add sequence data to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

 data <- strollur::new_dataset("miseq_sop")
 fasta_data <- strollur::read_fasta(strollur_example("final.fasta.gz"))
 strollur::xdev_add_sequences(data, fasta_data)

# With the additional parameters to add information about the reference

 data <- strollur::new_dataset("miseq_sop")
 fasta_data <- strollur::read_fasta(strollur_example("final.fasta.gz"))

 strollur::xdev_add_sequences(data, fasta_data,
               new_reference("silva.bacteria.fasta",
               "1.38.1",
               "alignment by mothur2 v1.0 using default options",
               "https://mothur.org/wiki/silva_reference_files/"))

# You can also add references using the `strollur::xdev_add_references`
# function.

}
