% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_add_sequences}
\alias{xdev_add_sequences}
\title{xdev_add_sequences}
\usage{
xdev_add_sequences(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_names",
  sequence = "sequences",
  comment = "comments",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing names, sequences(optional) and
comments(optional).}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}

\item{sequence, }{a string containing the name of the column in 'table' that
contains the sequence nucleotide strings. Default column name is
'sequences'.}

\item{comment, }{a string containing the name of the column in
'table' that contains the sequence comments. Default column name is
'comments'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
double containing the number of sequences added
}
\description{
Add sequence data to a \link{dataset} object
}
\examples{

 data <- new_dataset("miseq_sop", 2)
 fasta_data <- read_fasta(strollur_example("final.fasta.gz"))
 xdev_add_sequences(data, fasta_data)

# With the additional parameters to add information about the reference

 data <- new_dataset("miseq_sop", 2)
 fasta_data <- read_fasta(strollur_example("final.fasta.gz"))

 xdev_add_sequences(data, fasta_data,
               new_reference("silva.bacteria.fasta",
               "1.38.1",
               "alignment by mothur2 v1.0 using default options",
               "https://mothur.org/wiki/silva_reference_files/"))

# You can also add references using the 'add_references' function.

}
