% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_bin_taxonomy}
\alias{xdev_assign_bin_taxonomy}
\title{xdev_assign_bin_taxonomy}
\usage{
xdev_assign_bin_taxonomy(
  data,
  table,
  bin_type = "otu",
  reference = NULL,
  bin_name = "bin_name",
  taxonomy = "taxonomy",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table, }{a data.frame containing bin taxonomy assignments}

\item{bin_type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference, }{a list created by the function \link{new_reference}. Optional.}

\item{bin_name, }{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'bin_name'.}

\item{taxonomy, }{a string containing the name of the column in 'table' that
contains the bin taxonomies. Default column name is 'taxonomy'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Assign bin classifications to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
the concensus taxonomy for each bin for you.
}
\examples{

otu_data <- read_mothur_cons_taxonomy(strollur_example(
                        "final.cons.taxonomy"))

data <- new_dataset(dataset_name = "my_dataset")

# assign otu abundances
xdev_assign_bins(data = data, table = otu_data, bin_type = "otu")

# assign otu classifications
xdev_assign_bin_taxonomy(data = data, table = otu_data,
                         bin_type = "otu")

}
