% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_bins}
\alias{xdev_assign_bins}
\title{Add bin assignments to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_assign_bins(
  data,
  table,
  bin_type = "otu",
  reference = NULL,
  bin_name = "bin_name",
  abundance = "abundance",
  sample = "sample",
  sequence_name = "sequence_name",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{table}{a data.frame containing bin_data assignments}

\item{bin_type}{a string indicating the type of bin assignments. Default "otu".}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{bin_name}{a string containing the name of the column in 'table' that
contains the bin names. Default column name is 'bin_name'.}

\item{abundance}{a string containing the name of the column in 'table'
that contains the bin abundances. Default column name is 'abundance'. Note:
You must provide either abundance or sequence_name in the table.}

\item{sample}{a string containing the name of the column in 'table' that
contains the sample names for datasets where the abundances are broken down
by sample. Default column name is 'sample'.}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.
Note: You must provide either abundance or sequence_name in the table.}

\item{verbose}{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
Add bin assignments to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

  # To assign sequences to bins:

  data <- strollur::new_dataset(dataset_name = "miseq_sop")
  otu_data <- strollur::read_mothur_list(
                          list = strollur_example("final.opti_mcc.list.gz"))

  strollur::xdev_assign_bins(data, table = otu_data, bin_type = "otu")

  # To add abundance only bin assignments:

  data <- strollur::new_dataset(dataset_name = "miseq_sop")
  otu_data <- strollur::read_mothur_rabund(
                         rabund = strollur_example("final.opti_mcc.rabund"))

  strollur::xdev_assign_bins(data, table = otu_data, bin_type = "otu")

  # To add abundance bin assignments parsed by sample:

  data <- strollur::new_dataset(dataset_name = "miseq_sop")
  otu_data <- readRDS(strollur_example("miseq_shared_otu.rds"))

  strollur::xdev_assign_bins(data, table = otu_data, bin_type = "otu")

}
