% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sample_distances}
\alias{xdev_assign_sample_distances}
\title{Assign samples distances in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_assign_sample_distances(data, table, reference = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a 3 column data.frame (sample1 sample2 distance) containing distances between your samples}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{verbose}{a boolean indicating whether or not you want progress
messages. Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Assign samples distances in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

shared_file <- strollur_example("final.opti_mcc.shared")
dist_file <- strollur_example("final.opti_mcc.jclass.0.03.column.dist")
reference <- strollur::new_reference(name = "jclass estimator distances",
                                     vendor = "mothur_v1.48.6")

data <- strollur::new_dataset("my_dataset")
df <- read_mothur_shared(shared_file)
xdev_assign_bins(data, table = df, bin_type = "otu")

sample_dists <- readr::read_table(dist_file,
                                  col_names = FALSE,
                                  show_col_types = FALSE)
xdev_assign_sample_distances(data, table = sample_dists,
                             reference = reference)
data

}
