% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_abundance}
\alias{xdev_assign_sequence_abundance}
\title{Assign sequence abundance and optionally assign sample and treatment data to
a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_assign_sequence_abundance(
  data,
  table,
  sequence_name = "sequence_name",
  abundance = "abundance",
  sample = "sample",
  treatment = "treatment",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{table}{a data.frame containing sequence abundance assignments}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.}

\item{abundance}{a string containing the name of the column in 'table'
that contains the sequence abundances. Default column name is 'abundance'.}

\item{sample}{a string containing the name of the column in 'table'
that contains the sequence samples. Default column name is 'sample'.
(Optional)}

\item{treatment}{a string containing the name of the column in 'table'
that contains the sequence treatments. Default column name is 'treatment'.}

\item{verbose}{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
Assign sequence abundance and optionally assign sample and treatment data to
a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

data <- strollur::new_dataset("my_dataset")
abundance <- readRDS(strollur_example("miseq_abundance_by_sample.rds"))

strollur::xdev_assign_sequence_abundance(data,
                                           table = abundance)

}
