% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_abundance}
\alias{xdev_assign_sequence_abundance}
\title{xdev_assign_sequence_abundance}
\usage{
xdev_assign_sequence_abundance(
  data,
  table,
  sequence_name = "sequence_names",
  abundance = "abundances",
  sample = "samples",
  treatment = "treatments",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing sequence abundance assignments}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}

\item{abundance, }{a string containing the name of the column in 'table'
that contains the sequence abundances. Default column name is 'abundances'.}

\item{sample, }{a string containing the name of the column in 'table'
that contains the sequence samples. Default column name is 'samples'.
(Optional)}

\item{treatment, }{a string containing the name of the column in 'table'
that contains the sequence treatments. Default column name is 'treatments'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
double containing the number of sequences assigned
}
\description{
Assign sequence abundance and optionally assign sample and treatment data to
 a \link{dataset} object
}
\examples{

data <- new_dataset("my_dataset")
sequence_abundance <- readr::read_tsv(rdataset_example(
                                      "mothur2_count_table.tsv"),
                                      show_col_types = FALSE)

xdev_assign_sequence_abundance(data = data, table = sequence_abundance,
                               sequence_name = "names")

}
