% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_quality_scores}
\alias{xdev_assign_sequence_quality_scores}
\title{xdev_assign_sequence_quality_scores}
\usage{
xdev_assign_sequence_quality_scores(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_name",
  quality_score = "quality_score",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing quality data you are trying to add}

\item{reference}{a list created by the function \link{new_reference}. Optional.}

\item{sequence_name}{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_name'.}

\item{quality_score}{a string containing the name of the column in 'table'
that contains the quality scores stored as vector\if{html}{\out{<int>}}. Default column
name is 'quality_score'.}

\item{verbose}{a logical whether or not you want progress messages.
Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Add quality data to a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.
Scores are stored as vector\if{html}{\out{<int>}} for each sequence.
}
\examples{

qual_table <- strollur::read_quality(strollur_example("tiny.qual"))
fasta_table <- strollur::read_fasta(strollur_example("tiny.fasta"))

data <- strollur::new_dataset("example")
strollur::xdev_add_sequences(data, fasta_table)
strollur::xdev_assign_sequence_quality_scores(data, qual_table)

data

}
