% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_taxonomy}
\alias{xdev_assign_sequence_taxonomy}
\title{xdev_assign_sequence_taxonomy}
\usage{
xdev_assign_sequence_taxonomy(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_names",
  taxonomy = "taxonomies",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{table, }{a data.frame containing sequence taxonomy assignments}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}

\item{taxonomy, }{a string containing the name of the column in 'table' that
contains the sequence taxonomies. Default column name is 'taxonomies'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
double containing the number of sequence assigned
}
\description{
Assign sequence classifications to a \link{dataset} object

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the consensus taxonomy for each bin for you.
}
\examples{

sequence_classifications <- read_mothur_taxonomy(rdataset_example(
                        "final.taxonomy.gz"))

data <- new_dataset("my_dataset", 2)

xdev_assign_sequence_taxonomy(data, sequence_classifications)

# With the reference parameter you can add information about the reference
# you used to classify your sequences. You can also add references using the
# 'add_references' function.

reference <- new_reference("trainset9_032012.pds.zip", "9_032012",
              "classification by mothur2 v1.0 using default options", "",
"https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset9_032012.pds.zip")

xdev_assign_sequence_taxonomy(data, sequence_classifications, reference)

}
