% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_sequence_taxonomy_tidy}
\alias{xdev_assign_sequence_taxonomy_tidy}
\title{xdev_assign_sequence_taxonomy_tidy}
\usage{
xdev_assign_sequence_taxonomy_tidy(
  data,
  table,
  reference = NULL,
  sequence_name = "sequence_names",
  level = "levels",
  taxonomy = "taxonomies",
  confidence = "confidences",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table, }{a data.frame containing sequence taxonomy assignments}

\item{reference, }{a list created by the function [new_reference]. Optional.}

\item{sequence_name, }{a string containing the name of the column in 'table'
that contains the sequence names. Default column name is 'sequence_names'.}

\item{level, }{a string containing the name of the column in 'table' that
contains the taxonomy levels. Default column name is 'levels'.}

\item{taxonomy, }{a string containing the name of the column in 'table' that
contains the sequence taxonomies. Default column name is 'taxonomies'.}

\item{confidence, }{a string containing the name of the column in 'table'
that contains the taxonomies confidence. Default column name is 'confidences'.}

\item{verbose, }{a boolean whether or not you want progress messages.
Default = TRUE.}
}
\value{
double containing the number of sequence taxonomies assigned
}
\description{
Assign sequence classifications to a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object

Note, if you assign sequence taxonomies and assign bins, 'Dataset' will find
 the consensus taxonomy for each bin for you.
}
\examples{

sequence_classifications <- readRDS(strollur_example("miseq_tidy_taxonomy.rds"))

data <- new_dataset("my_dataset")

xdev_assign_sequence_taxonomy_tidy(data, sequence_classifications)

# With the reference parameter you can add information about the reference
# you used to classify your sequences. You can also add references using the
# 'add_references' function.

reference <- new_reference("trainset9_032012.pds.zip", "9_032012",
              "classification by mothur2 v1.0 using default options", "",
"https://mothur.s3.us-east-2.amazonaws.com/wiki/trainset9_032012.pds.zip")

xdev_assign_sequence_taxonomy_tidy(data, sequence_classifications, reference)

}
