% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_treatments}
\alias{xdev_assign_treatments}
\title{xdev_assign_treatments}
\usage{
xdev_assign_treatments(
  data,
  table,
  sample = "sample",
  treatment = "treatment",
  verbose = TRUE
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table}{a data.frame containing sample treatment assignments}

\item{sample}{a string containing the name of the column in 'table'
that contains the samples. Default column name is 'sample'.}

\item{treatment}{a string containing the name of the column in 'table'
that contains the treatments. Default column name is 'treatment'.}

\item{verbose}{a boolean indicating whether or not you want progress
messages. Default = TRUE.}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Assign samples to treatments in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- strollur::new_dataset("my_dataset")
abundance <- readRDS(strollur_example("miseq_abundance_by_sample.rds"))

strollur::xdev_assign_sequence_abundance(data,
                                           table = abundance)

sample_assignments <- readRDS(strollur_example("miseq_sample_design.rds"))

strollur::xdev_assign_treatments(data,
                                   table = sample_assignments)

}
