% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_treatments}
\alias{xdev_assign_treatments}
\title{xdev_assign_treatments}
\usage{
xdev_assign_treatments(
  data,
  table,
  sample = "samples",
  treatment = "treatments",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \link{strollur} object}

\item{table, }{a data.frame containing sample treatment assignments}

\item{sample, }{a string containing the name of the column in 'table'
that contains the samples. Default column name is 'samples'.}

\item{treatment, }{a string containing the name of the column in 'table'
that contains the treatments. Default column name is 'treatments'.}

\item{verbose, }{a boolean indicating whether or not you want progress
messages. Default = TRUE.}
}
\value{
double containing the number of samples assigned to treatments
}
\description{
Assign samples to treatments in a \link{strollur} object
}
\examples{

data <- new_dataset("my_dataset")
sequence_abundance <- readr::read_tsv(strollur_example(
                                      "mothur2_count_table.tsv.gz"),
                                      show_col_types = FALSE)

xdev_assign_sequence_abundance(data, sequence_abundance, "names")

sample_assignments <- readr::read_table(file = strollur_example("mouse.time.design"),
                             col_names = TRUE, show_col_types = FALSE)

xdev_assign_treatments(data, sample_assignments)

}
