% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_assign_treatments}
\alias{xdev_assign_treatments}
\title{xdev_assign_treatments}
\usage{
xdev_assign_treatments(
  data,
  table,
  sample = "samples",
  treatment = "treatments",
  verbose = TRUE
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{table, }{a data.frame containing sample treatment assignments}

\item{sample, }{a string containing the name of the column in 'table'
that contains the samples. Default column name is 'samples'.}

\item{treatment, }{a string containing the name of the column in 'table'
that contains the treatments. Default column name is 'treatments'.}

\item{verbose, }{a boolean indicating whether or not you want progress
messages. Default = TRUE.}
}
\value{
double containing the number of samples assigned to treatments
}
\description{
Assign samples to treatments in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- new_dataset("my_dataset")
sequence_abundance <- readRDS(strollur_example("miseq_abundance_by_sample.rds"))

xdev_assign_sequence_abundance(data, sequence_abundance)

sample_assignments <- readRDS(strollur_example("miseq_sample_design.rds"))

xdev_assign_treatments(data, sample_assignments)

}
