% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_alignment_length}
\alias{xdev_get_alignment_length}
\title{xdev_get_alignment_length}
\usage{
xdev_get_alignment_length(data)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
Integer containing the length of the alignment or -1 if unaligned.
}
\description{
Get the alignment length of sequences in your
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\examples{

 data <- strollur::miseq_sop_example()
 strollur::xdev_get_alignment_length(data)

}
