% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_bin_abundances_by_sample}
\alias{xdev_get_bin_abundances_by_sample}
\title{Get the sequence abundance data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} parsed by sample}
\usage{
xdev_get_bin_abundances_by_sample(data, bin_type = "otu")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{bin_type}{a string indicating the type of bin clusters. Default = "otu"}
}
\value{
2D vector of float containing data requested parsed by sample.
}
\description{
Get the sequence abundance data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object parsed by sample
}
\examples{

data <- strollur::miseq_sop_example()

# To get the `otu` bin abundances parsed by sample
otu_abunds <- strollur::xdev_get_bin_abundances_by_sample(data)

}
