% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_by_sample}
\alias{xdev_get_by_sample}
\title{xdev_get_by_sample}
\usage{
xdev_get_by_sample(
  data,
  type = "sequence_name",
  samples = as.character(c()),
  degap = FALSE
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{type, }{string containing the type of data you want the totals of.
Options include: "sequence_name", "sequence". Default = "sequence_name".}

\item{samples}{a vector of strings containing the names of the samples you
would like sequence names for. By default all samples are included.}

\item{degap}{a logical. Default = FALSE. When degap = \code{TRUE}, all gap
characters will be removed from the sequences.}
}
\value{
2D vector of strings num_seqs x num_samples containing data
requested parsed by sample.
}
\description{
Get the requested data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object parsed by sample
}
\examples{

data <- miseq_sop_example()

# To get the sequence names parsed by sample
xdev_get_by_sample(data, "sequence_name")

# To get the sequence nucleotide strings parsed by sample
parsed_sequences <- xdev_get_by_sample(data, "sequence")

}
