% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_by_sample}
\alias{xdev_get_by_sample}
\title{xdev_get_by_sample}
\usage{
xdev_get_by_sample(data, type = "sequence_names", samples = as.character(c()))
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{type, }{string containing the type of data you want the totals of.
Options include: "sequence_names", "sequences". Default = "sequence_names".}

\item{samples}{a vector of strings containing the names of the samples you
would like sequence names for. By default all samples are included.}
}
\value{
2D vector of strings ([num_seqs][num_samples]) containing data
requested parsed by sample.
}
\description{
Get the requested data in a \link{dataset} object parsed by sample
}
\examples{

data <- miseq_sop_example()

# To get the sequence names parsed by sample
xdev_get_by_sample(data, "sequence_names")

# To get the sequence nucleotide strings parsed by sample
xdev_get_by_sample(data, "sequences")

}
