% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_list_vector}
\alias{xdev_get_list_vector}
\title{xdev_get_list_vector}
\usage{
xdev_get_list_vector(data, type = "otu")
}
\arguments{
\item{data, }{a \link{strollur} object}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
vector of strings containing the names of the sequences in each bin
separated by commas
}
\description{
Get vector of strings containing the sequences bin data
}
\examples{

data <- miseq_sop_example()
xdev_get_list_vector(data)

}
