% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_sequence_indexes_by_sample}
\alias{xdev_get_sequence_indexes_by_sample}
\title{Get indexes of sequences parsed by sample}
\usage{
xdev_get_sequence_indexes_by_sample(data, samples = as.character(c()))
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{samples}{a vector of strings containing the names of the samples you
would like sequence names for. By default all samples are included.}
}
\value{
2D vector of strings indexes for use with xdev_get_sequences and xdev_get_names.
requested parsed by sample. (Indexes start at 1)
}
\description{
As a developer you may want to process data by sample. To save space you can
request the indexes parsed by sample, and then get a single copy of the sequences
and sequence names.
}
\examples{

data <- strollur::miseq_sop_example()

sequences <- strollur::xdev_get_sequences(data)
names <- strollur::xdev_names(data)

# To get the indexes of the names and sequences by sample
indexes <- strollur::xdev_get_sequence_indexes_by_sample(data)

# First sequence in first sample
names[indexes[[1]][1]]
sequences[indexes[[1]][1]]

}
