% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_get_sequences}
\alias{xdev_get_sequences}
\title{xdev_get_sequences}
\usage{
xdev_get_sequences(data, sample = "", degap = FALSE)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{sample}{a string containing the name of the sample you
would like sequence names for. For all samples in dataset, sample = "".}

\item{degap}{a logical. Default = FALSE. When degap = \code{TRUE}, all gap
characters ('-', '.') will be removed from the sequences.}
}
\value{
vector of string containing nucleotide strings of the sequences in
a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Get the nucleotide strings for each sequence in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

 data <- strollur::miseq_sop_example()
 strollur::xdev_get_sequences(data)

}
