% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_has_bin_taxonomy}
\alias{xdev_has_bin_taxonomy}
\title{Determine if a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} has bin taxonomy assignments}
\usage{
xdev_has_bin_taxonomy(data, type = "otu")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{type}{a string indicating the type of bin assignments. Default "otu".}
}
\value{
boolean
}
\description{
Determine if a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} has bin taxonomy assignments
}
\examples{

 data <- strollur::miseq_sop_example()
 strollur::xdev_has_bin_taxonomy(data)

}
