% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_merge_bins}
\alias{xdev_merge_bins}
\title{xdev_merge_bins}
\usage{
xdev_merge_bins(data, bin_names, reason = "merged", bin_type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{bin_names, }{a vector of strings containing the names of the bins you
would like merge. The resulting merged bin will be stored in the first
bin_id in the vector.}

\item{reason, }{a string indicating why you are merging bins. Default =
"merged".}

\item{bin_type, }{a string indicating the type of bin clusters.
Default = "otu"}
}
\description{
Designed with package integration in mind, the merge bins function allows
you to merge bins in a \link{dataset} object
}
\examples{

 data <- miseq_sop_example()

 # to merge otu5 and otu6

 bins_to_merge <- c("Otu005", "Otu006")

 xdev_merge_bins(data = data, bin_names = bins_to_merge)

 # If you look at the scrap report, you will see Otu006 with the trash code
 # set to "merged".

 report(data = data, type = "bin_scrap")

}
