% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_names}
\alias{xdev_names}
\title{xdev_names}
\usage{
xdev_names(
  data,
  type = "sequences",
  bin_type = "otu",
  samples = NULL,
  distinct = FALSE
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{type, }{string containing the type of data you would like. Options
include: "dataset", "sequences", "bins", "samples", "treatments", "reports".
Default = "sequences".}

\item{bin_type, }{string containing the bin type you would like the names
for. Default = "otu".}

\item{samples, }{vector of strings. samples is only used when 'type' =
"sequences" or 'type' = "bins" . samples should contain the names of the
samples you want names for. Default = NULL.}

\item{distinct, }{Boolean. distinct is used when 'type' =
"sequences" or 'type' = "bins" and the samples parameter is used. The
distinct parameter allows you to get the names that are unique to a given
set of samples. When distinct is TRUE, the names function will return the
names that ONLY contain data from the given samples. When distinct is FALSE
the data returned contains data from a given samples, but may ALSO contain
data from other samples. Default = FALSE.}
}
\value{
vector of strings, containing the names requested
}
\description{
Get the names of a given type of data in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

miseq <- miseq_sop_example()

# To get the name of the dataset
xdev_names(data = miseq, type = "dataset")

# To get the names of the sequences in the dataset
xdev_names(data = miseq, type = "sequences")

# To get the names of the sequences that are unique to sample 'F3D0'
xdev_names(data = miseq, type = "sequences", samples = c("F3D0"), distinct = TRUE)

# To get the names of the sequences that include sample 'F3D0'
xdev_names(data = miseq, type = "sequences", samples = c("F3D0"))

# To get the names of the samples in the dataset
xdev_names(data = miseq, type = "samples")

# To get the names of the treatments in the dataset
xdev_names(data = miseq, type = "treatments")

# To get the names of the bins in the dataset
xdev_names(data = miseq, type = "bins")

# To get the names of the bins in the dataset that are unique to 'F3D0'
xdev_names(data = miseq, type = "bins", samples = c("F3D0"), distinct = TRUE)

# To get the names of the bins in the dataset that include sequences
# from 'F3D0'
xdev_names(data = miseq, type = "bins", samples = c("F3D0"), distinct = FALSE)

# To get the names of the reports in the dataset
xdev_names(data = miseq, type = "reports")

}
