% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_bins}
\alias{xdev_remove_bins}
\title{xdev_remove_bins}
\usage{
xdev_remove_bins(data, bin_names, trash_tags, type = "otu")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{bin_names, }{a vector of strings containing the names of the bins you
would like removed.}

\item{trash_tags, }{a vector of strings containing the reasons you are
removing each bin}

\item{type}{a string indicating the type of clusters.}
}
\description{
Designed with package integration in mind, the remove bins function allows
you to remove bins from a \link{dataset} object
}
\examples{

  data <- new_dataset("my_dataset", 2)

  bin_names <- c("bin1", "bin2", "bin3")
  abundances <- c(110, 525, 80)

  assign_bins(data, data.frame(bin_names = bin_names,
                               abundances = abundances))

  get_num_bins(data)

  bins_to_remove <- c("bin1")
  trash_tag <- c("bad_bin")

  xdev_remove_bins(data, bins_to_remove, trash_tag)

  get_num_bins(data)

}
