% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_lineages}
\alias{xdev_remove_lineages}
\title{xdev_remove_lineages}
\usage{
xdev_remove_lineages(data, contaminants, reason = "contaminant")
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{contaminants, }{vector of strings containing the taxonomies you would
like to remove}

\item{reason, }{a string containing reason you are removing the lineages.
Default = "contaminant".}
}
\description{
Designed with package integration in mind, the remove lineages function
allows you to remove contaminents from a \link{dataset}
}
\examples{
data <- read_mothur(fasta = rdataset_example("final.fasta.gz"),
                       count = rdataset_example("final.count_table.gz"),
                       taxonomy = rdataset_example("final.taxonomy.gz"),
                       design = rdataset_example("mouse.time.design"),
                       otu_list = rdataset_example("final.opti_mcc.list.gz"),
                       dataset_name = "miseq_sop")

contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
 "Eukaryota")

xdev_remove_lineages(data = data, contaminants = contaminants)

}
