% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_samples}
\alias{xdev_remove_samples}
\title{xdev_remove_samples}
\usage{
xdev_remove_samples(data, samples)
}
\arguments{
\item{data, }{a \link{dataset} object.}

\item{samples, }{vector of strings containing the names of the samples to
remove.}
}
\description{
Designed with package integration in mind, the remove samples function allows
you to remove samples from a \link{dataset} object
}
\examples{

data <- miseq_sop_example()

get_num_samples(data)

# To remove samples 'F3D0' and 'F3D1'

xdev_remove_samples(data, c("F3D0", "F3D1"))

get_num_samples(data)

}
