% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_remove_sequences}
\alias{xdev_remove_sequences}
\title{xdev_remove_sequences}
\usage{
xdev_remove_sequences(data, sequence_names, trash_tags)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.}

\item{sequence_names}{vector of strings containing the names of the
sequences to remove}

\item{trash_tags}{vector of strings containing the reasons for the sequences
removals}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Designed with package integration in mind, the remove sequences function
allows you to remove sequences from a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\examples{

data <- strollur::miseq_sop_example()

strollur::count(data, type = "sequence")

# For the sake of example let's remove the first 3 sequences from
# miseq_sop_example:

seqs_to_remove <- c("M00967_43_000000000-A3JHG_1_2101_16474_12783",
                   "M00967_43_000000000-A3JHG_1_1113_12711_3318",
                   "M00967_43_000000000-A3JHG_1_2108_14707_9807")
trash_codes <- c("example", "removing", "sequences")

strollur::xdev_remove_sequences(data,
                                  sequence_names = seqs_to_remove,
                                  trash_tags = trash_codes)

# If you look at the scrap report, you the sequences names, listed with the
# trash codes set to "example", "removing", "sequences".

strollur::report(data, type = "sequence_scrap")

# You can see from the get_num_sequences function that the removed
# sequence's abundances are removed from the dataset.

strollur::count(data, type = "sequence")

}
