% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_abundance}
\alias{xdev_set_abundance}
\title{xdev_set_abundance}
\usage{
xdev_set_abundance(
  data,
  sequence_names,
  sequence_abundances,
  reason = "update"
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{sequence_names, }{a vector of strings containing sequence names}

\item{sequence_abundances, }{vector containing the abundances of each
sequence.}

\item{reason, }{a string containing the trash tag to be applied to any
sequences set to 0 abundance. Default = "update".}
}
\description{
Designed with package integration in mind, the set abundance function
allows you to change the abundances of sequences in a \link{dataset} object
without samples.
}
\examples{

names <- c("seq1", "seq2", "seq3",  "seq4")
abunds <- c(1250, 65, 50, 4)

data <- new_dataset("my_dataset", 2)

assign_sequence_abundance(data, data.frame(sequence_names = names,
                                           abundances = abunds))
get_sequence_abundances(data)

seqs_to_update <- c("seq1", "seq3")
new_abunds <- c(1000, 100)

xdev_set_abundance(data, seqs_to_update, new_abunds)

get_sequence_abundances(data)

}
