% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_abundance}
\alias{xdev_set_abundance}
\title{xdev_set_abundance}
\usage{
xdev_set_abundance(
  data,
  sequence_names,
  sequence_abundances,
  reason = "update"
)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{sequence_names, }{a vector of strings containing sequence names}

\item{sequence_abundances, }{vector containing the abundances of each
sequence.}

\item{reason, }{a string containing the trash tag to be applied to any
sequences set to 0 abundance. Default = "update".}
}
\description{
Designed with package integration in mind, the set abundance function
allows you to change the abundances of sequences in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
without samples.
}
\examples{

names <- c("seq1", "seq2", "seq3",  "seq4")
abunds <- c(1250, 65, 50, 4)

data <- new_dataset(dataset_name = "my_dataset")

xdev_assign_sequence_abundance(data = data, table = data.frame(sequence_names = names,
                                           abundances = abunds))
abundance(data = data, type = "sequences")

seqs_to_update <- c("seq1", "seq3")
new_abunds <- c(1000, 100)

xdev_set_abundance(data = data,
                   sequence_names = seqs_to_update,
                   sequence_abundances = new_abunds)

abundance(data = data, type = "sequences")

}
