% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_abundances}
\alias{xdev_set_abundances}
\title{Set abundances of sequences in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object} with sample data}
\usage{
xdev_set_abundances(data, sequence_names, abundances, reason = "update")
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}

\item{sequence_names}{a vector of strings containing sequence names}

\item{abundances}{2D vector num_seqs x num_samples containing
the abundances of each sequence parsed by sample.}

\item{reason}{a string containing the trash tag to be applied to any
sequences set to 0 abundance. Default = "update".}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
Designed with package integration in mind, the set abundances function
allows you to change the abundances of sequences in a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
with samples.
}
\examples{

data <- strollur::new_dataset(dataset_name = "my_dataset")

sequence_names <- c("seq1", "seq1", "seq1", "seq2", "seq2", "seq2", "seq3",
                    "seq3", "seq4")
samples <- c("sample2", "sample3", "sample4", "sample2", "sample3",
             "sample4", "sample2", "sample3", "sample4")
abundances <- c(250, 400, 500, 25, 40, 50, 25, 25, 4)
table <- data.frame(sequence_name = sequence_names,
                    abundance = abundances,
                    sample = samples)

strollur::xdev_assign_sequence_abundance(data,
                                           table = table)

seqs_to_update <- c("seq4")
new_abunds <- list(c(20, 10, 4))

strollur::xdev_set_abundances(data,
                    sequence_names = seqs_to_update,
                    abundances = new_abunds)

}
