% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_dataset_name}
\alias{xdev_set_dataset_name}
\title{Set the name of a \href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_set_dataset_name(data, dataset_name)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{dataset_name}{a string containing the desired name}
}
\value{
No return value, called for side effects.
}
\description{
Designed with package integration in mind, set the name of a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object.
}
\examples{

data <- strollur::new_dataset(dataset_name = "my_dataset")
strollur::xdev_set_dataset_name(data,
                                dataset_name = "new_dataset_name")

}
