% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_dataset_name}
\alias{xdev_set_dataset_name}
\title{xdev_set_dataset_name}
\usage{
xdev_set_dataset_name(data, dataset_name)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{dataset_name, }{a string containing the desired name}
}
\description{
Designed with package integration in mind, set the name of a \link{dataset} object.
}
\examples{

data <- new_dataset(dataset_name = "my_dataset")
xdev_set_dataset_name(data = data, dataset_name = "new_dataset_name")

}
