% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_sequences}
\alias{xdev_set_sequences}
\title{Set neucleotide sequence string in a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}}
\usage{
xdev_set_sequences(
  data,
  sequence_names,
  sequences,
  comments = as.character(c())
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}

\item{sequence_names}{a vector of strings containing sequence names}

\item{sequences}{a vector of strings containing sequence nucleotide strings}

\item{comments}{a vector of strings containing sequence comments.
(Optional)}
}
\value{
an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
}
\description{
Designed with package integration in mind, the set sequences function allows
you to change the nucleotide strings of sequences in a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur} object. For example, set_sequences may be used
after alignment to overwrite the unaligned sequences with aligned sequences.
}
\examples{

data <- strollur::new_dataset(dataset_name = "my_dataset")

strollur::xdev_add_sequences(data,
              table = data.frame(sequence_name = c("seq1", "seq2",
                                                  "seq3", "seq4")))

strollur::xdev_set_sequences(data,
                   sequence_names = c("seq1", "seq2","seq3", "seq4"),
                   sequences = c("ATTGC", "ACTGC", "AGTGC", "TTTGC"))

}
