% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xdev_set_sequences}
\alias{xdev_set_sequences}
\title{xdev_set_sequences}
\usage{
xdev_set_sequences(
  data,
  sequence_names,
  sequences,
  comments = as.character(c())
)
}
\arguments{
\item{data, }{a \link{dataset} object}

\item{sequence_names, }{a vector of strings containing sequence names}

\item{sequences, }{a vector of strings containing sequence nucleotide strings}

\item{comments, }{a vector of strings containing sequence comments.
(Optional)}
}
\description{
Designed with package integration in mind, the set sequences function
allows you to change the nucleotide strings of sequences in a \link{dataset}
object. For example, set_sequences may be used after alignment to overwrite
the unaligned sequences with aligned sequences.
}
\examples{

data <- new_dataset("my_dataset", 2)

add_sequences(data, data.frame(sequence_names = c("seq1", "seq2",
                                                  "seq3", "seq4")))

xdev_set_sequences(data, c("seq1", "seq2","seq3", "seq4"),
                    c("ATTGC", "ACTGC", "AGTGC", "TTTGC"))

}
