% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xint_copy_pointer}
\alias{xint_copy_pointer}
\title{xint_copy_pointer}
\usage{
xint_copy_pointer(data)
}
\arguments{
\item{data, }{a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object}
}
\value{
pointer to an instance of the C++ 'Dataset' class.
}
\description{
For internal use only, copy an instance of the C++ 'Dataset' class.
}
\keyword{internal}
